### 2019-09-04 #1707 @arfon
@doreenfan - many thanks for your submission to JOSS. Please merge this pull request which should fix your paper compilation issues: https://github.com/AMReX-Astro/MAESTROeX/pull/96

Also, does this paper accompany a submission to a AAS journal?
