### 2019-09-04 #1707 @doreenfan
Yes, we have also submitted a paper to ApJ. It's currently under review and
uploaded on Arxiv at https://arxiv.org/abs/1908.03634.
Thanks!
