### 2019-09-04 97da3c0
moved base state from make_gamma
