### 2019-09-05 da06511
GPU ppm type 0 (GPU hydro PR 3/n) (#78)
This offloads slope.F90 to the GPU (used by ppm type 0). It has been tested for the 2d and 3d reacting bubble problems compiled for MPI and GPU.

Note: It is likely that this version will not work if compiled with OpenMP (this will come in a future PR)
