+++ Source/Src_nd/mkutrans.F90
@@ -70,17 +70,9 @@ contains
 
     k = lo(3)
 
-    ! if (ppm_type .eq. 0) then
-    !    ! call slopex_2d(utilde(:,:,1:1),slopex,domlo,domhi,lo,hi,ng_ut,1,adv_bc(:,:,1:1))
-    !    ! call slopey_2d(utilde(:,:,2:2),slopey,domlo,domhi,lo,hi,ng_ut,1,adv_bc(:,:,2:2))
-    !
-    !    ! call slopex_2d(utilde(:,:,k,1:1),Ip(:,:,k,1:1),domlo,domhi,lo,hi,ng_ut,1,adv_bc(:,:,1:1))
-    !    ! call slopey_2d(utilde(:,:,k,2:2),Im(:,:,k,1:1),domlo,domhi,lo,hi,ng_ut,1,adv_bc(:,:,2:2))
-    ! else if (ppm_type .eq. 1 .or. ppm_type .eq. 2) then
-    !
-    ! end if
-
     if (idir == 1) then
+
+
 
        !******************************************************************
        ! create utrans
@@ -90,18 +82,15 @@ contains
           do i=lo(1),hi(1)
 
              if (ppm_type .eq. 0) then
-                ! extrapolate to edges
-                ! ul = utilde(i-1,j,1) + (HALF-(dt2/hx)*max(ZERO,ufull(i-1,j,1)))*slopex(i-1,j,1)
-                ! ur = utilde(i  ,j,1) - (HALF+(dt2/hx)*min(ZERO,ufull(i  ,j,1)))*slopex(i  ,j,1)
 
                 ulx = utilde(i-1,j,k,1) + (HALF-(dt2/hx)*max(ZERO,ufull(i-1,j,k,1)))*Ip(i-1,j,k,1)
                 urx = utilde(i  ,j,k,1) - (HALF+(dt2/hx)*min(ZERO,ufull(i  ,j,k,1)))*Ip(i  ,j,k,1)
+
              else if (ppm_type .eq. 1 .or. ppm_type .eq. 2) then
                 ! extrapolate to edges
                 ulx = Ip(i-1,j,k,1)
                 urx = Im(i  ,j,k,1)
              end if
-             ! end do
 
              ! impose lo i side bc's
              if (i .eq. lo(1) .and. lo(1) .eq. domlo(1)) then
@@ -163,10 +152,6 @@ contains
 
              if (ppm_type .eq. 0) then
                 ! ! extrapolate to edges
-                ! vly = utilde(i,j-1,2) + (HALF-(dt2/hy)*max(ZERO,ufull(i,j-1,2)))*slopey(i,j-1,1)
-                ! vry = utilde(i,j  ,2) - (HALF+(dt2/hy)*min(ZERO,ufull(i,j  ,2)))*slopey(i,j  ,1)
-
-                ! extrapolate to edges
                 vly = utilde(i,j-1,k,2) + (HALF-(dt2/hy)*max(ZERO,ufull(i,j-1,k,2)))*Im(i,j-1,k,1)
                 vry = utilde(i,j  ,k,2) - (HALF+(dt2/hy)*min(ZERO,ufull(i,j  ,k,2)))*Im(i,j  ,k,1)
 
@@ -247,9 +232,9 @@ contains
 
     integer         , intent(in   ) :: domlo(3), domhi(3), lo(3), hi(3)
     integer         , intent(in   ) :: ut_lo(3), ut_hi(3)
-    integer, value  , intent(in   ) :: lev, idir, nc_ut, ng_ut
+    integer, value  , intent(in   ) :: lev, idir, ng_ut, nc_ut
[…]
