+++ Source/Src_nd/make_edge_scal.F90
@@ -96,8 +96,8 @@ contains
     allocate(Ipf(ip_lo(1):ip_hi(1),ip_lo(2):ip_hi(2),ip_lo(3):ip_hi(3),1:2))
     allocate(Imf(im_lo(1):im_hi(1),im_lo(2):im_hi(2),im_lo(3):im_hi(3),1:2))
 
-    allocate(slopex(lo(1)-1:hi(1)+1,lo(2)-1:hi(2)+1,lo(3):hi(3),1:2))
-    allocate(slopey(lo(1)-1:hi(1)+1,lo(2)-1:hi(2)+1,lo(3):hi(3),1:2))
+    allocate(slopex(lo(1)-1:hi(1)+1,lo(2)-1:hi(2)+1,lo(3):hi(3),1))
+    allocate(slopey(lo(1)-1:hi(1)+1,lo(2)-1:hi(2)+1,lo(3):hi(3),1))
 
     ! Normal predictor states.
     ! Allocated from lo:hi+1 in the normal direction
@@ -126,8 +126,10 @@ contains
     k = lo(3)
 
     if (ppm_type .eq. 0) then
-       call slopex_2d(s(:,:,k,comp:comp),slopex(:,:,k,:),domlo,domhi,lo,hi,ng_s,1,adv_bc(:,:,bccomp:bccomp))
-       call slopey_2d(s(:,:,k,comp:comp),slopey(:,:,k,:),domlo,domhi,lo,hi,ng_s,1,adv_bc(:,:,bccomp:bccomp))
+       call slopex_2d(lo-1,hi+1,s(:,:,:,comp:comp),s_lo,s_hi,1, &
+                slopex(:,:,:,1:1),ip_lo,ip_hi,1,domlo,domhi,1,adv_bc,nbccomp,bccomp)
+       call slopey_2d(lo-1,hi+1,s(:,:,:,comp:comp),s_lo,s_hi,1, &
+                slopey(:,:,:,1:1),ip_lo,ip_hi,1,domlo,domhi,1,adv_bc,nbccomp,bccomp)
     else if (ppm_type .eq. 1 .or. ppm_type .eq. 2) then
 
        call ppm_2d(ip_lo,ip_hi,s,s_lo,s_hi,nc_s,&
@@ -604,11 +606,17 @@ contains
 
     if (ppm_type .eq. 0) then
 
-       do k = lo(3)-1,hi(3)+1
-          call slopex_2d(s(:,:,k,comp:),slopex(:,:,k,:),domlo,domhi,lo,hi,ng_s,1,adv_bc(:,:,bccomp:))
-          call slopey_2d(s(:,:,k,comp:),slopey(:,:,k,:),domlo,domhi,lo,hi,ng_s,1,adv_bc(:,:,bccomp:))
-       end do
-       call slopez_3d(s(:,:,:,comp:),slopez,domlo,domhi,lo,hi,ng_s,1,adv_bc(:,:,bccomp:))
+          call slopex_2d(lo-1,hi+1,s(:,:,:,comp:comp),s_lo,s_hi,1, &
+                        slopex,lo-1,hi+1,1,domlo,domhi,1, &
+                        adv_bc,nbccomp,bccomp)
+          call slopey_2d(lo-1,hi+1,s(:,:,:,comp:comp),s_lo,s_hi,1, &
+                        slopey,lo-1,hi+1,1,domlo,domhi,1, &
+                        adv_bc,nbccomp,bccomp)
+
+       call slopez_3d(lo-1,hi+1,s(:,:,:,comp:comp),s_lo,s_hi,1, &
+                       slopez,lo-1,hi+1,1, &
[…]
