+++ Source/Src_nd/ppm.F90
@@ -15,7 +15,6 @@ module ppm_module
   ! cell-centered velocity.
 
   use amrex_error_module
-  use amrex_mempool_module, only : bl_allocate, bl_deallocate
   use amrex_constants_module
   use meth_params_module, only: ppm_type, rel_eps
 
@@ -1643,6 +1642,7 @@ contains
              end do
           end do
        end do
+
 
     else if (ppm_type .eq. 2) then
 
@@ -3534,7 +3534,6 @@ contains
     end if
 
   end subroutine ppm_3d
-
 
 #endif
 
