### 2019-09-05 df731d9
moved base state from makeTfromRhoH
