+++ Source/MaestroRhoHT.cpp
@@ -18,14 +18,11 @@ Maestro::TfromRhoH (Vector<MultiFab>& scal,
 
     Vector<MultiFab> p0_cart(finest_level+1);
 
-    if (spherical == 1) {
-
-        for (int lev=0; lev<=finest_level; ++lev) {
-            p0_cart[lev].define(grids[lev], dmap[lev], 1, 0);
-            p0_cart[lev].setVal(0.);
-        }
-        Put1dArrayOnCart(p0,p0_cart,0,0,bcs_f,0);
+    for (int lev=0; lev<=finest_level; ++lev) {
+        p0_cart[lev].define(grids[lev], dmap[lev], 1, 0);
+        p0_cart[lev].setVal(0.);
     }
+    Put1dArrayOnCart(p0,p0_cart,0,0,bcs_f,0);
 
     for (int lev=0; lev<=finest_level; ++lev) {
 
@@ -46,19 +43,12 @@ Maestro::TfromRhoH (Vector<MultiFab>& scal,
             // use macros in AMReX_ArrayLim.H to pass in each FAB's data,
             // lo/hi coordinates (including ghost cells), and/or the # of components
             // We will also pass "validBox", which specifies the "valid" region.
-            if (spherical == 1) {
+
 #pragma gpu box(tileBox)
-                makeTfromRhoH_sphr(AMREX_INT_ANYD(tileBox.loVect()),
-                                   AMREX_INT_ANYD(tileBox.hiVect()),
-                                   BL_TO_FORTRAN_ANYD(scal_mf[mfi]),
-                                   BL_TO_FORTRAN_ANYD(p0_mf[mfi]));
-            } else {
-#pragma gpu box(tileBox)
-                makeTfromRhoH(AMREX_INT_ANYD(tileBox.loVect()),
-                              AMREX_INT_ANYD(tileBox.hiVect()), lev,
-                              BL_TO_FORTRAN_ANYD(scal_mf[mfi]),
-                              p0.dataPtr());
-            }
+            makeTfromRhoH(AMREX_INT_ANYD(tileBox.loVect()),
+                                AMREX_INT_ANYD(tileBox.hiVect()),
+                                BL_TO_FORTRAN_ANYD(scal_mf[mfi]),
+                                BL_TO_FORTRAN_ANYD(p0_mf[mfi]));
         }
 
     }
