### 2019-09-05 5fca2fb
moved base state from some more routines
