+++ Source/Maestro_F.H
@@ -781,29 +781,14 @@ extern "C"
                     const amrex::Real* normal, const int* no_lo, const int* no_hi);
 
     // rhcc = beta0 * (S - Sbar)
-    void make_rhcc_for_nodalproj(const int* lo, const int* hi, const int lev,
-                                 amrex::Real* rhcc,  const int* c_lo, const int* c_hi,
-                                 const amrex::Real* S_cc,  const int* s_lo, const int* s_hi,
-                                 const amrex::Real* Sbar,
-                                 const amrex::Real* beta0,
-                                 const amrex::Real* delta_gamma1_term, const int* dg_lo, const int* dg_hi);
-
-    void make_rhcc_for_nodalproj_sphr(const int* lo, const int* hi,
+    void make_rhcc_for_nodalproj(const int* lo, const int* hi,
 				      amrex::Real* rhcc,  const int* c_lo, const int* c_hi,
 				      const amrex::Real* S_cc, const int* s_lo, const int* s_hi,
 				      const amrex::Real* Sbar_cart, const int* sb_lo, const int* sb_hi,
 				      const amrex::Real* beta0_cart, const int* b_lo, const int* b_hi,
               const amrex::Real* delta_gamma1_term, const int* dg_lo, const int* dg_hi);
 
-    void create_correction_cc(const int* lo, const int* hi, const int lev,
-			      amrex::Real* correction_cc, const int* c_lo, const int* c_hi,
-			      const amrex::Real* delta_p_term, const int* dp_lo, const int* dp_hi,
-			      const amrex::Real* beta0,
-			      const amrex::Real* gamma1bar,
-			      const amrex::Real* p0,
-			      const amrex::Real dt);
-
-    void create_correction_cc_sphr(const int* lo, const int* hi,
+    void create_correction_cc(const int* lo, const int* hi,
 				   amrex::Real* correction_cc, const int* c_lo, const int* c_hi,
 				   const amrex::Real* delta_p_term, const int* dp_lo, const int* dp_hi,
 				   const amrex::Real* beta0_cart, const int* b_lo, const int* b_hi,
@@ -813,19 +798,7 @@ extern "C"
 				   const amrex::Real dt);
 
     // rhcc = beta0 * (S - Sbar) + beta0 * delta_chi
-    void make_rhcc_for_macproj(const int* lo, const int* hi, const int lev,
-                               amrex::Real* rhcc,  const int* c_lo, const int* c_hi,
-                               const amrex::Real* S_cc,  const int* s_lo, const int* s_hi,
-                               const amrex::Real* Sbar,
-                               const amrex::Real* beta0,
-                               const amrex::Real* delta_gamma1_term, const int* dg_lo, const int* dg_hi,
-                               const amrex::Real* gamma1bar,
-			                         const amrex::Real* p0,
-                               const amrex::Real* delta_p_term, const int* dp_lo, const int* dp_hi,
-			             amrex::Real* delta_chi, const int* dc_lo, const int* dc_hi,
-			       const amrex::Real dt, const int is_predictor);
-
-    void make_rhcc_for_macproj_sphr(const int* lo, const int* hi,
+    void make_rhcc_for_macproj(const int* lo, const int* hi,
 				    amrex::Real* rhcc,  const int* c_lo, const int* c_hi,
 				    const amrex::Real* S_cc, const int* s_lo, const int* s_hi,
 				    const amrex::Real* Sbar_cart, const int* sb_lo, const int* sb_hi,
@@ -865,8 +838,23 @@ extern "C"
 #if (AMREX_SPACEDIM == 3)
                            const amrex::Real* wmac,  const int* w_lo, const int* w_hi,
 #endif
[…]
