### 2019-09-09 46da875
moved base state from some more routines
