+++ Source/Maestro.H
@@ -1034,11 +1034,6 @@ private:
     /// @param p0       base state pressure
     /// @param mach     Mach number
     void MachfromRhoH (const amrex::Vector<amrex::MultiFab>& scal,
-                       const amrex::Vector<amrex::MultiFab>& vel,
-                       const RealVector& p0,
-                       amrex::Vector<amrex::MultiFab>& mach);
-
-    void MachfromRhoHSphr (const amrex::Vector<amrex::MultiFab>& scal,
                            const amrex::Vector<amrex::MultiFab>& vel,
                            const RealVector& p0,
                            const amrex::Vector<amrex::MultiFab>& w0cart,
