+++ Source/MaestroAdvection.cpp
@@ -83,6 +83,13 @@ Maestro::MakeUtrans (const Vector<MultiFab>& utilde,
     // timer for profiling
     BL_PROFILE_VAR("Maestro::MakeUtrans()",MakeUtrans);
 
+    Vector<MultiFab> w0_cart(finest_level+1);
+    for (int lev=0; lev<=finest_level; ++lev) {
+        w0_cart[lev].define(grids[lev], dmap[lev], 1, 1);
+    }
+
+    Put1dArrayOnCart(w0,w0_cart,0,0,bcs_u,0);
+
     for (int lev=0; lev<=finest_level; ++lev) {
 
         // Get the index space and grid spacing of the domain
@@ -99,10 +106,11 @@ Maestro::MakeUtrans (const Vector<MultiFab>& utilde,
               Im.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
 #if (AMREX_SPACEDIM == 3)
               MultiFab& wtrans_mf  = utrans[lev][2];
-	const MultiFab& w0macx_mf  = w0mac[lev][0];
-	const MultiFab& w0macy_mf  = w0mac[lev][1];
-	const MultiFab& w0macz_mf  = w0mac[lev][2];
+        const MultiFab& w0macx_mf  = w0mac[lev][0];
+        const MultiFab& w0macy_mf  = w0mac[lev][1];
+        const MultiFab& w0macz_mf  = w0mac[lev][2];
 #endif
+        const MultiFab& w0_mf = w0_cart[lev];
 
 #ifdef AMREX_USE_CUDA
         int* bc_f = prepare_bc(bcs_u[0].data(), 1);
@@ -191,7 +199,7 @@ Maestro::MakeUtrans (const Vector<MultiFab>& utilde,
 #pragma gpu box(xbx)
             mkutrans_2d(AMREX_INT_ANYD(xbx.loVect()),
                         AMREX_INT_ANYD(xbx.hiVect()),
-                        lev, 1,
+                        1,
                         AMREX_INT_ANYD(domainBox.loVect()),
                         AMREX_INT_ANYD(domainBox.hiVect()),
                         BL_TO_FORTRAN_ANYD(utilde_mf[mfi]), utilde_mf.nComp(), utilde_mf.nGrow(),
@@ -200,7 +208,8 @@ Maestro::MakeUtrans (const Vector<MultiFab>& utilde,
                         BL_TO_FORTRAN_ANYD(vtrans_mf[mfi]),
                         BL_TO_FORTRAN_ANYD(Ip[mfi]),
                         BL_TO_FORTRAN_ANYD(Im[mfi]),
-                        w0.dataPtr(), AMREX_REAL_ANYD(dx), dt, bc_f,
+                        BL_TO_FORTRAN_ANYD(w0_mf[mfi]), 
+                        AMREX_REAL_ANYD(dx), dt, bc_f,
                         phys_bc.dataPtr());
 
             if (ppm_type == 0) {
@@ -236,7 +245,7 @@ Maestro::MakeUtrans (const Vector<MultiFab>& utilde,
 #pragma gpu box(ybx)
            mkutrans_2d(AMREX_INT_ANYD(ybx.loVect()),
                        AMREX_INT_ANYD(ybx.hiVect()),
-                       lev, 2,
+                       2,
                        AMREX_INT_ANYD(domainBox.loVect()),
                        AMREX_INT_ANYD(domainBox.hiVect()),
                        BL_TO_FORTRAN_ANYD(utilde_mf[mfi]), utilde_mf.nComp(), utilde_mf.nGrow(),
[…]
