+++ Source/MaestroAdvection.cpp
@@ -66,7 +66,7 @@ Maestro::AdvancePremac (Vector<std::array< MultiFab, AMREX_SPACEDIM > >& umac,
 
 	int do_add_utilde_force = 1;
 	MakeVelForce(vel_force,utrans,sold,rho0_old,grav_cell_old,
-	             w0_force,w0_force_cart,do_add_utilde_force);
+	             w0_force_cart,do_add_utilde_force);
 
 	// add w0 to trans velocities
 	Addw0 (utrans,w0mac,1.);

[…]
