+++ Exec/SCIENCE/code_comp/initdata.f90
@@ -6,12 +6,13 @@ module initdata_module
   use amrex_constants_module
   use network, only: nspec
   use amrex_fort_module, only : amrex_spacedim, amrex_random
-  use base_state_geometry_module, only: nr_fine, max_radial_level
+  use base_state_geometry_module, only: nr_fine, max_radial_level, center
   use meth_params_module, only: nscal, rho_comp, rhoh_comp, temp_comp, spec_comp, pi_comp, &
        prob_lo, prob_hi
   use eos_module
   use eos_type_module
   use fill_3d_data_module, only: put_1d_array_on_cart_sphr
+  use probin_module, only: rho_0
 
   implicit none
 
@@ -38,7 +39,10 @@ contains
     double precision, intent(in   ) :: p0_init(0:max_radial_level,0:nr_fine-1)
     double precision, intent(in   ) :: dx(3)
 
-    integer          :: i,j,k,r
+    integer :: i, j, k, r
+    double precision :: x, y, z, fheat, rhopert
+
+    type(eos_t) :: eos_state
 
     if (parallel_IOProcessor()) then
        print*,"Create a local copy of initdata.f90 in your build directory"
@@ -52,23 +56,40 @@ contains
     vel(lo(1):hi(1),lo(2):hi(2),lo(3):hi(3),1:nc_v) = 0.d0
 
     do k=lo(3),hi(3)
+       z = prob_lo(3) + dx(3)*(dble(k) + HALF) - center(3)
        do j=lo(2),hi(2)
-          do i=lo(1),hi(1)
+          y = prob_lo(2) + dx(2)*(dble(j) + HALF) - center(2)
 
-             if (amrex_spacedim .eq. 1) then
-                r = i
-             else if (amrex_spacedim .eq. 2) then
+          if (y < 1.125d0 * 4.d8) then
+              fheat = sin(8.d0 * M_PI * (y/ 4.d8 - ONE))
+          else
+              fheat = ZERO
+          endif
+
+          do i=lo(1),hi(1)
+             x = prob_lo(1) + dx(1)*(dble(i) + HALF) - center(1)
+
+             rhopert = 5.d-5 * rho_0 * fheat * (sin(3.d0 * M_PI * x / 4.d8) + &
+                                                cos(M_PI * x / 4.d8)) * &
+                     (sin(3 * M_PI * z/4.d8) - cos(M_PI * z/4.d8))
+
+             if (amrex_spacedim .eq. 2) then
                 r = j
              else if (amrex_spacedim .eq. 3) then
                 r = k
              end if
 
+             eos_state % rho = s0_init(lev,r,rho_comp) + rhopert 
+             eos_state % p = p0_init(lev, r)
+             eos_state % xn(:) = s0_init(lev,r,spec_comp:spec_comp+nspec-1)
+
+             call eos(eos_input_rp, eos_state)
+
              ! set scalars using s0
-             scal(i,j,k,rho_comp)  = s0_init(lev,r,rho_comp)
-             scal(i,j,k,rhoh_comp) = s0_init(lev,r,rhoh_comp)
-             scal(i,j,k,temp_comp) = s0_init(lev,r,temp_comp)
-             scal(i,j,k,spec_comp:spec_comp+nspec-1) = &
-                  s0_init(lev,r,spec_comp:spec_comp+nspec-1)
+             scal(i,j,k,rho_comp)  = eos_state % rho
+             scal(i,j,k,rhoh_comp) = eos_state % rho * eos_state % h
+             scal(i,j,k,temp_comp) = eos_state % t
[…]
