### 2019-09-10 09f07ba
fixed setup and inputs file for code comp problem
