### 2019-09-11 47701e6
GPU velpred 2d (GPU hydro PR 4/n) (#79)
This offloads velpred_2d to the GPU. It has been tested for the 2d reacting bubble problem compiled for MPI and GPU.

Note: It is likely that this version will not work if compiled with OpenMP (this will come in a future PR)
