+++ Source/MaestroAdvection.cpp
@@ -448,6 +448,25 @@ Maestro::VelPred (const Vector<MultiFab>& utilde,
         const MultiFab& utrans_mf  = utrans[lev][0];
         const MultiFab& vtrans_mf  = utrans[lev][1];
               MultiFab& vmac_mf    = umac[lev][1];
+              MultiFab Ipu, Imu, Ipv, Imv;
+              Ipu.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+              Imu.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+              Ipv.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+              Imv.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+
+              MultiFab Ipfx, Imfx, Ipfy, Imfy;
+              Ipfx.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+              Imfx.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+              Ipfy.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+              Imfy.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+
+              MultiFab ulx, urx, uimhx, uly, ury, uimhy;
+              ulx.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+              urx.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+              uimhx.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+              uly.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+              ury.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+              uimhy.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
 #if (AMREX_SPACEDIM == 3)
         const MultiFab& wtrans_mf  = utrans[lev][2];
               MultiFab& wmac_mf    = umac[lev][2];
@@ -459,6 +478,8 @@ Maestro::VelPred (const Vector<MultiFab>& utilde,
 
         // loop over boxes (make sure mfi takes a cell-centered multifab as an argument)
 
+#if (AMREX_SPACEDIM == 3)
+
 	// NOTE: don't tile, but threaded in fortran subroutine
         for ( MFIter mfi(utilde_mf); mfi.isValid(); ++mfi ) {
 
@@ -469,31 +490,237 @@ Maestro::VelPred (const Vector<MultiFab>& utilde,
             // use macros in AMReX_ArrayLim.H to pass in each FAB's data,
             // lo/hi coordinates (including ghost cells), and/or the # of components
             // We will also pass "validBox", which specifies the "valid" region.
-#if (AMREX_SPACEDIM == 2)
-            velpred_2d(
-#elif (AMREX_SPACEDIM == 3)
-            velpred_3d(
-#endif
-                        &lev, ARLIM_3D(domainBox.loVect()), ARLIM_3D(domainBox.hiVect()),
+            velpred_3d(&lev, 
+                        ARLIM_3D(domainBox.loVect()), ARLIM_3D(domainBox.hiVect()),
                         ARLIM_3D(tileBox.loVect()), ARLIM_3D(tileBox.hiVect()),
                         BL_TO_FORTRAN_FAB(utilde_mf[mfi]), utilde_mf.nGrow(),
                         BL_TO_FORTRAN_FAB(ufull_mf[mfi]), ufull_mf.nGrow(),
                         BL_TO_FORTRAN_3D(utrans_mf[mfi]),
[…]
