### 2019-09-11 8621f80
fixed spherical issue and missing k = in slopey_2d
