+++ Source/Maestro_F.H
@@ -1029,7 +1029,8 @@ extern "C"
         const int* domlo, const int* domhi,
         const int* adv_bc,
         const amrex::Real* dx, const amrex::Real dt,
-        const int is_umac, const int comp, const int bccomp);
+        const int is_umac, const int comp, const int bccomp, 
+        const int nbccomp);
 #endif
     //////////////////////
 
