+++ Source/MaestroGamma.cpp
@@ -21,11 +21,11 @@ Maestro::MakeGamma1bar (const Vector<MultiFab>& scal,
     Vector<MultiFab> p0_cart(finest_level+1);
 
     for (int lev=0; lev<=finest_level; ++lev) {
-        gamma1[lev].define(grids[lev], dmap[lev], 1, 0);
+        gamma1[lev].define(grids[lev], dmap[lev], 1, 1);
     }
 
     for (int lev=0; lev<=finest_level; ++lev) {
-        p0_cart[lev].define(grids[lev], dmap[lev], 1, 0);
+        p0_cart[lev].define(grids[lev], dmap[lev], 1, 1);
         p0_cart[lev].setVal(0.);
     }
 
