### 2019-09-11 039772e
fixed test_stability
