### 2019-09-11 dce6223
fixed some boundary conditions
