+++ Source/Src_nd/make_gamma.F90
@@ -51,8 +51,6 @@ contains
              ! dens, pres, and xmass are inputs
              call eos(eos_input_rp, eos_state, pt_index)
 
-            !  write(*,*) "rho = ", eos_state % rho, "p = ", eos_state %p
-
              gamma(i,j,k) = eos_state%gam1
 
           end do
