### 2019-09-12 fd6f14c
this fixed #98 so that ppm_2d no longer assumes the bcs have AMREX_SPACEDIM components
