+++ Source/Src_nd/make_edge_scal.F90
@@ -135,12 +135,12 @@ contains
        call ppm_2d(ip_lo,ip_hi,s,s_lo,s_hi,nc_s,&
                     umac,u_lo,u_hi,vmac,v_lo,v_hi,&
                     Ip,ip_lo,ip_hi,Im,im_lo,im_hi, &
-                    domlo,domhi,adv_bc,dx,dt,1,comp,bccomp)
+                    domlo,domhi,adv_bc,dx,dt,1,comp,bccomp,nbccomp)
        if (ppm_trace_forces .eq. 1) then
           call ppm_2d(ip_lo,ip_hi,force,f_lo,f_hi,nc_f,&
                       umac,u_lo,u_hi,vmac,v_lo,v_hi,&
                       Ipf,ip_lo,ip_hi,Imf,im_lo,im_hi, &
-                      domlo,domhi,adv_bc,dx,dt,1,comp,bccomp)
+                      domlo,domhi,adv_bc,dx,dt,1,comp,bccomp,nbccomp)
        end if
 
     end if
