### 2019-09-13 0477ee2
fixed test_stability
