### 2019-09-13 df6fcd5
make_flux subroutines use split velocities instead of full velocities, so we need to delay adding w0 until after advection routines have been called
