### 2019-09-14 0bf75b0
rho0_old_cart was not initialized to zero, which caused failure in debug mode
