### 2019-09-14 878ce90
GPU velpred 3d (GPU hydro PR 5/n) (#80)
This offloads velpred_3d to the GPU. It has been tested for the 3d reacting bubble & wdconvect problems compiled for MPI and GPU.

Note: It is likely that this version will not work if compiled with OpenMP (this will come in a future PR)
