+++ Source/MaestroAdvection.cpp
@@ -449,6 +449,15 @@ Maestro::VelPred (const Vector<MultiFab>& utilde,
     // timer for profiling
     BL_PROFILE_VAR("Maestro::VelPred()",VelPred);
 
+    Vector<MultiFab> w0_cart(finest_level+1);
+
+    for (int lev=0; lev<=finest_level; ++lev) {
+        w0_cart[lev].define(grids[lev], dmap[lev], AMREX_SPACEDIM, 1);
+        w0_cart[lev].setVal(0.);
+    }
+
+    Put1dArrayOnCart(w0,w0_cart,0,1,bcs_u,0,1);
+
     for (int lev=0; lev<=finest_level; ++lev) {
 
         // Get the index space and grid spacing of the domain
@@ -483,46 +492,38 @@ Maestro::VelPred (const Vector<MultiFab>& utilde,
               uimhy.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
 #if (AMREX_SPACEDIM == 3)
         const MultiFab& wtrans_mf  = utrans[lev][2];
-              MultiFab& wmac_mf    = umac[lev][2];
-	const MultiFab& w0macx_mf  = w0mac[lev][0];
-	const MultiFab& w0macy_mf  = w0mac[lev][1];
-	const MultiFab& w0macz_mf  = w0mac[lev][2];
+        MultiFab& wmac_mf    = umac[lev][2];
+        const MultiFab& w0macx_mf  = w0mac[lev][0];
+        const MultiFab& w0macy_mf  = w0mac[lev][1];
+        const MultiFab& w0macz_mf  = w0mac[lev][2];
+
+        MultiFab Ipw, Imw;
+        Ipw.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+        Imw.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+
+        MultiFab Ipfz, Imfz;
+        Ipfz.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+        Imfz.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+
+        MultiFab ulz, urz, uimhz;
+        ulz.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+        urz.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+        uimhz.define(grids[lev],dmap[lev],AMREX_SPACEDIM,1);
+
+        MultiFab uimhyz, uimhzy, vimhxz, vimhzx, wimhxy, wimhyx;
+        uimhyz.define(grids[lev],dmap[lev],1,1);
+        uimhzy.define(grids[lev],dmap[lev],1,1);
+        vimhxz.define(grids[lev],dmap[lev],1,1);
+        vimhzx.define(grids[lev],dmap[lev],1,1);
+        wimhxy.define(grids[lev],dmap[lev],1,1);
+        wimhyx.define(grids[lev],dmap[lev],1,1);
 #endif
         const MultiFab& force_mf = force[lev];
+        const MultiFab& w0_mf = w0_cart[lev];
 
         // loop over boxes (make sure mfi takes a cell-centered multifab as an argument)
 
-#if (AMREX_SPACEDIM == 3)
-
-	// NOTE: don't tile, but threaded in fortran subroutine
-        for ( MFIter mfi(utilde_mf); mfi.isValid(); ++mfi ) {
-
-            // Get the index space of the valid region
-            const Box& tileBox = mfi.tilebox();
-
-            // call fortran subroutine
-            // use macros in AMReX_ArrayLim.H to pass in each FAB's data,
-            // lo/hi coordinates (including ghost cells), and/or the # of components
[…]

[…]
