### 2019-09-15 #1707 @jedbrown
This is in my domain, but I should note possible CoI with authors before offering to edit:
1. I co-authored the jupyter4edu book with Michael last year (along with several other JOSS editors)
2. I have been funded by projects that included Ann (e.g., DOE FASTMath). I'm no longer directly funded by such projects, but I still collaborate with many members (though not directly with Ann).
