### 2019-09-17 8fbbfc8
Always set i,j,k for burn states regardless of USE_NONAKA_PLOT
