### 2019-09-18 dfc1a3a
pass reaction start time to the burner
