+++ Source/Src_nd/burner_loop.F90
@@ -39,7 +39,7 @@ contains
        rho_Hext, e_lo, e_hi, &
        rho_odot, r_lo, r_hi, &
        rho_Hnuc, n_lo, n_hi, &
-       tempbar_init_in, dt_in, &
+       tempbar_init_in, dt_in, time_in, &
        mask,     m_lo, m_hi, use_mask) &
        bind (C,name="burner_loop")
 
@@ -58,6 +58,7 @@ contains
     double precision, intent (inout) :: rho_Hnuc(n_lo(1):n_hi(1),n_lo(2):n_hi(2),n_lo(3):n_hi(3))
     double precision, intent (in   ) :: tempbar_init_in(0:max_radial_level,0:nr_fine-1)
     double precision, value, intent (in) :: dt_in
+    double precision, value, intent (in) :: time_in
     integer         , intent (in   ) :: mask(m_lo(1):m_hi(1),m_lo(2):m_hi(2),m_lo(3):m_hi(3))
     integer, value  , intent (in   ) :: use_mask
 
@@ -131,7 +132,7 @@ contains
                    state_in % k = k
                    
                    call copy_burn_t(state_out, state_in)
-                   call burner(state_in, state_out, dt_in)
+                   call burner(state_in, state_out, dt_in, time_in)
                    do n = 1, nspec
                       x_out(n) = state_out % xn(n)
                    enddo
@@ -192,7 +193,7 @@ contains
        rho_Hext, e_lo, e_hi, &
        rho_odot, r_lo, r_hi, &
        rho_Hnuc, n_lo, n_hi, &
-       tempbar_init_cart, t_lo, t_hi, dt_in, &
+       tempbar_init_cart, t_lo, t_hi, dt_in, time_in, &
        mask,     m_lo, m_hi, use_mask) &
        bind (C,name="burner_loop_sphr")
 
@@ -211,6 +212,7 @@ contains
     integer         , intent (in   ) :: t_lo(3), t_hi(3)
     double precision, intent (in   ) :: tempbar_init_cart(t_lo(1):t_hi(1),t_lo(2):t_hi(2),t_lo(3):t_hi(3))
     double precision, value, intent (in) :: dt_in
+    double precision, value, intent (in) :: time_in
     integer         , intent (in   ) :: mask(m_lo(1):m_hi(1),m_lo(2):m_hi(2),m_lo(3):m_hi(3))
     integer, value  , intent (in   ) :: use_mask
 
@@ -279,7 +281,7 @@ contains
                    state_in % k = k
                    
                    call copy_burn_t(state_out, state_in)
-                   call burner(state_in, state_out, dt_in)
+                   call burner(state_in, state_out, dt_in, time_in)
                    do n = 1, nspec
                       x_out(n) = state_out % xn(n)
                    enddo
