+++ Source/Maestro_F.H
@@ -116,7 +116,7 @@ extern "C"
                            amrex::Real* rho_odot, const int* r_lo, const int* r_hi,
                            amrex::Real* rho_Hnuc, const int* n_lo, const int* n_hi,
                      const amrex::Real* tempbar_init_in,
-                     const amrex::Real dt_in,
+                     const amrex::Real dt_in, const amrex::Real time_in, 
                      const int* mask, const int* m_lo, const int* m_hi,
                      const int use_mask);
 
@@ -127,7 +127,7 @@ extern "C"
                                 amrex::Real* rho_odot, const int* r_lo, const int* r_hi,
                                 amrex::Real* rho_Hnuc, const int* n_lo, const int* n_hi,
                           const amrex::Real* tempbar_init_cart, const int* t_lo, const int* t_hi,
-                          const amrex::Real dt_in,
+                          const amrex::Real dt_in, const amrex::Real time_in, 
                           const int* mask, const int* m_lo, const int* m_hi,
                           const int use_mask);
 
