### 2019-09-18 5767728
added some analysis scripts for the code comp problem (taken from the Castro flame_wave directory)
