### 2019-09-18 #1707 @zingale
potential suggestions (some mined from the 'stars' on one of our python hydro projects):

Philipp Edelmann (@tukss)
James Guilochon (@guillochon)
Claire Kopenhafer (@clairekope)
