### 2019-09-19 #1707 @tukss
I would gladly be a reviewer. That would be my first JOSS review.
