### 2019-09-19 #1707 @arfon
> I would gladly be a reviewer. That would be my first JOSS review.

🎉 excellent thanks!

I'll go ahead and add you as a reviewer but I won't actually start the main review (with associated checklist etc.) until I've found a second reviewer. So for now, please sit tight!
