### 2019-09-19 d2bfea6
now all instances of the temporary w0_cart have only 1 ghost cell
cleaned up indentation in MaestroPlot.cpp
