+++ Source/MaestroPlot.cpp
@@ -17,8 +17,8 @@ void Maestro::WriteSmallPlotFile (const int step,
                                   Vector<MultiFab>& s_in,
                                   const Vector<MultiFab>& S_cc_in)
 {
-	WritePlotFile (step, t_in, dt_in, rho0_in, rhoh0_in, p0_in,
-	               gamma1bar_in, u_in, s_in, S_cc_in, true);
+    WritePlotFile (step, t_in, dt_in, rho0_in, rhoh0_in, p0_in,
+                   gamma1bar_in, u_in, s_in, S_cc_in, true);
 }
 
 // write plotfile to disk
@@ -35,112 +35,112 @@ Maestro::WritePlotFile (const int step,
                         const Vector<MultiFab>& S_cc_in,
                         const bool is_small)
 {
-	// timer for profiling
-	BL_PROFILE_VAR("Maestro::WritePlotFile()",WritePlotFile);
+    // timer for profiling
+    BL_PROFILE_VAR("Maestro::WritePlotFile()",WritePlotFile);
 
-	// wallclock time
-	const Real strt_total = ParallelDescriptor::second();
+    // wallclock time
+    const Real strt_total = ParallelDescriptor::second();
 
-	std::string plotfilename;
+    std::string plotfilename;
 
-	if (!is_small) {
-		plotfilename = plot_base_name;
-	} else {
-		plotfilename = small_plot_base_name;
-	}
+    if (!is_small) {
+        plotfilename = plot_base_name;
+    } else {
+        plotfilename = small_plot_base_name;
+    }
 
-	if (step == plotInitData) {
-		if (plotfilename.back() == '_') {
-			plotfilename += "InitData";
-		} else {
-			plotfilename += +"_InitData";
-		}
+    if (step == plotInitData) {
+        if (plotfilename.back() == '_') {
+            plotfilename += "InitData";
+        } else {
+            plotfilename += +"_InitData";
+        }
 
-	}
-	else if (step == plotInitProj) {
-		if (plotfilename.back() == '_') {
-			plotfilename += "after_InitProj";
-		} else {
-			plotfilename += +"_after_InitProj";
-		}
-	}
-	else if (step == plotDivuIter) {
-		if (plotfilename.back() == '_') {
-			plotfilename += "after_DivuIter";
-		} else {
-			plotfilename += +"_after_DivuIter";
-		}
-	}
-	else {
-		PlotFileName(step, &plotfilename);
-	}
+    }
+    else if (step == plotInitProj) {
+        if (plotfilename.back() == '_') {
+            plotfilename += "after_InitProj";
+        } else {
+            plotfilename += +"_after_InitProj";
+        }
+    }
+    else if (step == plotDivuIter) {
+        if (plotfilename.back() == '_') {
+            plotfilename += "after_DivuIter";
+        } else {
+            plotfilename += +"_after_DivuIter";
+        }
+    }
+    else {
+        PlotFileName(step, &plotfilename);
+    }
 
-	// convert rho0 to multi-D MultiFab
-	Vector<MultiFab> rho0_cart(finest_level+1);
-	for (int lev=0; lev<=finest_level; ++lev) {
-		rho0_cart[lev].define(grids[lev], dmap[lev], 1, 0);
-	}
-	Put1dArrayOnCart(rho0_in,rho0_cart,0,0);
[…]
