+++ Source/Maestro_F.H
@@ -90,14 +90,14 @@ extern "C"
     // in bc_fill_nd.F90
 
     void scalarfill (amrex::Real* phi, const int* phi_lo, const int* phi_hi,
-		  const int* domlo, const int* domhi,
-		  const amrex::Real* dx, const amrex::Real* gridlo,
-		  const amrex::Real* time, const int* bc, const int icomp);
-
+                     const int* domlo, const int* domhi,
+                     const amrex::Real* dx, const amrex::Real* gridlo,
+                     const amrex::Real* time, const int* bc, const int icomp);
+    
     void velfill (amrex::Real* vel, const int* vel_lo, const int* vel_hi,
-    	  const int* domlo, const int* domhi,
-    	  const amrex::Real* dx, const amrex::Real* gridlo,
-    	  const amrex::Real* time, const int* bc, const int icomp);
+                  const int* domlo, const int* domhi,
+                  const amrex::Real* dx, const amrex::Real* gridlo,
+                  const amrex::Real* time, const int* bc, const int icomp);
 
     //////////////////////
 
@@ -160,7 +160,7 @@ extern "C"
 		    const amrex::Real* force, const int* f_lo, const int* f_hi, const int* nc_f,
 		    const amrex::Real* divu, const int* d_lo, const int* d_hi,
 		    const amrex::Real* dSdt, const int* t_lo, const int* t_hi,
-            const amrex::Real* w0_cart, const int* w_lo, const int* w_hi,
+                    const amrex::Real* w0_cart, const int* w_lo, const int* w_hi,
 		    const amrex::Real* w0macx, const int* x_lo, const int* x_hi,
 		    const amrex::Real* w0macy, const int* y_lo, const int* y_hi,
 		    const amrex::Real* w0macz, const int* z_lo, const int* z_hi,
@@ -446,24 +446,24 @@ extern "C"
 #elif (AMREX_SPACEDIM == 3)
     void make_edge_scal_3d(
 #endif
-                      const int* domlo, const int* domhi,
-                      const int* lo, const int* hi,
-                      const amrex::Real* s,      const int* s_lo, const int* s_hi, const int nc_s, int ng_s,
-                            amrex::Real* sedgex, const int* x_lo, const int* x_hi, const int nc_x,
-                            amrex::Real* sedgey, const int* y_lo, const int* y_hi, const int nc_y,
+                           const int* domlo, const int* domhi,
+                           const int* lo, const int* hi,
+                           const amrex::Real* s,      const int* s_lo, const int* s_hi, const int nc_s, int ng_s,
+                           amrex::Real* sedgex, const int* x_lo, const int* x_hi, const int nc_x,
+                           amrex::Real* sedgey, const int* y_lo, const int* y_hi, const int nc_y,
 #if (AMREX_SPACEDIM == 3)
-                            amrex::Real* sedgez, const int* z_lo, const int* z_hi, const int nc_z,
+                           amrex::Real* sedgez, const int* z_lo, const int* z_hi, const int nc_z,
 #endif
-                      const amrex::Real* umac,   const int* u_lo, const int* u_hi,
-                      const amrex::Real* vmac,   const int* v_lo, const int* v_hi,
+                           const amrex::Real* umac,   const int* u_lo, const int* u_hi,
+                           const amrex::Real* vmac,   const int* v_lo, const int* v_hi,
 #if (AMREX_SPACEDIM == 3)
-                      const amrex::Real* wmac,   const int* w_lo, const int* w_hi,
+                           const amrex::Real* wmac,   const int* w_lo, const int* w_hi,
 #endif
-                      const int ng_um,
-                      const amrex::Real* force,  const int* f_lo, const int* f_hi, const int nc_f,
[…]
