### 2019-09-19 fac1970
w0_cart is now part of the MAESTRO class
there are no more temporary w0_carts created

this does not address the fact that filling in w0_cart from w0 may be happening more times than it should
