+++ Source/MaestroAdvection.cpp
@@ -83,12 +83,9 @@ Maestro::MakeUtrans (const Vector<MultiFab>& utilde,
     // timer for profiling
     BL_PROFILE_VAR("Maestro::MakeUtrans()",MakeUtrans);
 
-    Vector<MultiFab> w0_cart(finest_level+1);
     for (int lev=0; lev<=finest_level; ++lev) {
-        w0_cart[lev].define(grids[lev], dmap[lev], AMREX_SPACEDIM, 1);
         w0_cart[lev].setVal(0.);
     }
-
     Put1dArrayOnCart(w0,w0_cart,0,1,bcs_u,0,1);
 
     for (int lev=0; lev<=finest_level; ++lev) {
@@ -449,13 +446,9 @@ Maestro::VelPred (const Vector<MultiFab>& utilde,
     // timer for profiling
     BL_PROFILE_VAR("Maestro::VelPred()",VelPred);
 
-    Vector<MultiFab> w0_cart(finest_level+1);
-
     for (int lev=0; lev<=finest_level; ++lev) {
-        w0_cart[lev].define(grids[lev], dmap[lev], AMREX_SPACEDIM, 1);
         w0_cart[lev].setVal(0.);
     }
-
     Put1dArrayOnCart(w0,w0_cart,0,1,bcs_u,0,1);
 
     for (int lev=0; lev<=finest_level; ++lev) {
@@ -1843,9 +1836,7 @@ Maestro::UpdateVel (const Vector<std::array< MultiFab, AMREX_SPACEDIM > >& umac,
     if (not_launched) Gpu::setLaunchRegion(true);
 #endif
 
-    Vector<MultiFab> w0_cart(finest_level+1);
     for (int lev=0; lev<=finest_level; ++lev) {
-        w0_cart[lev].define(grids[lev], dmap[lev], AMREX_SPACEDIM, 1);
         w0_cart[lev].setVal(0.);
     }
     Put1dArrayOnCart(w0,w0_cart,0,1,bcs_u,0,1);
