+++ Source/MaestroInit.cpp
@@ -11,294 +11,295 @@ using namespace amrex;
 void
 Maestro::Init ()
 {
-	// timer for profiling
-	BL_PROFILE_VAR("Maestro::Init()",Init);
+    // timer for profiling
+    BL_PROFILE_VAR("Maestro::Init()",Init);
 
-	Print() << "Calling Init()" << std::endl;
+    Print() << "Calling Init()" << std::endl;
 
-	if (restart_file == "") {
+    if (restart_file == "") {
 
-		start_step = 1;
+        start_step = 1;
 
-		// fill in multifab and base state data
-		InitData();
+        // fill in multifab and base state data
+        InitData();
 
-		if (plot_int > 0 || plot_deltat > 0) {
+        if (plot_int > 0 || plot_deltat > 0) {
 
-			// Need to fill normal vector to compute velrc in plotfile
-			if (spherical) { MakeNormal(); }
+            // Need to fill normal vector to compute velrc in plotfile
+            if (spherical) { MakeNormal(); }
 
-			Print() << "\nWriting plotfile "<< plot_base_name << "InitData after InitData" << std::endl;
-			WritePlotFile(plotInitData,t_old,0,rho0_old,rhoh0_old,p0_old,gamma1bar_old,uold,sold,S_cc_old);
+            Print() << "\nWriting plotfile "<< plot_base_name << "InitData after InitData" << std::endl;
+            WritePlotFile(plotInitData,t_old,0,rho0_old,rhoh0_old,p0_old,gamma1bar_old,uold,sold,S_cc_old);
 
-		} else if (small_plot_int > 0 || small_plot_deltat > 0) {
+        } else if (small_plot_int > 0 || small_plot_deltat > 0) {
 
-			// Need to fill normal vector to compute velrc in plotfile
-			if (spherical) { MakeNormal(); }
+            // Need to fill normal vector to compute velrc in plotfile
+            if (spherical) { MakeNormal(); }
 
-			Print() << "\nWriting small plotfile "<< small_plot_base_name << "InitData after InitData" << std::endl;
-			WriteSmallPlotFile(plotInitData,t_old,0,rho0_old,rhoh0_old,p0_old,gamma1bar_old,uold,sold,S_cc_old);
+            Print() << "\nWriting small plotfile "<< small_plot_base_name << "InitData after InitData" << std::endl;
+            WriteSmallPlotFile(plotInitData,t_old,0,rho0_old,rhoh0_old,p0_old,gamma1bar_old,uold,sold,S_cc_old);
 
-		}
-	}
-	else {
-		Print() << "Initializing from checkpoint " << restart_file << std::endl;
+        }
+    }
+    else {
+        Print() << "Initializing from checkpoint " << restart_file << std::endl;
 
-		// read in checkpoint file
-		// this builds (defines) and fills the following MultiFabs:
-		//
-		// snew, unew, gpi, dSdt, S_cc_new
-		//
-		// and also fills in the 1D arrays:
-		//
-		// rho0_new, p0_new, gamma1bar_new, rhoh0_new, beta0_new, psi, tempbar, etarho_cc, tempbar_init
-		ReadCheckPoint();
+        // read in checkpoint file
+        // this builds (defines) and fills the following MultiFabs:
+        //
+        // snew, unew, gpi, dSdt, S_cc_new
+        //
+        // and also fills in the 1D arrays:
+        //
+        // rho0_new, p0_new, gamma1bar_new, rhoh0_new, beta0_new, psi, tempbar, etarho_cc, tempbar_init
+        ReadCheckPoint();
 
[…]
