### 2019-09-19 032feac
only copy w0 onto w0_cart immediately after make_w0
