+++ Source/MaestroAdvanceAvg.cpp
@@ -38,7 +38,6 @@ Maestro::AdvanceTimeStepAverage (bool is_initIter) {
     Vector<MultiFab>   scal_force(finest_level+1);
     Vector<MultiFab>    delta_chi(finest_level+1);
     Vector<MultiFab>       sponge(finest_level+1);
-    Vector<MultiFab>         w0cc(finest_level+1);
 
     // face-centered in the dm-direction (planar only)
     Vector<MultiFab> etarhoflux_dummy(finest_level+1);
@@ -143,7 +142,6 @@ Maestro::AdvanceTimeStepAverage (bool is_initIter) {
 	}
 	delta_chi   [lev].define(grids[lev], dmap[lev],       1,    0);
 	sponge      [lev].define(grids[lev], dmap[lev],       1,    0);
-	w0cc     [lev].define(grids[lev], dmap[lev], AMREX_SPACEDIM, 0);
 
 	// face-centered in the dm-direction (planar only)
 	AMREX_D_TERM(etarhoflux_dummy[lev].define(convert(grids[lev],nodal_flag_x), dmap[lev], 1, 1); ,
@@ -306,6 +304,8 @@ Maestro::AdvanceTimeStepAverage (bool is_initIter) {
                     etarho_ec.dataPtr(),etarho_cc.dataPtr(),delta_chi_w0_dummy.dataPtr(),
                     r_cc_loc.dataPtr(),r_edge_loc.dataPtr(),&dt,&dtold,&is_predictor);
 
+            Put1dArrayOnCart(w0, w0_cart, 1, 1, bcs_u, 0, 1);
+        
             if (spherical == 1) {
                 // put w0 on Cartesian edges
                 MakeW0mac(w0mac);
@@ -641,6 +641,8 @@ Maestro::AdvanceTimeStepAverage (bool is_initIter) {
                     etarho_ec.dataPtr(),etarho_cc.dataPtr(),delta_chi_w0_dummy.dataPtr(),
                     r_cc_loc.dataPtr(),r_edge_loc.dataPtr(),&dt,&dtold,&is_predictor);
 
+            Put1dArrayOnCart(w0, w0_cart, 1, 1, bcs_u, 0, 1);
+            
             if (spherical == 1) {
                 // put w0 on Cartesian edges
                 MakeW0mac(w0mac);
@@ -895,10 +897,8 @@ Maestro::AdvanceTimeStepAverage (bool is_initIter) {
                     etarho_ec.dataPtr(),etarho_cc.dataPtr(),delta_chi_w0_dummy.dataPtr(),
                     r_cc_loc.dataPtr(),r_edge_loc.dataPtr(),&dt,&dtold,&is_predictor);
 
-            if (spherical == 1) {
-                // put w0 on Cartesian cell-centers
-                Put1dArrayOnCart(w0, w0cc, 1, 1, bcs_u, 0, 1);
-            }
+            Put1dArrayOnCart(w0, w0_cart, 1, 1, bcs_u, 0, 1);
+        
         }
     }
 
@@ -924,8 +924,8 @@ Maestro::AdvanceTimeStepAverage (bool is_initIter) {
     if (spherical == 1 && evolve_base_state && split_projection) {
 	// subtract w0 from uold and unew for nodal projection
 	for (int lev = 0; lev <= finest_level; ++lev) {
-	    MultiFab::Subtract(uold[lev],w0cc[lev],0,0,AMREX_SPACEDIM,0);
-	    MultiFab::Subtract(unew[lev],w0cc[lev],0,0,AMREX_SPACEDIM,0);
[…]
