+++ Source/MaestroDensityAdvance.cpp
@@ -452,7 +452,14 @@ Maestro::DensityAdvanceSDC (int which_step,
 	MultiFab::Add(scal_force[lev],intra[lev],FirstSpec,FirstSpec,NumSpec,0);
     }
 
+    Vector<MultiFab> p0_new_cart(finest_level+1);
+    for (int lev=0; lev<=finest_level; ++lev) {
+        p0_new_cart[lev].define(grids[lev], dmap[lev], 1, 1);
+    }
+
+    Put1dArrayOnCart(p0_new,p0_new_cart,0,0,bcs_f,0);
+
     // p0 only used in rhoh update so it's an optional parameter
-    UpdateScal(scalold, scalnew, sflux, scal_force, FirstSpec, NumSpec);
+    UpdateScal(scalold, scalnew, sflux, scal_force, FirstSpec, NumSpec, p0_new_cart);
     
 }
