+++ Source/MaestroPlot.cpp
@@ -238,278 +238,6 @@ Maestro::PlotFileMF (const int nPlot,
                      const RealVector& gamma1bar_in,
                      const Vector<MultiFab>& S_cc_in)
 {
-<<<<<<< HEAD
-	// timer for profiling
-	BL_PROFILE_VAR("Maestro::PlotFileMF()",PlotFileMF);
-
-	// MultiFab to hold plotfile data
-	Vector<const MultiFab*> plot_mf;
-
-	// temporary MultiFab to hold plotfile data
-	Vector<MultiFab*> plot_mf_data(finest_level+1);
-
-	// temporary MultiFab for calculations
-	Vector<MultiFab> tempmf(finest_level+1);
-	Vector<MultiFab> tempmf_scalar1(finest_level+1);
-	Vector<MultiFab> tempmf_scalar2(finest_level+1);
-	RealVector tempbar_plot ((max_radial_level+1)*nr_fine);
-	tempbar_plot.shrink_to_fit();
-	std::fill(tempbar_plot.begin(), tempbar_plot.end(), 0.);
-
-	int dest_comp = 0;
-
-	// build temporary MultiFab to hold plotfile data
-	for (int i = 0; i <= finest_level; ++i) {
-		plot_mf_data[i] = new MultiFab((s_in[i]).boxArray(),(s_in[i]).DistributionMap(),nPlot,0);
-		tempmf[i].define(grids[i],dmap[i],AMREX_SPACEDIM,0);
-
-		tempmf_scalar1[i].define(grids[i],dmap[i],1,0);
-		tempmf_scalar2[i].define(grids[i],dmap[i],1,0);
-	}
-
-	// velocity
-	for (int i = 0; i <= finest_level; ++i) {
-		plot_mf_data[i]->copy(u_in[i],0,dest_comp,AMREX_SPACEDIM);
-	}
-	dest_comp += AMREX_SPACEDIM;
-
-	// magvel
-	MakeMagvel(u_in, tempmf);
-	for (int i = 0; i <= finest_level; ++i) {
-		plot_mf_data[i]->copy(tempmf[i],0,dest_comp,1);
-	}
-	++dest_comp;
-
-	// momentum = magvel * rho
-	for (int i = 0; i <= finest_level; ++i) {
-		plot_mf_data[i]->copy(tempmf[i],0,dest_comp,1);
-		MultiFab::Multiply(*plot_mf_data[i], s_in[i], Rho, dest_comp, 1, 0);
-	}
-	++dest_comp;
-
-	// vorticity
-	MakeVorticity(u_in, tempmf);
-	for (int i = 0; i <= finest_level; ++i) {
-		plot_mf_data[i]->copy(tempmf[i],0,dest_comp,1);
-	}
-	++dest_comp;
-
-	// rho
-	for (int i = 0; i <= finest_level; ++i) {
-		plot_mf_data[i]->copy(s_in[i],Rho,dest_comp,1);
-	}
-	++dest_comp;
-
-	// rhoh
-	for (int i = 0; i <= finest_level; ++i) {
-		plot_mf_data[i]->copy(s_in[i],RhoH,dest_comp,1);
-	}
-	++dest_comp;
-
-	// h
-	for (int i = 0; i <= finest_level; ++i) {
-		plot_mf_data[i]->copy(s_in[i],RhoH,dest_comp,1);
-		MultiFab::Divide(*plot_mf_data[i], s_in[i], Rho, dest_comp, 1, 0);
-	}
-	++dest_comp;
-
-	// rhoX
-	for (int i = 0; i <= finest_level; ++i) {
[…]
