+++ Source/MaestroAdvanceSdc.cpp
@@ -196,14 +196,13 @@ Maestro::AdvanceTimeStepSDC (bool is_initIter) {
 	w0mac_dummy[lev][1].define(convert(grids[lev],nodal_flag_y), dmap[lev], 1, 1);
 	w0mac_dummy[lev][2].define(convert(grids[lev],nodal_flag_z), dmap[lev], 1, 1);
     }
-    if (spherical == 1) {
-	for (int lev=0; lev<=finest_level; ++lev) {
-	    w0_force_cart_dummy[lev].define(grids[lev], dmap[lev], AMREX_SPACEDIM, 1);
-	}
-    }
 #endif
-
-
+    
+    for (int lev=0; lev<=finest_level; ++lev) {
+	w0_force_cart_dummy[lev].define(grids[lev], dmap[lev], AMREX_SPACEDIM, 1);
+	w0_force_cart_dummy[lev].setVal(0.);
+    }
+    
     // set etarhoflux_dummy to zero
     for (int lev=0; lev<=finest_level; ++lev) {
 	etarhoflux_dummy[lev].setVal(0.);
@@ -215,11 +214,6 @@ Maestro::AdvanceTimeStepSDC (bool is_initIter) {
 	for (int d=0; d<AMREX_SPACEDIM; ++d) {
 	    w0mac[lev][d].setVal(0.);
 	    w0mac_dummy[lev][d].setVal(0.);
-	}
-    }
-    if (spherical == 1) {
-	for (int lev=0; lev<=finest_level; ++lev) {
-	    w0_force_cart_dummy[lev].setVal(0.);
 	}
     }
 #endif
