+++ Source/MaestroEnthalpyAdvance.cpp
@@ -610,13 +610,11 @@ Maestro::EnthalpyAdvanceSDC (int which_step,
     }
     
     Vector<MultiFab> p0_new_cart(finest_level+1);
-    if (spherical == 1) {
-        for (int lev=0; lev<=finest_level; ++lev) {
-            p0_new_cart[lev].define(grids[lev], dmap[lev], 1, 1);
-        }
-
-        Put1dArrayOnCart(p0_new,p0_new_cart,0,0,bcs_f,0);
+    for (int lev=0; lev<=finest_level; ++lev) {
+	p0_new_cart[lev].define(grids[lev], dmap[lev], 1, 1);
     }
+
+    Put1dArrayOnCart(p0_new,p0_new_cart,0,0,bcs_f,0);
 
     UpdateScal(scalold, scalnew, sflux, scal_force, RhoH, 1, p0_new_cart);
 }
