### 2019-09-20 5330d0e
declare w0_force_cart in non-spherical cases as well
