### 2019-09-20 febd745
corner ghost cell issue fixed
