+++ Source/MaestroAdvection.cpp
@@ -24,9 +24,11 @@ Maestro::AdvancePremac (Vector<std::array< MultiFab, AMREX_SPACEDIM > >& umac,
 	FillPatch(t_new, utilde, uold, uold, 0, 0, AMREX_SPACEDIM, 0, bcs_u, 1);
 
 	// create a MultiFab to hold uold + w0
-	Vector<MultiFab>      ufull(finest_level+1);
+	Vector<MultiFab> ufull(finest_level+1);
 	for (int lev=0; lev<=finest_level; ++lev) {
 		ufull[lev].define(grids[lev], dmap[lev], AMREX_SPACEDIM, ng_adv);
+                // needed to avoid NaNs in filling corner ghost cells with 2 physical boundaries
+                ufull[lev].setVal(0.);
 	}
 
 	// create ufull = uold + w0
