### 2019-09-21 a687137
GPU make edge scal 2d (GPU hydro PR 6/n) (#81)
This offloads make_edge_scal_2d to the GPU. It has been tested for the 2d reacting bubble problem compiled for MPI and GPU.

Note: It is likely that this version will not work if compiled with OpenMP (this will come in a future PR)
