+++ Source/Maestro_F.H
@@ -441,30 +441,53 @@ extern "C"
 
     //////////////////////
     // in make_edge_scal.F90
-#if (AMREX_SPACEDIM == 2)
-    void make_edge_scal_2d(
-#elif (AMREX_SPACEDIM == 3)
+#if (AMREX_SPACEDIM == 3)
     void make_edge_scal_3d(
-#endif
-                           const int* domlo, const int* domhi,
-                           const int* lo, const int* hi,
-                           const amrex::Real* s,      const int* s_lo, const int* s_hi, const int nc_s, int ng_s,
-                           amrex::Real* sedgex, const int* x_lo, const int* x_hi, const int nc_x,
-                           amrex::Real* sedgey, const int* y_lo, const int* y_hi, const int nc_y,
-#if (AMREX_SPACEDIM == 3)
-                           amrex::Real* sedgez, const int* z_lo, const int* z_hi, const int nc_z,
-#endif
-                           const amrex::Real* umac,   const int* u_lo, const int* u_hi,
-                           const amrex::Real* vmac,   const int* v_lo, const int* v_hi,
-#if (AMREX_SPACEDIM == 3)
-                           const amrex::Real* wmac,   const int* w_lo, const int* w_hi,
-#endif
-                           const int ng_um,
-                           const amrex::Real* force,  const int* f_lo, const int* f_hi, const int nc_f,
-                           const amrex::Real* dx, const amrex::Real dt, const int is_vel,
-                           const int* adv_bc, const int nbccomp, const int comp, const int bccomp,
-                           const int is_conservative);
+                      const int* domlo, const int* domhi,
+                      const int* lo, const int* hi,
+                      const amrex::Real* s,      const int* s_lo, const int* s_hi, const int nc_s, int ng_s,
+                            amrex::Real* sedgex, const int* x_lo, const int* x_hi, const int nc_x,
+                            amrex::Real* sedgey, const int* y_lo, const int* y_hi, const int nc_y,
+                            amrex::Real* sedgez, const int* z_lo, const int* z_hi, const int nc_z,
+                      const amrex::Real* umac,   const int* u_lo, const int* u_hi,
+                      const amrex::Real* vmac,   const int* v_lo, const int* v_hi,
+                      const amrex::Real* wmac,   const int* w_lo, const int* w_hi,
+                      const int ng_um,
+                      const amrex::Real* force,  const int* f_lo, const int* f_hi, const int nc_f,
+                      const amrex::Real* dx, const amrex::Real dt, const int is_vel,
+                      const int* adv_bc, const int nbccomp, const int comp, const int bccomp,
+                      const int is_conservative);
 
+#elif (AMREX_SPACEDIM == 2)
+  void make_edge_scal_predictor_2d(const int* lo, const int* hi, const int idir,
+                    const int* domlo, const int* domhi,
+                    const amrex::Real* s, const int* s_lo, const int* s_hi, const int nc_s, int ng_s,
+                    const amrex::Real* umac,   const int* u_lo, const int* u_hi,
+                    const amrex::Real* vmac,   const int* v_lo, const int* v_hi,
+                    const amrex::Real* Ip, const int* ip_lo, const int* ip_hi,
+                    const amrex::Real* Im, const int* im_lo, const int* im_hi,
+                    amrex::Real* sl, const int* sl_lo, const int* sl_hi,
+                    amrex::Real* sr, const int* sr_lo, const int* sr_hi,
+                    amrex::Real* simh, const int* si_lo, const int* si_hi,
+                    const amrex::Real* dx, const amrex::Real dt, const int is_vel,
[…]
