+++ Source/Src_nd/make_edge_scal.F90
@@ -26,466 +26,438 @@ module make_edge_scal_module
 contains
 
 #if (AMREX_SPACEDIM == 2)
-  subroutine make_edge_scal_2d(domlo, domhi, lo, hi, &
-       s,      s_lo, s_hi, nc_s, ng_s, &
-       sedgex, x_lo, x_hi, nc_x, &
-       sedgey, y_lo, y_hi, nc_y, &
-       umac,   u_lo, u_hi, &
-       vmac,   v_lo, v_hi, ng_um, &
-       force,  f_lo, f_hi, nc_f, &
-       dx, dt, is_vel, adv_bc, nbccomp, &
-       comp, bccomp, is_conservative) bind(C,name="make_edge_scal_2d")
+subroutine make_edge_scal_predictor_2d(lo, hi, idir, domlo, domhi, &
+     s,      s_lo, s_hi, nc_s, ng_s, &
+     umac,   u_lo, u_hi, &
+     vmac,   v_lo, v_hi,&
+     Ip, ip_lo, ip_hi, &
+     Im, im_lo, im_hi, &
+     sl, sl_lo, sl_hi, &
+     sr, sr_lo, sr_hi, &
+     simh, si_lo, si_hi, &
+     dx, dt, is_vel, adv_bc, nbccomp, &
+     comp, bccomp) bind(C,name="make_edge_scal_predictor_2d")
 
-    integer         , intent(in   ) :: domlo(3), domhi(3), lo(3), hi(3)
-    integer         , intent(in   ) :: s_lo(3), s_hi(3)
-    integer, value,   intent(in   ) :: nc_s,ng_s
-    integer         , intent(in   ) :: x_lo(3), x_hi(3)
-    integer, value,   intent(in   ) :: nc_x
-    integer         , intent(in   ) :: y_lo(3), y_hi(3)
-    integer, value,   intent(in   ) :: nc_y
-    integer         , intent(in   ) :: u_lo(3), u_hi(3)
-    integer         , intent(in   ) :: v_lo(3), v_hi(3)
-    integer, value,   intent(in   ) :: ng_um
-    integer         , intent(in   ) :: f_lo(3), f_hi(3)
-    integer, value,   intent(in   ) :: nc_f
-    double precision, intent(in   ) :: s     (s_lo(1):s_hi(1),s_lo(2):s_hi(2),s_lo(3):s_hi(3),1:nc_s)
-    double precision, intent(inout) :: sedgex(x_lo(1):x_hi(1),x_lo(2):x_hi(2),x_lo(3):x_hi(3),1:nc_x)
-    double precision, intent(inout) :: sedgey(y_lo(1):y_hi(1),y_lo(2):y_hi(2),y_lo(3):y_hi(3),1:nc_y)
-    double precision, intent(in   ) :: umac  (u_lo(1):u_hi(1),u_lo(2):u_hi(2),u_lo(3):u_hi(3))
-    double precision, intent(in   ) :: vmac  (v_lo(1):v_hi(1),v_lo(2):v_hi(2),v_lo(3):v_hi(3))
-    double precision, intent(in   ) :: force (f_lo(1):f_hi(1),f_lo(2):f_hi(2),f_lo(3):f_hi(3),nc_f)
-    double precision, intent(in   ) :: dx(3)
-    double precision, value, intent(in   ) :: dt
-    integer, value, intent(in   ) :: is_vel, nbccomp, comp, bccomp, is_conservative
-    integer         , intent(in   ) :: adv_bc(2,2,nbccomp)
[…]

[…]
