+++ docs/_sources/pert.rst.txt
@@ -87,18 +87,20 @@ Basic equations
 The full density evolution equation is
 
 .. math::
-
    \begin{aligned}
    \frac{\partial\rho}{\partial t} &=& -\nabla\cdot(\rho\Ub) \nonumber \\
-   &=& -\Ub\cdot\nabla\rho - \rho\nabla\cdot\Ub \, . \label{rho equation}\end{aligned}
+   &=& -\Ub\cdot\nabla\rho - \rho\nabla\cdot\Ub \, . 
+   \end{aligned}
+   :label: rho equation
 
 The species are evolved according to
 
 .. math::
-
    \begin{aligned}
    \frac{\partial(\rho X_k)}{\partial t} &=& -\nabla\cdot(\rho\Ub X_k) + \rho \omegadot_k \nonumber \\
-   &=& -\Ub\cdot\nabla(\rho X_k) - \rho X_k \nabla\cdot\Ub + \rho \omegadot_k \, . \label{species equation}\end{aligned}
+   &=& -\Ub\cdot\nabla(\rho X_k) - \rho X_k \nabla\cdot\Ub + \rho \omegadot_k \, .
+   \end{aligned}
+   :label: species equation
 
 In practice, only the species evolution equation is evolved, and the
 total density is set as
@@ -114,12 +116,10 @@ subscript ‘edge’ to indicate what quantity was *predicted* to the
 edges. In MAESTROeX, the different methods of computing :math:`(\rho X)` on
 edges are controlled by the ``species_pred_type parameter``. The
 quantities predicted to edges and the
-resulting edge state are shown in the table \ `[table:pred:species] <#table:pred:species>`__
+resulting edge state are shown in the :numref:`table:pred:species`.
 
-.. raw:: latex
 
-   \centering
-
+.. _table:pred:species:
 .. table:: Summary of species edge state construction
 
    +--------------------------+---------------------------------------+---------------------------------------------------------------------------------------------+
@@ -150,13 +150,12 @@ We predict both :math:`\rho'` and :math:`\rho_0` to edges separately and later u
 reconstruct :math:`\rho` at edges. The base state density evolution equation is
 
 .. math::
-
    \frac{\partial\rho_0}{\partial t} = -\nabla\cdot(\rho_0 w_0 \eb_r) =
    -w_0\frac{\partial\rho_0}{\partial r}
    \underbrace{-\rho_0\frac{\partial w_0}{\partial r}}_{``\rho_0 ~ \text{force"}}.
-   \label{rho0 equation}
+   :label: rho0 equation
 
-Subtract (`[rho0 equation] <#rho0 equation>`__) from (`[rho equation] <#rho equation>`__) and rearrange
+Subtract :eq:`rho0 equation` from :eq:`rho equation` and rearrange
 terms, noting that :math:`\Ub = \Ubt + w_o\eb_r`, to obtain the
 perturbational density equation,
 
@@ -167,7 +166,7 @@ perturbational density equation,
    \label{rhoprime equation}
 
 We also need :math:`X_k` at the edges. Here, we subtract :math:`X_k \times`
-Eq. \ `[rho equation] <#rho equation>`__ from Eq. \ `[species equation] <#species equation>`__ to obtain
+:eq:`rho equation` from :eq:`species equation` to obtain
 
 .. math:: \frac{\partial X_k}{\partial t} = -\Ub \cdot \nabla X_k + \omegadot_k
 
@@ -191,7 +190,7 @@ Predicting :math:`\rho_0` at edges
 There are two ways to predict :math:`\rho_0` at edges.
 
 #. We call make_edge_state_1d using the underbraced term
[…]
