+++ docs/_sources/thermo_notes.rst.txt
@@ -32,21 +32,20 @@ system:
 where we have defined the mass fraction, :math:`X_i = \frac{\rho_i}{\rho}`. Note
 
 .. math::
-
-   \label{eq:mass fraction sums to 1}
-     \sum_i X_i = 1.
+   \sum_i X_i = 1.
+   :label: eq:mass fraction sums to 1
 
 We write the average molar mass and average proton number as:
 
 .. math::
+   \bar{A} = \frac{\sum_i A_i n_i}{\sum_i n_i} = \left(\sum_i X_i\right)
+     \left(\sum_i \frac{X_i}{A_i}\right)^{-1}
+   :label: eq:abar
 
-   \begin{aligned}
-   \label{eq:abar}
-     \bar{A} &= \frac{\sum_i A_i n_i}{\sum_i n_i} = \left(\sum_i X_i\right)
-     \left(\sum_i \frac{X_i}{A_i}\right)^{-1}\\
-     \label{eq:zbar}
-     \bar{Z} &= \frac{\sum_i Z_i n_i}{\sum_i n_i} = \left(\sum_i Z_i
-     \frac{X_i}{A_i}\right)\left(\sum_i \frac{X_i}{A_i}\right)^{-1}.\end{aligned}
+.. math::
+   \bar{Z} &= \frac{\sum_i Z_i n_i}{\sum_i n_i} = \left(\sum_i Z_i
+     \frac{X_i}{A_i}\right)\left(\sum_i \frac{X_i}{A_i}\right)^{-1}.
+   :label: eq:zbar
 
 Our algorithm requires terms involving the derivative thermodynamic variables
 (:math:`p` or :math:`e`, e.g.) with respect to composition. Our EOS does not return such
@@ -60,21 +59,21 @@ to :math:`\bar{A}` and :math:`\bar{Z}`. Using the chain rule, we have
      \frac{\ptl p}{\ptl \bar{A}}\frac{\ptl \bar{A}}{\ptl X_k} + 
      \frac{\ptl p}{\ptl \bar{Z}}\frac{\ptl \bar{Z}}{\ptl X_k}.
 
-From `[eq:abar] <#eq:abar>`__ and `[eq:zbar] <#eq:zbar>`__ we have
+From :eq:`eq:abar` and :eq:`eq:zbar` we have
 
 .. math::
+   \frac{\ptl \bar{A}}{\ptl X_k} = \left(\sum_i \frac{X_i}{A_i}\right)^{-1}
+     - \frac{\bar{A}^2}{A_k} = -\frac{\bar{A}}{A_k}\left(\bar{A} - A_k\right)
+   :label: eq:abar_X_k
 
-   \begin{aligned}
-   \label{eq:abar_X_k}
-     \frac{\ptl \bar{A}}{\ptl X_k} &= \left(\sum_i \frac{X_i}{A_i}\right)^{-1}
-     - \frac{\bar{A}^2}{A_k} = -\frac{\bar{A}}{A_k}\left(\bar{A} - A_k\right)\\
-     \label{eq:zbar_X_k}
-     \frac{\ptl \bar{Z}}{\ptl X_k} &= 
+.. math::
+   \frac{\ptl \bar{Z}}{\ptl X_k} = 
      \left(\frac{Z_k}{A_k}\right)\left(\sum_i \frac{X_i}{A_i}\right)^{-1}
      - \frac{\bar{Z}}{A_k}\left(\sum_i \frac{X_i}{A_i}\right)^{-1} = 
-     -\frac{\bar{A}}{A_k}\left(\bar{Z} - Z_k\right),\end{aligned}
+     -\frac{\bar{A}}{A_k}\left(\bar{Z} - Z_k\right),
+   :label: eq:zbar_X_k
 
[…]
