### 2019-09-22 a5430ef
a bunch more reference fixes
