+++ docs/_sources/pert.rst.txt
@@ -892,175 +892,95 @@ amount of mixing.
 Initial Observations
 --------------------
 
-With ``use_tfromp = T`` and ``cflfac = 0.7`` there is a large difference
-between ``species_pred_type = 1`` and species_pred_type = 2,3 as
-seen in Figure `[fig:spec1_vs_23] <#fig:spec1_vs_23>`__. ``species_pred_type = 1`` shows
-quick heating (peak T vs. t) and there is ok agreement between ``tfromh``
-and ``tfromp``. ``species_pred_type = 2,3`` show cooling (peak T vs. t)
-and ``tfromh`` looks completely unphysical (see Figure
-`[fig:tfromh_unphysical] <#fig:tfromh_unphysical>`__). There are also strange filament type features in
-the momentum plots shown in Figure `[fig:mom_filaments] <#fig:mom_filaments>`__.
+With ``use_tfromp = T`` and ``cflfac = 0.7`` there is a large
+difference between ``species_pred_type = 1`` and species_pred_type =
+2,3 as seen in :numref:`fig:spec1_vs_23``. ``species_pred_type = 1``
+shows quick heating (peak T vs. t) and there is ok agreement between
+``tfromh`` and ``tfromp``. ``species_pred_type = 2,3`` show cooling
+(peak T vs. t) and ``tfromh`` looks completely unphysical (see
+:numref:`fig:tfromh_unphysical`). There are also strange filament type
+features in the momentum plots shown in :numref:`fig:mom_filaments`.
 
-.. raw:: latex
 
-   \centering
-
-.. figure:: \pertfigpath/spec1_vs_23
-   :alt: Compare species_pred_type = 1,2,3 with use_tfromp =
-   T, enthalpy_pred_type = 1, cflfac = 0.7
+.. _fig:spec1_vs_23:
+.. figure:: spec1_vs_23.png
+   :align: center
 
    Compare species_pred_type = 1,2,3 with use_tfromp =
    T, enthalpy_pred_type = 1, cflfac = 0.7
 
-[fig:spec1_vs_23]
 
-.. raw:: latex
-
-   \vspace{0pt}
-
-.. raw:: latex
-
-   \centering
-
-.. figure:: \pertfigpath/tfromh_unphysical
-   :alt: There are strange filament type features at the bottom of the
-   domain. ``species_pred_type = 2``, ``enthalpy_pred_type = 1``, ``cflfac = 0.7``,
-   ``use_tfromp = T``
+.. _fig:tfromh_unphysical:
+.. figure:: tfromh_unphysical.png
+   :align: center
 
    There are strange filament type features at the bottom of the
    domain. ``species_pred_type = 2``, ``enthalpy_pred_type = 1``, ``cflfac = 0.7``,
    ``use_tfromp = T``
 
-[fig:tfromh_unphysical]
-
-.. raw:: latex
-
-   \hspace{0.5cm}
-
-.. raw:: latex
-
-   \vspace{0pt}
-
-.. raw:: latex
-
-   \centering
-
-.. figure:: \pertfigpath/mom_filaments
-   :alt: There are strange filament type features at the bottom of the
-   domain. ``species_pred_type = 2``, ``enthalpy_pred_type = 1``, ``cflfac = 0.7``,
-   ``use_tfromp = T``
+.. _fig:mom_filaments:
+.. figure:: mom_filaments.png
+   :align: center
 
    There are strange filament type features at the bottom of the
[…]
