+++ sphinx_docs/source/Godunov.rst
@@ -56,9 +56,6 @@ For spherical problems, we compute the radial bin centered gradient using
 Then we put :math:`\partial w_0/\partial r` onto a Cartesian grid
 using put_1d_array_on_cart_3d_sphr.
 
-.. raw:: latex
-
-   \newpage
 
 Computing :math:`\Ubt^{\trans}` in MAESTROeX
 ============================================
@@ -89,9 +86,6 @@ Computing :math:`\Ubt^{\trans}` in MAESTROeX
 
   .. math:: \ut_{R,\ib-\half\eb_x} = \ut_{\ib}^n + \left[\frac{1}{2} - \frac{\dt}{2h}\min(0,\ut_{\ib}^n)\right]\partial\ut_{\ib}^n
 
-.. raw:: latex
-
-   \newpage
 
 2D Cartesian Case
 -----------------
@@ -135,9 +129,6 @@ that we upwind based on the full velocity.
    \wt_{R,\ib-\half\eb_r}, & w_{L,\ib-\half\eb_r} + w_{R,\ib-\half\eb_r} < 0, \\
    \end{cases}
 
-.. raw:: latex
-
-   \newpage
 
 .. _d-cartesian-case-1:
 
@@ -164,9 +155,6 @@ computing :math:`\ut^{\trans}`. We predict :math:`\vt` to y-faces using the
    \vt_{R,\ib-\half\eb_y}, & v_{L,\ib-\half\eb_y} + v_{R,\ib-\half\eb_y} < 0. \\
    \end{cases}
 
-.. raw:: latex
-
-   \newpage
 
 3D Spherical Case
 -----------------
@@ -179,9 +167,6 @@ that the advection velocities, as well as
 the upwind velocity, is done with the full velocity, not the
 perturbational velocity.
 
-.. raw:: latex
-
-   \newpage
 
 Computing :math:`\Ubt^{\mac,*}` in MAESTROeX
 ============================================
@@ -201,9 +186,6 @@ Computing :math:`\Ubt^{\mac,*}` in MAESTROeX
   (:math:`\ut_{\ib-\half\eb_x}^{\trans}`, :math:`\vt_{\ib-\half\eb_y}^{\trans}`,
   and :math:`\wt_{\ib-\half\eb_r}^{\trans}`)
 
-.. raw:: latex
-
-   \newpage
 
 .. _d-cartesian-case-2:
 
@@ -298,9 +280,6 @@ Solve a Riemann problem:
    \wt_{R,\ib-\half\eb_r}^{\mac,*}, & w_L^{\mac,*} + w_R^{\mac,*} < 0.
    \end{cases}
 
-.. raw:: latex
-
-   \newpage
 
 .. _d-cartesian-case-3:
 
@@ -429,9 +408,6 @@ the effects of corner coupling.
      \wt_{R,\ib-\half\eb_r}^{\mac,*} &=& \wt_{R,\ib-\half\eb_r}^{\mac,*} -
      \frac{\dt}{4h}\left(\wt^{\trans}_{\ib-\half\eb_r} + \wt^{\trans}_{\ib-\frac{3}{2}\eb_r}\right)\left(w_{0,\ib-\half\eb_r}-w_{0,\ib-\frac{3}{2}\eb_r}\right)\end{aligned}
 
-.. raw:: latex
-
-   \newpage
 
 .. _d-spherical-case-1:
 
@@ -444,9 +420,6 @@ in the full dimensional extrapolation for each direction.
[…]
