+++ docs/_sources/pert.rst.txt
@@ -424,7 +424,7 @@ depend on ``species_pred_type``.
 
 The general procedure for making the :math:`(\rho h)` edge state is as follows:
 
-#. predict :math:`(rho h)`, :math:`(\rho h)'`, :math:`h`, or :math:`T`
+#. predict :math:`(\rho h)`, :math:`(\rho h)'`, :math:`h`, or :math:`T`
    to the edges (depending on ``enthalpy_pred_type`` ) using
    ``make_edge_scal`` and the forces identified in the appropriate
    evolution equation (:eq:`rhohprime equation`, :eq:`T equation
@@ -455,210 +455,193 @@ it through the velocity field and :math:`S`). As a result, the
 not work. MAESTROeX will abort if the code is run with this
 combination of parameters.
 
-:numref:`table:pred:hoverview` gives a summary of the
-``enthalpy_pred_type`` behavior.
+The behavior of ``enthalpy_pred_type`` is:
 
-.. _table:pred:hforce:
-.. table:: Forcing term into ``make_edge_scal``
+* ``enthalpy_pred_type = 0`` (``predict_rhoh``) :
+  we predict :math:`(\rho h)`.  The advective force is:
 
-   +-----------------------------------+-----------------------------------+
-   | enthalpy_pred_type                | advective force                   |
-   +===================================+===================================+
-   | 0 / predict_rhoh :math:`(\rho h)` | :math:`\left [\psi + (\Ubt \cdot  |
-   |                                   | \er)                              |
-   |                                   |   \frac{\partial p_0}{\partial r} |
-   |                                   |  + \nabla \cdot \kth \nabla T \ri |
-   |                                   | ght ]`                            |
-   +-----------------------------------+-----------------------------------+
-   | 1 / predict_rhohprime             | :math:`-(\rho h)^\prime \; \nabla |
-   | :math:`((\rho h)^\prime)`         |  \cdot (\Ubt+w_0 \er) -           |
-   |                                   |  \nabla \cdot (\Ubt (\rho h)_0) + |
-   |                                   |  (\Ubt \cdot \er) \frac{\partial  |
-   |                                   | p_0}{\partial r} + \nabla \cdot \ |
-   |                                   | kth \nabla T`                     |
-   +-----------------------------------+-----------------------------------+
-   | 2 / predict_h :math:`(h)`         | :math:`\frac{1}{\rho} \left [\psi |
-   |                                   |  + (\Ubt \cdot \er)               |
-   |                                   |   \frac{\partial p_0}{\partial r} |
-   |                                   |  + \nabla \cdot \kth \nabla T \ri |
-   |                                   | ght ]`                            |
-   +-----------------------------------+-----------------------------------+
-   | 3 / predict_T_then_rhohprime      | :math:`\frac{1}{\rho c_p} \left \ |
-   | :math:`(T)`                       | { (1 - \rho h_p)                  |
-   |                                   |    \left [\psi + (\Ubt \cdot \er) |
-   |                                   |  \frac{\partial p_0}{\partial r}  |
-   |                                   | \right ] + \nabla \cdot \kth \nab |
-   |                                   | la T \right \}`                   |
-   +-----------------------------------+-----------------------------------+
-   | 4 / predict_T_then_h :math:`(T)`  | :math:`\frac{1}{\rho c_p} \left\{ |
-   |                                   |  (1 - \rho h_p) \left [\psi + (\U |
-   |                                   | bt \cdot \er)                     |
-   |                                   | \frac{\partial p_0}{\partial r}\r |
[…]
