### 2019-09-23 #1757 @arfon
@tukss, @pgrete - thanks again for agreeing to review this submission. Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
