### 2019-09-23 6fbdda3
fixed GPU compilation
