+++ Source/MaestroSetup.cpp
@@ -294,34 +294,6 @@ void Maestro::VariableSetup ()
     if (ppm_type == 2 || bds_type == 1) {
         ng_s = 4;
     }
-
-#ifdef AMREX_USE_CUDA
-  // Set the minimum number of threads needed per
-  // threadblock to do BC fills with CUDA. We will
-  // force this to be 8. The reason is that it is
-  // not otherwise guaranteed for our thread blocks
-  // to be aligned with the grid in such a way that
-  // the synchronization logic in amrex_filccn works
-  // out. We need at least NUM_GROW + 1 threads in a
-  // block for CTU. If we used this minimum of 5, we
-  // would hit cases where this doesn't work since
-  // our blocking_factor is usually a power of 2, and
-  // the thread blocks would not be aligned to guarantee
-  // that the threadblocks containing the ghost zones
-  // contained all of the ghost zones, as well as the
-  // required interior zone. And for reflecting BCs,
-  // we need NUM_GROW * 2 == 8 threads anyway. This logic
-  // then requires that blocking_factor be a multiple
-  // of 8. It is a little wasteful for MOL/SDC and for
-  // problems that only have outflow BCs, but the BC
-  // fill is not the expensive part of the algorithm
-  // for our production science problems anyway, so
-  // we ignore this extra cost in favor of safety.
-
-  for (int dim = 0; dim < AMREX_SPACEDIM; ++dim) {
-      numBCThreadsMin[dim] = 8;
-  }
-#endif
 }
 
 void
