### 2019-09-23 ba05907
RealVector is now an alias for amrex::GPU:ManagedVector<amrex::Real>
