+++ Source/MaestroGamma.cpp
@@ -6,7 +6,7 @@ using namespace amrex;
 void
 Maestro::MakeGamma1bar (const Vector<MultiFab>& scal,
                         RealVector& gamma1bar,
-                        const RealVector& p0)
+                        const Vector<MultiFab>& p0_cart)
 {
     // timer for profiling
     BL_PROFILE_VAR("Maestro::MakeGamma1bar()",MakeGamma1bar);
@@ -18,17 +18,17 @@ Maestro::MakeGamma1bar (const Vector<MultiFab>& scal,
 #endif
 
     Vector<MultiFab> gamma1(finest_level+1);
-    Vector<MultiFab> p0_cart(finest_level+1);
+    // Vector<MultiFab> p0_cart(finest_level+1);
 
     for (int lev=0; lev<=finest_level; ++lev) {
         gamma1[lev].define(grids[lev], dmap[lev], 1, 1);
         gamma1[lev].setVal(0.);
 
-        p0_cart[lev].define(grids[lev], dmap[lev], 1, 1);
-        p0_cart[lev].setVal(0.);
+        // p0_cart[lev].define(grids[lev], dmap[lev], 1, 1);
+        // p0_cart[lev].setVal(0.);
     }
 
-    Put1dArrayOnCart(p0,p0_cart,0,0,bcs_f,0);
+    // Put1dArrayOnCart(p0,p0_cart,0,0,bcs_f,0);
 
     for (int lev=0; lev<=finest_level; ++lev) {
 
