### 2019-09-24 38ea7d4
some doc fixes
