### 2019-09-24 394d91f
think I've got rid of the roundoff errors
