### 2019-09-26 5171fad
enthalpy_pred_type=predict_rhohprime and species_pred_type=predict_rhoprime_and_X now works with sdc when evolve_base_state=F
