+++ Source/MaestroAdvanceSdc.cpp
@@ -28,6 +28,7 @@ Maestro::AdvanceTimeStepSDC (bool is_initIter) {
     Vector<MultiFab>       rho_Hext(finest_level+1);
     Vector<MultiFab>     sdc_source(finest_level+1);
     Vector<MultiFab>           aofs(finest_level+1);
+    Vector<MultiFab>    intra_rhoh0(finest_level+1);
     
     Vector<MultiFab> delta_gamma1_term(finest_level+1);
     Vector<MultiFab>      delta_gamma1(finest_level+1);
@@ -83,6 +84,7 @@ Maestro::AdvanceTimeStepSDC (bool is_initIter) {
     RealVector     gamma1bar_nph    ( (max_radial_level+1)*nr_fine );
     RealVector delta_gamma1_termbar ( (max_radial_level+1)*nr_fine );
     RealVector delta_chi_w0_dummy   ( (max_radial_level+1)*nr_fine );
+    RealVector     delta_rhoh0      ( (max_radial_level+1)*nr_fine );
 
     // vectors store the multilevel 1D states as one very long array
     // these are edge-centered
@@ -103,6 +105,7 @@ Maestro::AdvanceTimeStepSDC (bool is_initIter) {
     delta_gamma1_termbar.shrink_to_fit();
     w0_old.shrink_to_fit();
     delta_chi_w0_dummy.shrink_to_fit();
+    delta_rhoh0.shrink_to_fit();
 
     int is_predictor;
 
@@ -113,9 +116,7 @@ Maestro::AdvanceTimeStepSDC (bool is_initIter) {
 
     if (evolve_base_state) {
 	Abort("evolve_base_state not supported with SDC");
-    } else if (enthalpy_pred_type == 1) {
-	Abort("enthalpy_pred_type == 1 not supported with SDC");
-    }
+    } 
 
     if (maestro_verbose > 0) {
     	Print() << "Cell Count:" << std::endl;
@@ -143,6 +144,7 @@ Maestro::AdvanceTimeStepSDC (bool is_initIter) {
 	rho_Hext    [lev].define(grids[lev], dmap[lev],       1,    0);
 	sdc_source  [lev].define(grids[lev], dmap[lev],   Nscal,    0);
 	aofs        [lev].define(grids[lev], dmap[lev],   Nscal,    0);
+	intra_rhoh0 [lev].define(grids[lev], dmap[lev],       1,    0);
 	delta_gamma1_term[lev].define(grids[lev], dmap[lev],  1,    0);
 	delta_gamma1[lev].define(grids[lev], dmap[lev],       1,    0);
         peosbar_cart[lev].define(grids[lev], dmap[lev],       1,    0);
@@ -185,6 +187,9 @@ Maestro::AdvanceTimeStepSDC (bool is_initIter) {
 	// initialize umac
 	for (int d=0; d < AMREX_SPACEDIM; ++d)
 	    umac[lev][d].setVal(0.);
+
+	// initialize intra_rhoh0
+	intra_rhoh0[lev].setVal(0.);
     }
 
 #if (AMREX_SPACEDIM == 3)
@@ -478,8 +483,7 @@ Maestro::AdvanceTimeStepSDC (bool is_initIter) {
     // base state enthalpy update
     if (evolve_base_state) {
 	// compute rhoh0_old by "averaging"
-	Average(s1, rhoh0_old, RhoH);
-	Average(s2, rhoh0_new, RhoH); // -> rhoh0_new = rhoh0_old (bad?)
+	Average(sold, rhoh0_old, RhoH);
     }
[…]
