### 2019-09-27 df8ef62
bug fix: set sdc_off parameter to false
