### 2019-09-27 cdf18dc
Changed make_flux to use full umac velocities instead of split umac and w0mac. This is done so that the split mac velocities are only used during projection steps
