+++ Source/MaestroDensityAdvance.cpp
@@ -50,6 +50,16 @@ Maestro::DensityAdvance (int which_step,
 
     Put1dArrayOnCart(rho0_old,rho0_old_cart,0,0,bcs_s,Rho);
 
+    /////////////////////////////////////////////////////////////////
+    // Subtract w0 from MAC velocities (MAC velocities has w0 already).
+    /////////////////////////////////////////////////////////////////
+
+    Addw0(umac,w0mac,-1.);
+
+    /////////////////////////////////////////////////////////////////
+    // Compute source terms
+    /////////////////////////////////////////////////////////////////
+    
     // ** density source term **
 
     // Make source term for rho or rho'
@@ -75,7 +85,7 @@ Maestro::DensityAdvance (int which_step,
 
 
     /////////////////////////////////////////////////////////////////
-    // Add w0 to MAC velocities (trans velocities already have w0).
+    // Add w0 back to MAC velocities (trans velocities already have w0).
     /////////////////////////////////////////////////////////////////
 
     Addw0(umac,w0mac,1.);
@@ -143,14 +153,6 @@ Maestro::DensityAdvance (int which_step,
         // convert X --> (rho X) in scalold
         ConvertRhoXToX(scalold,false);
     }
-
-
-    /////////////////////////////////////////////////////////////////
-    // Subtract w0 from MAC velocities.
-    /////////////////////////////////////////////////////////////////
-
-    Addw0(umac,w0mac,-1.);
-
 
     /////////////////////////////////////////////////////////////////
     // Compute fluxes
@@ -285,6 +287,16 @@ Maestro::DensityAdvanceSDC (int which_step,
 
     Put1dArrayOnCart(rho0_old,rho0_old_cart,0,0,bcs_s,Rho);
 
+    /////////////////////////////////////////////////////////////////
+    // Subtract w0 from MAC velocities (MAC velocities has w0 already).
+    /////////////////////////////////////////////////////////////////
+
+    Addw0(umac,w0mac,-1.);
+
+    /////////////////////////////////////////////////////////////////
+    // Compute source terms
+    /////////////////////////////////////////////////////////////////
+
     // ** density source term **
 
     // Make source term for rho or rho'
@@ -378,13 +390,6 @@ Maestro::DensityAdvanceSDC (int which_step,
         // convert X --> (rho X) in scalold
         ConvertRhoXToX(scalold,false);
     }
-
-    /////////////////////////////////////////////////////////////////
-    // Subtract w0 from MAC velocities.
-    /////////////////////////////////////////////////////////////////
-
-    Addw0(umac,w0mac,-1.);
-
 
     /////////////////////////////////////////////////////////////////
     // Compute fluxes
