+++ Source/Src_nd/make_flux.F90
@@ -131,18 +131,15 @@ contains
              if (species_pred_type == predict_rhoprime_and_X) then
                 ! edge states are rho' and X.  To make the (rho X) flux,
                 ! we need the edge state of rho0
-                sfluxy(i,j,comp) = &
-                     (vmac(i,j)+w0(lev,j))*(rho0_edge+sedgey(i,j,rho_comp))*sedgey(i,j,comp)
+                sfluxy(i,j,comp) = vmac(i,j)*(rho0_edge+sedgey(i,j,rho_comp))*sedgey(i,j,comp)
 
              else if (species_pred_type == predict_rhoX) then
                 ! edge states are (rho X)
-                sfluxy(i,j,comp) = &
-                     (vmac(i,j)+w0(lev,j))*sedgey(i,j,comp)
+                sfluxy(i,j,comp) = vmac(i,j)*sedgey(i,j,comp)
 
              else if (species_pred_type == predict_rho_and_X) then
                 ! edge state are rho and X
-                sfluxy(i,j,comp) = &
-                     (vmac(i,j)+w0(lev,j))*sedgey(i,j,rho_comp)*sedgey(i,j,comp)
+                sfluxy(i,j,comp) = vmac(i,j)*sedgey(i,j,rho_comp)*sedgey(i,j,comp)
 
              endif
 
@@ -296,16 +293,16 @@ contains
                 if (species_pred_type == predict_rhoprime_and_X) then
                    ! edge states are rho' and X.  To make the (rho X)
                    ! flux, we need the edge state of rho0
-                   sfluxz(i,j,k,comp) = (wmac(i,j,k)+w0(lev,k))* &
+                   sfluxz(i,j,k,comp) = wmac(i,j,k)* &
                         (rho0_edge+sedgez(i,j,k,rho_comp))*sedgez(i,j,k,comp)
 
                 else if (species_pred_type == predict_rhoX) then
                    ! edge states are (rho X)
-                   sfluxz(i,j,k,comp) = (wmac(i,j,k)+w0(lev,k))*sedgez(i,j,k,comp)
+                   sfluxz(i,j,k,comp) = wmac(i,j,k)*sedgez(i,j,k,comp)
 
                 else if (species_pred_type == predict_rho_and_X) then
                    ! edge states are rho and X
-                   sfluxz(i,j,k,comp) = (wmac(i,j,k)+w0(lev,k))* &
+                   sfluxz(i,j,k,comp) = wmac(i,j,k)* &
                         sedgez(i,j,k,rho_comp)*sedgez(i,j,k,comp)
 
                 endif
@@ -369,9 +366,6 @@ contains
        umac,   u_lo, u_hi, &
        vmac,   v_lo, v_hi, &
        wmac,   w_lo, w_hi, &
-       w0macx, wx_lo, wx_hi, &
-       w0macy, wy_lo, wy_hi, &
-       w0macz, wz_lo, wz_hi, &
        rho0_edgex, rx_lo, rx_hi, &
        rho0_edgey, ry_lo, ry_hi, &
        rho0_edgez, rz_lo, rz_hi, &
@@ -397,12 +391,6 @@ contains
     double precision, intent(in   ) :: vmac  (v_lo(1):v_hi(1),v_lo(2):v_hi(2),v_lo(3):v_hi(3))
     integer         , intent(in   ) :: w_lo(3), w_hi(3)
[…]
