+++ Microphysics/networks/general_null/write_network.py
@@ -1,51 +1,47 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 import sys
-import getopt
+import argparse
 
-Header="""
-! DO NOT EDIT THIS FILE!!!  
-!  
-! This file is automatically generated by write_network.py at 
-! compile-time.  
-!  
-! To modify the species carried by the network, edit the appropriate inputs 
+HEADER = """
+! DO NOT EDIT THIS FILE!!!
+!
+! This file is automatically generated by write_network.py at
+! compile-time.
+!
+! To modify the species carried by the network, edit the appropriate inputs
 ! file.
 
 """
 
-
-#=============================================================================
-# the species class holds the properties of a single species
-#=============================================================================
-class Species(object):
-
+class Species:
+    """the species class holds the properties of a single species"""
     def __init__(self):
         self.name = ""
-        self.shortName = ""
+        self.short_name = ""
         self.A = -1
         self.Z = -1
 
-#=============================================================================
-# convenience class for an auxilliary variable
-#=============================================================================
-class AuxVar(object):
-    
+    def __str__(self):
+        return "species {}, (A,Z) = {},{}".format(self.name, self.A, self.Z)
+
+class AuxVar:
+    """convenience class for an auxilliary variable"""
     def __init__(self):
         self.name = ""
 
+    def __str__(self):
+        return "auxillary variable {}".format(self.name)
 
-#=============================================================================
-# get_next_line returns the next, non-blank line, with comments stripped
-#=============================================================================
+
 def get_next_line(fin):
-
+    """get_next_line returns the next, non-blank line, with comments
+    stripped"""
     line = fin.readline()
 
     pos = str.find(line, "#")
 
     while (pos == 0 or str.strip(line) == "") and line:
-
         line = fin.readline()
         pos = str.find(line, "#")
 
@@ -54,37 +50,36 @@ def get_next_line(fin):
     return line
 
 
-
-#=============================================================================
-# get_object_index looks through the list and returns the index corresponding to
-# the network object (species or auxvar) specified by name
-#=============================================================================
 def get_object_index(objs, name):
+    """look through the list and returns the index corresponding to the
+    network object (species or auxvar) specified by name
+
+    """
 
     index = -1
 
-    for n in range(len(objs)):
-        if objs[n].name == name:
+    for n, o in enumerate(objs):
+        if o.name == name:
             index = n
             break
 
     return index
 
 
-#=============================================================================
-# parse_net_file read all the species listed in a given network inputs file
-# and adds the valid species to the species list
-#=============================================================================
 def parse_net_file(species, aux_vars, net_file):
+    """parse_net_file read all the species listed in a given network
+    inputs file and adds the valid species to the species list
+
+    """
 
     err = 0
 
-    try: f = open(net_file, "r")
+    try:
+        f = open(net_file, "r")
     except IOError:
+        sys.exit("write_network.py: ERROR: file "+str(net_file)+" does not exist")
 
-        print("write_network.py: ERROR: file "+str(net_file)+" does not exist")
-        sys.exit(2)
-        
+
     print("write_network.py: working on network file "+str(net_file)+"...")
 
     line = get_next_line(f)
@@ -95,19 +90,20 @@ def parse_net_file(species, aux_vars, net_file):
 
         # read the species or auxiliary variable from the line
         net_obj, err = parse_network_object(fields)
-        if net_obj is None: return err
+        if net_obj is None:
+            return err
 
         objs = species
[…]
