### 2019-09-30 b04369c
add a rudimentary test problem for planar diagnostics
