### 2019-09-30 e2574c6
compute gradp0 outside of fortran subroutine so that we don't need to access ghost cells for p0
