+++ Source/MaestroMacProj.cpp
@@ -18,12 +18,12 @@ Maestro::MacProj (Vector<std::array< MultiFab, AMREX_SPACEDIM > >& umac,
 {
     // timer for profiling
     BL_PROFILE_VAR("Maestro::MacProj()", MacProj);
-// 
-// #ifdef AMREX_USE_CUDA
-//     auto not_launched = Gpu::notInLaunchRegion();
-//     // turn on GPU
-//     if (not_launched) Gpu::setLaunchRegion(true);
-// #endif
+
+#ifdef AMREX_USE_CUDA
+    auto not_launched = Gpu::notInLaunchRegion();
+    // turn on GPU
+    if (not_launched) Gpu::setLaunchRegion(true);
+#endif
 
     // this will hold solver RHS = macrhs - div(beta0*umac)
     Vector<MultiFab> solverrhs(finest_level+1);
@@ -225,10 +225,10 @@ Maestro::MacProj (Vector<std::array< MultiFab, AMREX_SPACEDIM > >& umac,
         FillPatchUedge(umac);
     }
 
-// #ifdef AMREX_USE_CUDA
-//     // turn off GPU
-//     if (not_launched) Gpu::setLaunchRegion(false);
-// #endif
+#ifdef AMREX_USE_CUDA
+    // turn off GPU
+    if (not_launched) Gpu::setLaunchRegion(false);
+#endif
 
 }
 
