### 2019-10-01 bfb2339
added invert threshold density flag so can turn off burning *above* a certain density threshold
