### 2019-10-01 43930a1
remove hard-coded dimensional factor
