+++ Source/Src_nd/make_grav.F90
@@ -1,0 +1,368 @@
+module make_grav_module
+
+  use amrex_constants_module
+  use base_state_geometry_module, only: nr_fine, &
+                                        max_radial_level, nr, numdisjointchunks, &
+                                        r_start_coord, r_end_coord, finest_radial_level
+  use meth_params_module, only: spherical, grav_const, base_cutoff_density, &
+                                do_planar_invsq_grav, planar_invsq_mass, do_2d_planar_octant
+  use fundamental_constants_module, only: Gconst
+
+  implicit none
+
+contains
+
+  subroutine make_grav_cell(grav_cell,rho0,r_cc_loc,r_edge_loc) &
+       bind(C, name="make_grav_cell")
+    ! compute the base state gravitational acceleration at the cell
+    ! centers.  The base state uses 0-based indexing, so grav_cell
+    ! does too.
+
+     use base_state_geometry_module, only: restrict_base, fill_ghost_base
+
+    double precision, intent(  out) ::  grav_cell(0:max_radial_level,0:nr_fine-1)
+    double precision, intent(in   ) ::       rho0(0:max_radial_level,0:nr_fine-1)
+    double precision, intent(in   ) ::   r_cc_loc(0:max_radial_level,0:nr_fine-1)
+    double precision, intent(in   ) :: r_edge_loc(0:max_radial_level,0:nr_fine  )
+
+    ! Local variables
+    integer :: r, n, i
+    double precision, allocatable :: m(:,:)
+    double precision              :: term1, term2
+
+    call bl_proffortfuncstart("Maestro::make_grav_cell")
+
+    if (spherical .eq. 0) then
+
+       if (do_planar_invsq_grav)  then
+
+          ! we are doing a plane-parallel geometry with a 1/r**2
+          ! gravitational acceleration.  The mass is assumed to be
+          ! at the origin.  The mass in the computational domain
+          ! does not contribute to the gravitational acceleration.
+          do n=0,finest_radial_level
+             do r = 0, nr(n)-1
+                grav_cell(n,r) = -Gconst*planar_invsq_mass / r_cc_loc(n,r)**2
+             enddo
+          enddo
+
+       else if (do_2d_planar_octant .eq. 1) then
+
+          ! compute gravity as in the spherical case
+
+          allocate(m(0:finest_radial_level,0:nr_fine-1))
+
+          n = 0
+          m(n,0) = FOUR3RD*M_PI*rho0(n,0)*r_cc_loc(n,0)**3
+          grav_cell(n,0) = -Gconst * m(n,0) / r_cc_loc(n,0)**2
+
+          do r=1,nr(n)-1
+
+             ! the mass is defined at the cell-centers, so to compute
+             ! the mass at the current center, we need to add the
+             ! contribution of the upper half of the zone below us and
+             ! the lower half of the current zone.
+
+             ! don't add any contributions from outside the star --
+             ! i.e.  rho < base_cutoff_density
+             if (rho0(n,r-1) > base_cutoff_density) then
+                term1 = FOUR3RD*M_PI*rho0(n,r-1) * &
+                     (r_edge_loc(n,r) - r_cc_loc(n,r-1)) * &
+                     (r_edge_loc(n,r)**2 + &
+                     r_edge_loc(n,r)*r_cc_loc(n,r-1) + &
+                     r_cc_loc(n,r-1)**2)
+             else
+                term1 = ZERO
+             endif
+
+             if (rho0(n,r) > base_cutoff_density) then
+                term2 = FOUR3RD*M_PI*rho0(n,r  )*&
[…]
