### 2019-10-01 2d25d2a
delete print statement / unnecessary print statement
