### 2019-10-01 a385ebb
Add diagnostics for planar geometries (#90)
Diagnostics currently are spherical only, this PR extends the same measurements to a planar geometry. In 2D, things like z-components are just recorded as zero, and spherical-specific things such as measurements at the star centre are not recorded.

The fortran now has separate diag and diag_sphr subroutines (where diag_sphr is what was previously diag). They basically do the same thing in this PR. Gravitational potential energy is defined differently so it too has separate planar and spherical subroutines.

* add diagnostics for planar mode

* tidy 
[…]
