+++ Source/MaestroDiag.cpp
@@ -4,6 +4,10 @@
 
 using namespace amrex;
 
+// Choose precision of all output files
+const int outfilePrecision = 10;
+const int setwVal = outfilePrecision+2+4+4; // 0. + precision + 4 for exp + 4 for gap
+
 // write diagnostics files to disk
 // We hold many timesteps-worth of diagnostic information in a buffer
 // and output to the files only when flush_diag() is called.  This
@@ -17,23 +21,16 @@ Maestro::DiagFile (const int step,
                    const Vector<MultiFab>& s_in,
                    int& index)
 {
-    if (spherical == 0) {
-        if (ParallelDescriptor::IOProcessor()) {
-            Warning("WARNING: WriteDiagFile() not written for non-spherical geometry");
-        }
-        return;
-    }
 
     // timer for profiling
     BL_PROFILE_VAR("Maestro::DiagFile()",DiagFile);
 
-    // w0mac will contain an edge-centered w0 on a Cartesian grid,
-    // for use in computing divergences.
+    // -- w0mac will contain an edge-centered w0 on a Cartesian grid,
+    // -- for use in computing divergences.
     Vector<std::array< MultiFab, AMREX_SPACEDIM > > w0mac(finest_level+1);
-
-    // w0r_cart is w0 but onto a Cartesian grid in cell-centered as
-    // a scalar.  Since w0 is the radial expansion velocity, w0r_cart
-    // is the radial w0 in a zone
+    // -- w0r_cart is w0 but onto a Cartesian grid in cell-centered as
+    // -- a scalar.  Since w0 is the radial expansion velocity, w0r_cart
+    // -- is the radial w0 in a zone
     Vector<MultiFab> w0r_cart(finest_level+1);
 
     // rho_Hnuc and rho_Hext are used to determine energy generation
@@ -64,22 +61,20 @@ Maestro::DiagFile (const int step,
         // expansion velocity)
         Put1dArrayOnCart(w0,w0r_cart,1,0,bcs_u,0,1);
 
-        // compute rho_Hext and rho_Hnuc
-        for (int lev=0; lev<=finest_level; ++lev) {
-            stemp             [lev].define(grids[lev], dmap[lev],   Nscal, 0);
-            rho_Hext          [lev].define(grids[lev], dmap[lev],       1, 0);
-            rho_omegadot      [lev].define(grids[lev], dmap[lev], NumSpec, 0);
-            rho_Hnuc          [lev].define(grids[lev], dmap[lev],       1, 0);
-        }
+    } 
 
-        if (dt < small_dt) {
-            React(s_in, stemp, rho_Hext, rho_omegadot, rho_Hnuc, p0_in, small_dt, t_in);
-        } else {
-            React(s_in, stemp, rho_Hext, rho_omegadot, rho_Hnuc, p0_in, dt*0.5, t_in);
-        }
+    // compute rho_Hext and rho_Hnuc
+    for (int lev=0; lev<=finest_level; ++lev) {
+        stemp             [lev].define(grids[lev], dmap[lev],   Nscal, 0);
+        rho_Hext          [lev].define(grids[lev], dmap[lev],       1, 0);
+        rho_omegadot      [lev].define(grids[lev], dmap[lev], NumSpec, 0);
+        rho_Hnuc          [lev].define(grids[lev], dmap[lev],       1, 0);
+    }
 
+    if (dt < small_dt) {
+        React(s_in, stemp, rho_Hext, rho_omegadot, rho_Hnuc, p0_in, small_dt, t_in);
     } else {
-        Abort("ERROR: WriteDiagFile() not supported for non-spherical geometry");
+        React(s_in, stemp, rho_Hext, rho_omegadot, rho_Hnuc, p0_in, dt*0.5, t_in);
     }
 
[…]
