### 2019-10-02 edeeb35
move instantaneous_reaction_rates subroutine to burner_loop_module and limit computations to region within burning_cutoff_density
