+++ Source/Maestro_F.H
@@ -155,6 +155,10 @@ extern "C"
 			      const int use_mask);
 #endif
     
+    void instantaneous_reaction_rates(const int* lo, const int* hi,
+				      amrex::Real* rho_omegadot, const int* o_lo, const int* o_hi,
+				      amrex::Real* rho_Hext, const int* h_lo, const int* h_hi,
+				      const amrex::Real* scal, const int* s_lo, const int* s_hi);
     //////////////////////
 
     //////////////////////
@@ -777,14 +781,6 @@ extern "C"
     // 			const amrex::Real* p0_old,
     // 			const amrex::Real* p0_new,
     // 			const amrex::Real* dt);
-    //////////////////////
-
-    //////////////////////
-    // in make_reaction_rates.F90
-    void instantaneous_reaction_rates(const int* lo, const int* hi,
-				      amrex::Real* rho_omegadot, const int* o_lo, const int* o_hi,
-				      amrex::Real* rho_Hext, const int* h_lo, const int* h_hi,
-				      const amrex::Real* scal, const int* s_lo, const int* s_hi);
     //////////////////////
 
     //////////////////////
