### 2019-10-07 AMReX-Astro/MAESTROeX#103 @tukss
list of dependencies in the documentation
I am currently reviewing the submission to JOSS in openjournals/joss-reviews#1757. One required points in the checklist is "a clearly-stated list of dependencies" in the installation instructions. I couldn't find that in the documentation.

I did some tests in a minimal container to see what is required. I came up with this list:
- GNU make
- C++ compiler
- Fortran compiler
- Python

In Debian/Ubuntu the packages would be: `make g++ gfortran python`. I guess MPI will most likely be needed in most real-world scenarios.

That being said, the dependencies for MAESTROeX are so minimal that most people won't have to worry about that. Still, I would consider explicitly mentioning them a selling point actually.

I found a comment on the minimum required Fortran standard (F95) in the FAQ. Do you know what the requirement for the C++ side of the code is? I think that would be useful to know since some C++17 features might not be present in older compiler versions.
