### 2019-10-08 0ac47ae
added a section describing software requirements in docs
