### 2019-10-08 AMReX-Astro/MAESTROeX#103 @harpolea
There's some information in the AMReX docs about this, but it could definitely do with being repeated in the MAESTROeX docs. I've added a section on software requirements to the _Getting Started_ section of the docs in 0ac47ae2.
