+++ sphinx_docs/source/getting_started.rst
@@ -11,7 +11,8 @@ Requirements
 
 MAESTROeX requires a C++ compiler that supports the C++11 standard, a 
 Fortran compiler that supports the Fortran 2003 standard, and a C compiler 
-that supports the C99 standard. Several compiler suites are supported, included GNU, Intel, PGI and Cray. GNU Make is also required, as is Python 
+that supports the C99 standard. Several compiler suites are supported, 
+including GNU, Intel, PGI and Cray. GNU Make is also required, as is Python 
 (>= 2.7, including 3) and standard tools available in any Unix-like 
 environments (e.g., Perl and ``sed``). 
 
