+++ Source/MaestroPlot.cpp
@@ -361,9 +361,9 @@ Maestro::PlotFileMF (const int nPlot,
     }
 #else	
     if (dt_in < small_dt) {
-	ReactSDC(s_in, stemp, rho_Hext, p0_in, small_dt, t_old, sdc_source);
+	ReactSDC(s_in, stemp, rho_Hext, p0_in, small_dt, t_in, sdc_source);
     } else {
-	ReactSDC(s_in, stemp, rho_Hext, p0_in, dt_in*0.5, t_old, sdc_source);
+	ReactSDC(s_in, stemp, rho_Hext, p0_in, dt_in*0.5, t_in, sdc_source);
     }
     
     MakeReactionRates(rho_omegadot,rho_Hnuc,s_in);
