+++ Microphysics/conductivity/conductivity.F90
@@ -44,13 +44,9 @@ contains
     ! call the EOS, passing through the arguments we called conducteos with
     call eos(input, state)
 
-#ifndef SDC
-    call actual_conductivity(state)
-#else
     call actual_conductivity(state, conduct)
 
     state % conductivity = conduct
-#endif
 
   end subroutine conducteos
 
@@ -65,13 +61,9 @@ contains
 
     real(rt) :: conduct
 
-#ifndef SDC
-    call actual_conductivity(state)
-#else
     call actual_conductivity(state, conduct)
 
     state % conductivity = conduct
-#endif
 
   end subroutine conductivity
 
