### 2019-10-11 AMReX-Astro/MAESTROeX#106 @tukss
community guidelines: where to seek support
This is a issue related to the JOSS submission: openjournals/joss-reviews#1757

There are clear instructions on how to contribute though pull requests both in the README file and the documentation. This is good.

The README file mentions a mailing list as the place to get help. When I follow the [link](https://groups.google.com/forum/#!forum/maestro-help), I just get a message that the group doesn't exist or I don't have permission to access it. If the group is indeed private, there should be information on how to get invited.

There is no explicit information on how to report issues. You might want to indicate your preference on using GitHub's issue tracker or sending emails in the README.
