### 2019-10-11 AMReX-Astro/MAESTROeX#107 @tukss
[JOSS] improvements for Paper
This is part of the review of the JOSS paper, which is part of this repository.
openjournals/joss-reviews#1757

The paper is mostly fine, but I have a few points for improvement to meet the checklist for the review.
- Fan et al. (2019) has been submitted to ApJ, but is already on the arXiv. Can you add a link to the [arXiv submission](https://arxiv.org/abs/1908.03634) to the bibliography?
- To accurately describe what the target audience is, can you add a statement on the range of applicability? For example: The algorithm becomes inaccurate above Mach numbers of …
- JOSS asks for a description of the state of the field. Can you briefly describe how the MAESTROeX approach compares to other commonly used hydro schemes? For example, fully compressible or anelastic.
