### 2019-10-15 0eb8340
changed getting help section of the README to ask users to submit github issues, addressing #106
