+++ README.md
@@ -148,6 +148,6 @@ status as a core developer.
 
 ## Getting help
 
-Join the mailing list to ask for help or stay up-to-date:
+For answers to frequently asked questions and solutions to common problems, consult the [FAQ section](https://amrex-astro.github.io/MAESTROeX/docs/faq.html) of the User's Guide. 
 
-https://groups.google.com/forum/#!forum/maestro-help
+If you discover any bugs in the code or would like to request further help, then please submit a issue to the MAESTROeX Github repository describing your problem.
