### 2019-10-15 AMReX-Astro/MAESTROeX#106 @harpolea
I've changed the `Getting Help` section of the README in 0eb8340 so it now directs users to the FAQ section of the User's Guide and asks them to submit a github issue if they require further assistance or wish to report a bug
