### 2019-10-15 AMReX-Astro/MAESTROeX#106 @harpolea
It turns out the visibility of the mailing list group was incorrect, so now that's been fixed I've added it back to the README
