### 2019-10-17 b76d17a
add subroutines to enable enthalpy_pred_type = predict_T_then_rhohprime, predict_T_then_h, predict_Tprime_then_h during enthalpy advance
