+++ Source/Src_nd/rhoh_vs_t.F90
@@ -1,5 +1,6 @@
 module rhoh_vs_t_module
 
+  use amrex_constants_module
   use eos_type_module
   use eos_module
   use network, only: nspec
@@ -8,6 +9,20 @@ module rhoh_vs_t_module
   use base_state_geometry_module, only:  max_radial_level, nr_fine
 
   implicit none
+
+  private
+  
+  integer, parameter :: predict_rhoh             = 0;
+  integer, parameter :: predict_rhohprime        = 1;
+  integer, parameter :: predict_h                = 2;
+  integer, parameter :: predict_T_then_rhohprime = 3;
+  integer, parameter :: predict_T_then_h         = 4;
+  integer, parameter :: predict_hprime           = 5;
+  integer, parameter :: predict_Tprime_then_h    = 6;
+
+  integer, parameter :: predict_rhoprime_and_X   = 1;
+  integer, parameter :: predict_rhoX             = 2;
+  integer, parameter :: predict_rho_and_X        = 3;
 
 contains
 
@@ -347,6 +362,464 @@ contains
     endif
 
   end subroutine makeCsfromRhoH
+
+
+  subroutine makeHfromRhoT_edge(lo,hi,sedgex,x_lo,x_hi, &
+       sedgey,y_lo,y_hi, &
+#if (AMREX_SPACEDIM == 3)
+       sedgez,z_lo,z_hi, &
+#endif
+       rho0_cart,r_lo,r_hi, &
+       rhoh0_cart,rh_lo,rh_hi, &
+       temp0_cart,t_lo,t_hi, &
+       rho0_edge_cart,re_lo,re_hi, &
+       rhoh0_edge_cart,rhe_lo,rhe_hi, &
+       temp0_edge_cart,te_lo,te_hi) &
+       bind(C,name="makeHfromRhoT_edge")
+    
+    use meth_params_module, only: species_pred_type, enthalpy_pred_type, small_temp
+    
+    integer         , intent (in   ) :: lo(3), hi(3)
+    integer         , intent (in   ) :: x_lo(3), x_hi(3)
+    double precision, intent (inout) :: sedgex(x_lo(1):x_hi(1),x_lo(2):x_hi(2),x_lo(3):x_hi(3),nscal)
+    integer         , intent (in   ) :: y_lo(3), y_hi(3)
+    double precision, intent (inout) :: sedgey(y_lo(1):y_hi(1),y_lo(2):y_hi(2),y_lo(3):y_hi(3),nscal)
+#if (AMREX_SPACEDIM == 3)
+    integer         , intent (in   ) :: z_lo(3), z_hi(3)
+    double precision, intent (inout) :: sedgez(z_lo(1):z_hi(1),z_lo(2):z_hi(2),z_lo(3):z_hi(3),nscal)
+#endif
+    integer         , intent (in   ) :: r_lo(3), r_hi(3)
+    double precision, intent (in   ) :: rho0_cart (r_lo(1):r_hi(1),r_lo(2):r_hi(2),r_lo(3):r_hi(3))
+    integer         , intent (in   ) :: rh_lo(3), rh_hi(3)
+    double precision, intent (in   ) :: rhoh0_cart (rh_lo(1):rh_hi(1),rh_lo(2):rh_hi(2),rh_lo(3):rh_hi(3))
+    integer         , intent (in   ) :: t_lo(3), t_hi(3)
+    double precision, intent (in   ) :: temp0_cart (t_lo(1):t_hi(1),t_lo(2):t_hi(2),t_lo(3):t_hi(3))
+    integer         , intent (in   ) :: re_lo(3), re_hi(3)
[…]
