+++ Source/Src_nd/make_scal_force.F90
@@ -2,7 +2,9 @@ module make_scal_force_module
 
   use amrex_mempool_module, only : bl_allocate, bl_deallocate
   use base_state_geometry_module, only:  max_radial_level, nr_fine, dr, nr, base_cutoff_density_coord
-  use meth_params_module, only: enthalpy_pred_type, use_exact_base_state, base_cutoff_density, spherical
+  use meth_params_module, only: enthalpy_pred_type, use_exact_base_state, base_cutoff_density, spherical, &
+       nscal, rho_comp, temp_comp, spec_comp
+  use amrex_constants_module
 
   implicit none
 
@@ -165,6 +167,177 @@ contains
 
   end subroutine mkrhohforce
 
+  subroutine mktempforce(lo, hi, lev, &
+       temp_force, f_lo, f_hi, &
+       scal, s_lo, s_hi, &
+       umac, u_lo, u_hi, &
+       vmac, v_lo, v_hi, &
+#if (AMREX_SPACEDIM == 3)
+       wmac, w_lo, w_hi, &
+#endif
+       thermal, t_lo, t_hi, &
+       p0_cart, p_lo, p_hi, &
+       psi_cart, ps_lo, ps_hi, &
+       dx, domhi)  bind(C,name="mktempforce")
+
+    use eos_module
+    use eos_type_module
+    use network, only: nspec
+    
+    integer         , intent(in   ) :: lo(3),hi(3)
+    integer, value  , intent(in   ) :: lev
+    integer         , intent(in   ) :: f_lo(3), f_hi(3)
+    double precision, intent(inout) :: temp_force(f_lo(1):f_hi(1),f_lo(2):f_hi(2),f_lo(3):f_hi(3))
+    integer         , intent(in   ) :: s_lo(3), s_hi(3)
+    double precision, intent(in   ) :: scal(s_lo(1):s_hi(1),s_lo(2):s_hi(2),s_lo(3):s_hi(3),nscal)
+    integer         , intent(in   ) :: u_lo(3), u_hi(3)
+    double precision, intent(in   ) :: umac(u_lo(1):u_hi(1),u_lo(2):u_hi(2),u_lo(3):u_hi(3))
+    integer         , intent(in   ) :: v_lo(3), v_hi(3)
+    double precision, intent(in   ) :: vmac(v_lo(1):v_hi(1),v_lo(2):v_hi(2),v_lo(3):v_hi(3))
+#if (AMREX_SPACEDIM == 3)
+    integer         , intent(in   ) :: w_lo(3), w_hi(3)
+    double precision, intent(in   ) :: wmac(w_lo(1):w_hi(1),w_lo(2):w_hi(2),w_lo(3):w_hi(3))
+#endif
+    integer         , intent(in   ) :: t_lo(3), t_hi(3)
+    double precision, intent(in   ) :: thermal(t_lo(1):t_hi(1),t_lo(2):t_hi(2),t_lo(3):t_hi(3))
+    integer         , intent(in   ) :: p_lo(3), p_hi(3)
+    double precision, intent(in   ) :: p0_cart(p_lo(1):p_hi(1),p_lo(2):p_hi(2),p_lo(3):p_hi(3))
+    integer         , intent(in   ) :: ps_lo(3), ps_hi(3)
+    double precision, intent(in   ) :: psi_cart(ps_lo(1):ps_hi(1),ps_lo(2):ps_hi(2),ps_lo(3):ps_hi(3))
+    double precision, intent(in   ) :: dx(3)
+    integer         , intent(in   ) :: domhi(3)
+
+    ! Local variable
+    double precision :: gradp0,veladv,dhdp
+    integer          :: i,j,k
+    
[…]

[…]
