### 2019-10-17 #1757 @tukss
The proof PDF contains 10.21105/joss.01757, but it doesn't work yet. @arfon, can you comment if the DOI will change after acceptance?
