### 2019-10-17 #1757 @arfon
> 10.21105/joss.01757

@crawfordsm ☝️ this will be the DOI.
