### 2019-10-17 #1757 @XinlongSBU
Hi,

This is the commit that I want to modify Exec/SCIENCE/wdconvect
<https://github.com/XinlongSBU/MAESTROeX/commit/bbbc157984b9c26254cb54b1e9f100d4fabc3bdf#diff-1ecb3a9669f1bc882c83705728af8ec6>.
Other changes are mistakes. I don't think this is merged into
AMReX-Astro/MAESTROeX <https://github.com/AMReX-Astro/MAESTROeX>.

Best,
Xinlong

Philipp Grete <notifications@github.com> 于2019年10月17日周四 上午11:46写道：

> I'm trying to understand the contribution of @XinlongSBU
> <https://github.com/XinlongSBU> specifically this commit bbbc157984b9c2
> Is the changeset that large because of the different parent or is this a
> significant contribution to the code (I'm asking as @XinlongSBU
> <https://github.com/XinlongSBU> is not listed as author on the paper as
> far as I can tell).
>
