### 2019-10-21 AMReX-Astro/MAESTROeX#107 @pgrete
I agree with with these recommendations.
Moreover, with respect to the third item it'd also be good to refer to a few comparable codes that also implement low Mach number schemes (from the guide: _A list of key references, including to other software addressing related needs._)
