### 2019-10-21 AMReX-Astro/MAESTROeX#109 @pgrete
[JOSS] Issue/problem reporting
I currently cannot find instructions on the process of reporting issues and/or problems (beyond asking for help on the mailing list).
It'd be good to have that info already in the README and optimal (but not necessary) to have a template.

linking issue: https://github.com/openjournals/joss-reviews/issues/1757
