### 2019-10-21 AMReX-Astro/MAESTROeX#110 @pgrete
[JOSS] Development model documention
The documentation https://amrex-astro.github.io/MAESTROeX/docs/getting_started.html#development-model does not mention the requirement of using a fork (as does the README).
Moreover, it'd be great (but not necessary) to provide a template for pull requests -- especially as it would make the resulting squashed message more accessible to new users and more streamlined in general.

linking issue: openjournals/joss-reviews#1757
