### 2019-10-21 AMReX-Astro/MAESTROeX#111 @pgrete
[JOSS] Testing instructions
I was not able to find detailed instructions on how to run the test suite locally.
I infer that I'd have to install the AMReX suite and modify an ini file.
It'd be great if there's an example on how to run the tests, which may facilitate contributions by new developers who don't have access the main repository (and the associated central test infrastructure).
Moreover, this would ensure that (new) userd correctly setup their environment.

linking issue: openjournals/joss-reviews#1757
