### 2019-10-21 AMReX-Astro/MAESTROeX#112 @pgrete
[JOSS] Example installation instructions
When I first tried to follow the example in the README or homepage (https://amrex-astro.github.io/MAESTROeX/) my compilation failed as I used an Intel environment.
A small clarification, e.g., "(assuming a gnu compiler suite environment)" would be helpful (as already described in the detailed docs).

linking issue: openjournals/joss-reviews#1757
