### 2019-10-21 AMReX-Astro/MAESTROeX#113 @pgrete
[JOSS] GPU instructions
The JOSS paper mentions increasing GPU support, which I interpret as "some parts already run on GPUs, others don't".
However, I didn't find more detail on what is already supported and how I compile MAESTROeX
 with GPU support.

linking issue: openjournals/joss-reviews#1757
