### 2019-10-21 AMReX-Astro/MAESTROeX#114 @pgrete
[JOSS] Author list
Why is the author list in the software paper dissimilar from the authorlist on the homepage https://amrex-astro.github.io/MAESTROeX/who.html ?

linking issue: openjournals/joss-reviews#1757
