### 2019-10-21 AMReX-Astro/MAESTROeX#115 @pgrete
[JOSS] Plotting instructions are out of date
I tried to follow https://amrex-astro.github.io/MAESTROeX/docs/visualization.html#python-visualization-scripts and the AMReX structure probably changed.
I found the fsnapshot tool in `amrex/Tools/Plotfile`.

Also the `yt` example that are supposed to be in the `Util` folder are missing.

linking issue: openjournals/joss-reviews#1757
