### 2019-10-21 AMReX-Astro/MAESTROeX#112 @pgrete
In addition, the point 8 (Examine the output. https://amrex-astro.github.io/MAESTROeX/docs/getting_started.html) mentioned plotting and restart file.
The test run I conducted resulted in neither `plt` nor `chk` directories. I only have `serial` directories.
Did I do something wrong or have the default been changed (or something else)?
