### 2019-10-21 #1757 @pgrete
I'm done with my first pass of the review and opened issues for all things I spotted.
