### 2019-10-22 AMReX-Astro/MAESTROeX#109 @harpolea
If you check out the development branch, I added instructions in the README that direct users to the User's Guide FAQ and invite them to submit GitHub issues if they discover any bugs.
