### 2019-10-22 AMReX-Astro/MAESTROeX#114 @harpolea
The author list on the website also includes developers of the previous MAESTRO code, whereas the author list on the paper only lists developers of the new MAESTROeX code.
