### 2019-10-22 646058f
updated development instructions
instructs users to submit a pull request from their fork
