### 2019-10-22 AMReX-Astro/MAESTROeX#110 @harpolea
I've updated the User's Guide (115f582f27) and the home page (646058f) to instruct users to submit a PR from their fork.
