### 2019-10-22 b05960c
add requirements to README and update reacting_bubble inputs file
