### 2019-10-22 AMReX-Astro/MAESTROeX#112 @harpolea
I've added the requirements to the README (b05960c5).

Point 8 says that the plotfiles will be named `pltnnnnn` and the checkpoint files `chknnnn` **by default**. If you look at the inputs file, for this problem `maestro.plot_base_name` had been set to `serial` (so overriding the default). I've reverted this to `plt`. There are were no checkpoint files output as `maestro.chk_int` was set to -1. I've changed this to 1000 so that you should also get one of these output for the example problem. (b05960c5)

There is a plot of the outputs of the `reacting_bubble` problem in the section following the Getting Started section: https://amrex-astro.github.io/MAESTROeX/docs/getting_started.html#working-with-the-output.
