### 2019-10-22 AMReX-Astro/MAESTROeX#111 @harpolea
The test suite runs regression tests that compare the outputs of several test problems against previously computed benchmarks. As the exact outputs of any simulation will vary to some degree between different computational setups (e.g. due to different compilers / versions / hardware), it would not make sense for us to provide a set of benchmarks for others to download and compare their simulations against, as they will almost certainly not agree.

Before we merge any PR's, we test them using our local test suite. If there are any conflicts, we work with the submitter to resolve these conflicts. At this point, it's normally only a small number of test cases that fail. The webpages listing the results of the test suite describe the full set up of each test case, so it's possible for a user to run the individual tests on their system by using the setups listed.

As described in the README, we provide a set of unit tests in the `Exec/UNIT_TESTS` directory for simple problems that have analytic / more repeatable solutions. These can be used by a new user to test that they have MAESTROeX properly set up.
