### 2019-10-22 3e63ed6
added script for producing sliceplots
