### 2019-10-22 AMReX-Astro/MAESTROeX#114 @ajnonaka
Wondering if we should put a note in the acknowledgements section of the JOSS paper those who were involved with original MAESTRO algorithm+software development.
