### 2019-10-22 b32bffe
added more functionality to plotsinglevar.py
