### 2019-10-22 c5b6960
contour compare script
