### 2019-10-22 52e0165
clarify COMP=gnu in the homepage example
