### 2019-10-22 AMReX-Astro/MAESTROeX#112 @ajnonaka
clarified usage of gnu (added "COMP=gnu") to README on homepage) in 52e0165e0d93721ef9d7a11ea024ff17120a5662
