### 2019-10-22 AMReX-Astro/MAESTROeX#109 @ajnonaka
I also added a note about submitting an issue on github to the gh-pages front page (it already directed to the User's Guide and the mailing list)
