### 2019-10-22 fbf9db7
updated docs to reflect new scripts
