### 2019-10-22 AMReX-Astro/MAESTROeX#115 @harpolea
I've updated the plotting instructions in the docs and put the required scripts in the `Util/yt` directory
