### 2019-10-23 d9a5b83
JOSS paper updates
