### 2019-10-23 AMReX-Astro/MAESTROeX#107 @ajnonaka
Addressed in fbf9db7b0b5ccbf085d7b9d7a47783b16c67f484.  For comparison to other low Mach schemes (point #3), I instead added a reference to the MAESTROeX paper where it is discussed more thoroughly.
