### 2019-10-23 649338d
changed subroutine to be used as gpu example in user's guide
