+++ docs/_sources/gpu.rst.txt
@@ -1,0 +1,285 @@
+.. _sec:gpu:
+
+***
+GPU
+***
+
+In this chapter, we will present the GPU support in MAESTROeX,
+including necessary build parameters, how to offload a routine
+to GPU, and some basic profiling and debugging options.
+Note that currently MAESTROeX only supports Nvidia GPUs. 
+
+Requirements
+============
+
+Since MAESTROeX uses primarily CUDA C++ and CUDA Fortran,
+it requires a recent version of CUDA (e.g., >= 9) and the device
+must have compute capability of >= 6. 
+
+.. _sec:gpubuild:
+
+Building GPU Support
+====================
+
+To build MAESTROeX_with GPU support, add the following arguments
+to the GNUmakefile:
+
+::
+
+      USE_CUDA := TRUE
+      USE_GPU_PRAGMA := TRUE
+
+We also need to set ``USE_OMP = FALSE`` because OpenMP is currently
+not compatible with building with CUDA. 
+``USE_CUDA = TRUE`` and ``USE_OMP = TRUE`` will fail to compile.
+However, you may use MPI with CUDA for additional parallelization.
+
+Only IBM and PGI support CUDA Fortran, so the compiler should be set as:
+
+::
+
+      COMP := pgi 
+
+Finally, the integrator package in microphysics must also be defined
+for full GPU support.
+
+::
+
+    INTEGRATOR_DIR   := VODE90
+
+.. _sec:gpuporting:
+    
+Offloading a routine to GPU
+===============
+
+In order to offload a routine to the GPU, we insert a ``#pragma gpu``
+statement on the line before the call. This tells the preprocessor to
+generate a CUDA device version of the function, with all the
+additional CUDA GPU management. In addition to this, there are a few
+other modifications required to ensure the function operates correctly
+on the GPU. 
+
+C++
+---
+
+- Make sure that the box ``lo`` and ``hi`` are the first two arguments
+  and use ``AMREX_INT_ANYD`` as the macro wrapping ``bx.loVect()`` and
+  ``bx.hiVect()``
+
+- Likewise, inplace of ``ZFILL()``, use ``AMREX_REAL_ANYD``
+
+- Scalars must be passed by value to Fortran functions (not by
+  reference)
+
+- Make sure that you change the variable definitions in the Fortran
+  code to include value so that the Fortran knows the variables are
+  being passed by value rather than by reference. If you don’t do
+  this, the code is liable to segfault
+
+- FArrayBox functions like `setVal()` and `saxpy()` are not on the
+  GPU, so you should explicitly write out these operations in C++.
+  The MultiFab counterparts are on the GPU.
+
+- Use ``BL_TO_FORTRAN_ANYD()`` to wrap MultiFab arguments (and in the header file wrap with ``BL_FORT_FAB_ARG_3D``)
+
+To illustrate these modifications, consider the function ``mk_sponge``. To call the function on the CPU, we would write
+
+.. code-block:: c++
+
+    for (MFIter mfi(sponge_mf, true); mfi.isValid(); ++mfi) {
+
+        const Box& tileBox = mfi.tilebox();
+
+        mk_sponge(ARLIM_3D(tileBox.loVect()), ARLIM_3D(tileBox.hiVect()),
+                  BL_TO_FORTRAN_ANYD(sponge_mf[mfi]),
+		  ZFILL(dx), dt);
+     }
+
+Implementing the changes described above to offload this to GPU, this becomes
+
+.. code-block:: c++
+
+    for (MFIter mfi(sponge_mf, true); mfi.isValid(); ++mfi) {
+
+        const Box& tileBox = mfi.tilebox();
+
[…]
