### 2019-10-23 AMReX-Astro/MAESTROeX#113 @doreenfan
A section on GPU support has been added to User's Guide in the development branch. Please let me know if any information is missing.
