### 2019-10-23 e61359d
some minor edits to gpu docs
