+++ Exec/SCIENCE/code_comp/analysis/plot_generator.py
@@ -24,6 +24,7 @@ def parse_args():
     bounds_help = "The bounds for the colorbar."
     cmap_help = "The colormap for the variable to plot."
     log_help = "If provided, sets the plot to a logarithmic scale."
+    linthresh_help = "If provided, sets the linear threshold for a symlog plot"
     time_help = "If provided, adds a timestamp to each plot with the given precision."
     ext_help = "The extension of the file format to save to. PNG by default."
     sort_help = """A floating point number specifying the digits to sort file names by. Digits preceding the decimal point
@@ -31,6 +32,7 @@ def parse_args():
             descending order."""
     xlim_help = "The x-axis limits."
     ylim_help = "The y-axis limits."
+    zlim_help = "The z-axis limits."
     normal_help = "The normal direction"
 
     # Construct parser and parse
@@ -42,11 +44,13 @@ def parse_args():
     parser.add_argument('-b', '--bounds', nargs=2, type=float, metavar=('LOWER', 'UPPER'), help=bounds_help)
     parser.add_argument('-c', '--cmap', metavar=('NAME',), help=cmap_help)
     parser.add_argument('--log', action='store_true', help=log_help)
+    parser.add_argument('--linthresh', type=float, help=linthresh_help)
     parser.add_argument('-t', '--time', type=int, metavar=('PRECISION',), help=time_help)
     parser.add_argument('-e', '--ext', type=lambda s: s.lower(), default='png', help=ext_help)
     parser.add_argument('-s', '--sort', type=float, default=0.0, help=sort_help)
     parser.add_argument('-x', '--xlim', nargs=2, type=float, metavar=('UPPER', 'LOWER'), help=xlim_help)
     parser.add_argument('-y', '--ylim', nargs=2, type=float, metavar=('UPPER', 'LOWER'), help=ylim_help)
+    parser.add_argument('-z', '--zlim', nargs=2, type=float, metavar=('UPPER', 'LOWER'), help=zlim_help)
     parser.add_argument('-n', '--normal', default='z', help=normal_help)
 
     return parser.parse_args(sys.argv[1:])
@@ -96,11 +100,15 @@ def plot_generator(args):
     def get_width(ds, xlim=None, ylim=None, zlim=None):
         """ Get the width of the plot. """
 
-        if xlim is None: xlim = ds.domain_left_edge[0], ds.domain_right_edge[0]
-        else: xlim = xlim[0] * cm, xlim[1] * cm
+        if xlim is None: 
+            xlim = ds.domain_left_edge[0], ds.domain_right_edge[0]
+        else: 
+            xlim = xlim[0] * cm, xlim[1] * cm
 
-        if ylim is None: ylim = ds.domain_left_edge[1], ds.domain_right_edge[1]
-        else: ylim = ylim[0] * cm, ylim[1] * cm
+        if ylim is None: 
+            ylim = ds.domain_left_edge[1], ds.domain_right_edge[1]
+        else: 
+            ylim = ylim[0] * cm, ylim[1] * cm
 
         xwidth = (xlim[1] - xlim[0]).in_cgs()
         ywidth = (ylim[1] - ylim[0]).in_cgs()
@@ -108,8 +116,10 @@ def plot_generator(args):
         if ds.domain_dimensions[2] == 1:
             zwidth = 0.0
         else:
-            if zlim is None: zlim = ds.domain_left_edge[2], ds.domain_right_edge[2]
-            else: zlim = zlim[0] * cm, zlim[1] * cm
+            if zlim is None: 
+                zlim = ds.domain_left_edge[2], ds.domain_right_edge[2]
+            else: 
[…]
