+++ sphinx_docs/source/filelist.rst
@@ -1,14 +1,0 @@
-File list
-=========
-
-.. toctree::
-   :maxdepth: 2
-
-   file/BaseState_8H.rst
-   file/Maestro_8H.rst
-   file/MaestroBCThreads_8H.rst
-   file/MaestroPlot_8H.rst
-   file/PhysBCFunctMaestro_8H.rst
-   Src_nd_files
-   param_includes_files
-   
