+++ sphinx_docs/source/Src_nd_files.rst
@@ -1,52 +1,0 @@
-Src_nd
-======
-
-.. toctree::
-   :maxdepth: 2
-
-   file/advect_base_f
-   file/average_f
-   file/base_state_f
-   file/base_state_geometry_f
-   file/bc_fill_nd_f
-   file/burner_loop_f
-   file/cell_to_edge_f
-   file/compute_dt_f
-   file/compute_grad_phi_f
-   file/debug_f
-   file/diag_f
-   file/enforce_HSE_f
-   file/fill_3d_data_f
-   file/fill_umac_ghost_f
-   file/initdata_f
-   file/inlet_bc_f
-   file/mac_solver_f
-   file/maestro_init_f
-   file/make_S_f
-   file/make_beta0_f
-   file/make_edge_scal_f
-   file/make_edge_state_f
-   file/make_eta_f
-   file/make_explicit_thermal_f
-   file/make_flux_f
-   file/make_gamma_f
-   file/make_grav_f
-   file/make_heating_f
-   file/make_pi_cc_f
-   file/make_plot_variables_f
-   file/make_psi_f
-   file/make_scal_force_f
-   file/make_vel_force_f
-   file/make_w0_f
-   file/mkutrans_f
-   file/ppm_f
-   file/regrid_f
-   file/rhoh_vs_t_f
-   file/rotation_f
-   file/slope_f
-   file/sponge_f
-   file/tagging_f
-   file/update_scal_f
-   file/update_vel_f
-   file/velpred_f
-   
