+++ sphinx_docs/source/param_includes_files.rst
@@ -1,12 +1,0 @@
-Param_includes
-==============
-
-.. toctree::
-   :maxdepth: 2
-
-   file/maestro__defaults_8H
-   file/maestro__params_8H
-   file/maestro__queries_8H
-   file/meth_params_f
-   file/state_sizes_f
-   
