### 2019-10-24 638c50f
outflow should be neumann for thermal operator and solver
