### 2019-10-24 0152621
update inflow boundary conditions and enthalpy_pred_types in sdc subroutines
