+++ Source/MaestroThermal.cpp
@@ -628,7 +628,7 @@ Maestro::ThermalConductSDC (int which_step,
         else {
             // lo-side BCs
             if (bcs_s[RhoH].lo(idim) == BCType::foextrap) {
-                mlmg_lobc[idim] = LinOpBCType::Dirichlet;
+                mlmg_lobc[idim] = LinOpBCType::Neumann;
             } else if (bcs_s[RhoH].lo(idim) == BCType::ext_dir) {
                 mlmg_lobc[idim] = LinOpBCType::Dirichlet;
             } else {
@@ -637,7 +637,7 @@ Maestro::ThermalConductSDC (int which_step,
 
             // hi-side BCs
             if (bcs_s[RhoH].hi(idim) == BCType::foextrap) {
-                mlmg_hibc[idim] = LinOpBCType::Dirichlet;
+                mlmg_hibc[idim] = LinOpBCType::Neumann;
             } else if (bcs_s[RhoH].hi(idim) == BCType::ext_dir) {
                 mlmg_hibc[idim] = LinOpBCType::Dirichlet;
             } else {
