### 2019-10-25 34bed0a
fixed sign of div(hcoeff grad h) term
