### 2019-10-25 bbd0eae
fix for dpdt_factor>0 to work properly
