### 2019-10-25 03ef599
fixed one restart bug (still may be others).  Need to fill in w0_cart after reading in w0 from checkpoint. This problem was introduced once we decided to make w0_cart persistent for GPU purposes
