+++ Source/MaestroRegrid.cpp
@@ -379,7 +379,7 @@ Maestro::MakeNewLevelFromCoarse (int lev, Real time, const BoxArray& ba,
     S_cc_new[lev].define          (ba, dm,              1, 0);
     gpi[lev].define               (ba, dm, AMREX_SPACEDIM, 0);
     dSdt[lev].define              (ba, dm,              1, 0);
-    w0_cart[lev].define           (ba, dm, AMREX_SPACEDIM, 1);
+    w0_cart[lev].define           (ba, dm, AMREX_SPACEDIM, 2);
     rhcc_for_nodalproj[lev].define(ba, dm,              1, 1);
 
     pi[lev].define(convert(ba,nodal_flag), dm, 1, 0);     // nodal
