### 2019-10-28 4207025
put the HfromRhoTedge routines on the gpu
