+++ Source/Src_nd/rhoh_vs_t.F90
@@ -364,11 +364,7 @@ contains
   end subroutine makeCsfromRhoH
 
 
-  subroutine makeHfromRhoT_edge(lo,hi,sedgex,x_lo,x_hi, &
-       sedgey,y_lo,y_hi, &
-#if (AMREX_SPACEDIM == 3)
-       sedgez,z_lo,z_hi, &
-#endif
+  subroutine makeHfromRhoT_edge(lo,hi,idir,sedge,x_lo,x_hi, &
        rho0_cart,r_lo,r_hi, &
        rhoh0_cart,rh_lo,rh_hi, &
        temp0_cart,t_lo,t_hi, &
@@ -380,14 +376,9 @@ contains
     use meth_params_module, only: species_pred_type, enthalpy_pred_type, small_temp
     
     integer         , intent (in   ) :: lo(3), hi(3)
+    integer,   value, intent (in   ) :: idir
     integer         , intent (in   ) :: x_lo(3), x_hi(3)
-    double precision, intent (inout) :: sedgex(x_lo(1):x_hi(1),x_lo(2):x_hi(2),x_lo(3):x_hi(3),nscal)
-    integer         , intent (in   ) :: y_lo(3), y_hi(3)
-    double precision, intent (inout) :: sedgey(y_lo(1):y_hi(1),y_lo(2):y_hi(2),y_lo(3):y_hi(3),nscal)
-#if (AMREX_SPACEDIM == 3)
-    integer         , intent (in   ) :: z_lo(3), z_hi(3)
-    double precision, intent (inout) :: sedgez(z_lo(1):z_hi(1),z_lo(2):z_hi(2),z_lo(3):z_hi(3),nscal)
-#endif
+    double precision, intent (inout) :: sedge(x_lo(1):x_hi(1),x_lo(2):x_hi(2),x_lo(3):x_hi(3),nscal)
     integer         , intent (in   ) :: r_lo(3), r_hi(3)
     double precision, intent (in   ) :: rho0_cart (r_lo(1):r_hi(1),r_lo(2):r_hi(2),r_lo(3):r_hi(3))
     integer         , intent (in   ) :: rh_lo(3), rh_hi(3)
@@ -406,224 +397,229 @@ contains
     integer :: pt_index(3)
     type (eos_t) :: eos_state
 
-!!$    !$gpu
+    !$gpu
 
-    ! x-edge
-    do k = lo(3), hi(3)
-       do j = lo(2), hi(2)
-          do i = lo(1), hi(1)+1
-             
-             ! get edge-centered temperature
-             if (enthalpy_pred_type .eq. predict_Tprime_then_h) then
-                eos_state%T = max(sedgex(i,j,k,temp_comp)+temp0_cart(i,j,k),small_temp)
-             else
-                eos_state%T = max(sedgex(i,j,k,temp_comp),small_temp)
-             end if
+    if (idir == 1) then
+        ! x-edge
+        do k = lo(3), hi(3)
+            do j = lo(2), hi(2)
+                do i = lo(1), hi(1)
+                    
+                    ! get edge-centered temperature
+                    if (enthalpy_pred_type .eq. predict_Tprime_then_h) then
+                        eos_state%T = max(sedge(i,j,k,temp_comp)+temp0_cart(i,j,k),small_temp)
+                    else
+                        eos_state%T = max(sedge(i,j,k,temp_comp),small_temp)
+                    end if
 
-             ! get edge-centered density and species
-             if (species_pred_type .eq. predict_rhoprime_and_X) then
-                
-                ! interface states are rho' and X
-                eos_state%rho = sedgex(i,j,k,rho_comp) + rho0_cart(i,j,k)
+                    ! get edge-centered density and species
[…]
