### 2019-10-29 29b98a3
fixed bug in 3d rhoh_vs_t
