### 2019-10-29 17c3dd2
disable openMP from tagging routines for the moment; planar version causes segfaults
