+++ Source/MaestroRegrid.cpp
@@ -206,21 +206,60 @@ Maestro::ErrorEst (int lev, TagBoxArray& tags, Real time, int ng)
     const int clearval = TagBox::CLEAR;
     const int   tagval = TagBox::SET;
 
-    const Real* dx      = geom[lev].CellSize();
+    const Real* dx = geom[lev].CellSize();
 
     const MultiFab& state = sold[lev];
+    
+    Vector<int> itags;
 
-#ifdef _OPENMP
-#pragma omp parallel
-#endif
-    {
-        Vector<int>  itags;
+    // if you add openMP here, make sure to collect tag_array across threads
+    for (MFIter mfi(state, true); mfi.isValid(); ++mfi) {
 
+        const Box& tilebox  = mfi.tilebox();
+
+        TagBox& tagfab  = tags[mfi];
+
+        // We cannot pass tagfab to Fortran becuase it is BaseFab<char>.
+        // So we are going to get a temporary integer array.
+        // set itags initially to 'untagged' everywhere
+        // we define itags over the tilebox region
+        tagfab.get_itags(itags, tilebox);
+
+        // data pointer and index space
+        int*        tptr    = itags.dataPtr();
+        const int*  tlo     = tilebox.loVect();
+        const int*  thi     = tilebox.hiVect();
+
+        // tag cells for refinement
+        // for planar problems, we keep track of when a cell at a particular
+        // latitude is tagged using tag_array
+        state_error(tptr,  ARLIM_3D(tlo), ARLIM_3D(thi),
+                    BL_TO_FORTRAN_3D(state[mfi]),
+                    &tagval, &clearval,
+                    ARLIM_3D(tilebox.loVect()), ARLIM_3D(tilebox.hiVect()),
+                    ZFILL(dx), &time,
+                    &lev, tag_array.dataPtr());
+
+        //
+        // Now update the tags in the TagBox in the tilebox region
+        // to be equal to itags
+        //
+        tagfab.tags_and_untags(itags, tilebox);
+    }
+
+    // for planar refinement, we need to gather tagged entries in arrays
+    // from all processors and then re-tag tileboxes across each tagged
+    // height
+    if (spherical == 0) {
+        
+        ParallelDescriptor::ReduceIntMax(tag_array.dataPtr(),(max_radial_level+1)*nr_fine);
+
+        // NOTE: adding OpenMP breaks the code - not exactly sure why
         for (MFIter mfi(state, true); mfi.isValid(); ++mfi) {
 
             const Box& tilebox  = mfi.tilebox();
 
-            TagBox&     tagfab  = tags[mfi];
+            TagBox& tagfab  = tags[mfi];
 
             // We cannot pass tagfab to Fortran becuase it is BaseFab<char>.
             // So we are going to get a temporary integer array.
@@ -233,60 +272,19 @@ Maestro::ErrorEst (int lev, TagBoxArray& tags, Real time, int ng)
             const int*  tlo     = tilebox.loVect();
             const int*  thi     = tilebox.hiVect();
 
-            // tag cells for refinement
-	    // for planar problems, we keep track of when a cell at a particular
-            // latitude is tagged using tag_array
-            state_error(tptr,  ARLIM_3D(tlo), ARLIM_3D(thi),
-                        BL_TO_FORTRAN_3D(state[mfi]),
-                        &tagval, &clearval,
-                        ARLIM_3D(tilebox.loVect()), ARLIM_3D(tilebox.hiVect()),
-                        ZFILL(dx), &time,
-			&lev, tag_array.dataPtr());
+            // tag all cells at a given height if any cells at that height were tagged
[…]
