### 2019-10-30 9ab5232
GPU make edge scal 3d (GPU hydro PR 7/n) (#82)
This offloads make_edge_scal_3d to the GPU. It has been tested for the 3d reacting bubble & wdconvect problems compiled for MPI and GPU.

This also adds in the CUDA launch statements (and does some restructuring of the ppm-dependent routines to enable register capping) so all of the hydro routines offloaded so far now actually run on the GPU.
