### 2019-10-31 #1757 @zingale
@pgrete I think we've addressed most of the issues.  Can you take a look?
