### 2019-11-03 AMReX-Astro/MAESTROeX#110 @pgrete
For the purpose of the JOSS review this issue may be closed.
However, feel free to keep it open as a reminder in case you consider adding pull request and/or issue templates.
