### 2019-11-03 #1757 @pgrete
Looks good so far. Just testing the testing and GPU instructions.

@arfon what's your take on " State of the field: Do the authors describe how this software compares to other commonly-used packages?" with respect o referring to the MAESTROeX paper instead, see https://github.com/AMReX-Astro/MAESTROeX/issues/107 ?
