### 2019-11-03 AMReX-Astro/MAESTROeX#113 @pgrete
I tried to compile the example using the GPU build instructions and came across:
```bash
$ make
[...]
Loading /mnt/research/turbulence/src/amrex/Tools/GNUMake/comps/pgi.mak...
Loading /mnt/research/turbulence/src/amrex/Tools/GNUMake/sites/Make.unknown...
Compiling MaestroRhoHT.cpp ...
nvcc -ccbin=g++ -Xcompiler=' -Werror=return-type -g -O3  -pthread --std=c++14' --std=c++14 -Wno-deprecated-gpu-targets -m64 -arch=compute_60 -code=sm_60 -maxrregcount=255 -lineinfo --ptxas-options=-O3 --ptxas-options=-v --use_fast_math  -dc -x cu --expt-relaxed-constexpr --expt-extended-lambda -DNDEBUG -DCUDA -DAMREX_GIT_VERSION=\"19.11\" -DAMREX_USE_CUDA_FORTRAN -DAMREX_NVCC_VERSION=10.0 -DAMREX_NVCC_MAJOR_VERSION=10 -DAMREX_NVCC_MINOR_VERSION=0 -DBL_GCC_VERSION=6.4.0 -DBL_GCC_MAJOR_VERSION=6 -DBL_GCC_MINOR_VERSION=4 -DAMREX_USE_CUDA -DAMREX_USE_GPU_PRAGMA -DAMREX_LAUNCH='attributes(global)' -DAMREX_DEVICE='attributes(device)' -DAMREX_CUDA_FORT_GLOBAL='attributes(global)' -DAMREX_CUDA_FORT_DEVICE='attributes(device)' -DAMREX_CUDA_FORT_HOST='attributes(host)' -DAMREX_CUDA_FORT_HOST_DEVICE='attributes(host) attributes(device)' -DAMREX_USE_GPU -DBL_COALESCE_FABS -DBL_SPACEDIM=2 -DAMREX_SPACEDIM=2 -DBL_FORT_USE_UNDERSCORE -DAMREX_FORT_USE_UNDERSCORE -DBL_Linux -DAMREX_Linux -DAMREX_GPU_MAX_THREADS=256 -DCRSEGRNDOMP -DEXTRA_THERMO -DINTEGRATOR=3 -DCUDA -Itmp_build_dir/s/2d.pgi.CUDA.EXE -I. -I/opt/software/CUDA/10.0.130/include -I. -I../../../Source -I../../../Source/Src_nd -I../../../Source/param_includes -I../../../constants -I../../../Util/model_parser -I../../../Util/simple_log -I../../../Microphysics/EOS -I/mnt/research/turbulence/src/Microphysics/EOS/helmholtz -I../../../Microphysics/networks -I/mnt/research/turbulence/src/Microphysics/networks/ignition_simple -I/mnt/research/turbulence/src/Microphysics/EOS -I/mnt/research/turbulence/src/Microphysics/networks -I../../../Microphysics/conductivity -I../../../Microphysics/conductivity -I/mnt/research/turbulence/src/Microphysics/conductivity/stellar -I/mnt/research/turbulence/src/amrex/Src/Base -I/mnt/research/turbulence/src/amrex/Src/Boundary -I/mnt/research/turbulence/src/amrex/Src/AmrCore -I/mnt/research/turbulence/src/amrex/Src/LinearSolvers/C_CellMG -I/mnt/research/turbulence/src/amrex/Src/LinearSolvers/MLMG -I/mnt/research/turbulence/src/Microphysics/util -I/mnt/research/turbulence/src/Microphysics/integration/VODE90 -I/mnt/research/turbulence/src/Microphysics/integration/VODE90/cuVODE/source/ -I/mnt/research/turbulence/src/Microphysics/integration/utils -I/mnt/research/turbulence/src/Microphysics/integration -I/mnt/research/turbulence/src/Microphysics/screening -I. -I../../../Source -I../../../Source/Src_nd -I../../../Source/param_includes -I../../../constants -I../../../Util/model_parser -I../../../Util/simple_log -I../../../Microphysics/EOS -I/mnt/research/turbulence/src/Microphysics/EOS/helmholtz -I../../../Microphysics/networks -I/mnt/research/turbulence/src/Microphysics/networks/ignition_simple -I/mnt/research/turbulence/src/Microphysics/EOS -I/mnt/research/turbulence/src/Microphysics/networks -I../../../Microphysics/conductivity -I../../../Microphysics/conductivity -I/mnt/research/turbulence/src/Microphysics/conductivity/stellar -I/mnt/research/turbulence/src/amrex/Tools/C_scripts -c tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp -o tmp_build_dir/o/2d.pgi.CUDA.EXE/MaestroRhoHT.o

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(542): error: argument of type "int" is incompatible with parameter of type "const int *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(542): error: argument of type "int" is incompatible with parameter of type "const int *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(543): error: argument of type "int" is incompatible with parameter of type "const int *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(543): error: argument of type "int" is incompatible with parameter of type "const int *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(544): error: argument of type "amrex::Real *" is incompatible with parameter of type "const int *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(544): error: argument of type "const int *" is incompatible with parameter of type "const amrex::Real *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(545): error: argument of type "amrex::Real *" is incompatible with parameter of type "const int *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(545): error: argument of type "const int *" is incompatible with parameter of type "const amrex::Real *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(549): error: argument of type "const amrex::Real *" is incompatible with parameter of type "const int *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(549): error: argument of type "const int *" is incompatible with parameter of type "const amrex::Real *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(550): error: argument of type "const amrex::Real *" is incompatible with parameter of type "const int *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(550): error: argument of type "const int *" is incompatible with parameter of type "const amrex::Real *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(551): error: argument of type "const amrex::Real *" is incompatible with parameter of type "const int *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(551): error: argument of type "const int *" is incompatible with parameter of type "const amrex::Real *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(552): error: argument of type "const amrex::Real *" is incompatible with parameter of type "const int *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(552): error: argument of type "const int *" is incompatible with parameter of type "const amrex::Real *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(553): error: argument of type "const amrex::Real *" is incompatible with parameter of type "const int *"

tmp_build_dir/s/2d.pgi.CUDA.EXE/MaestroRhoHT.cpp(553): error: too many arguments in function call

18 errors detected in the compilation of "/tmp/tmpxft_00007d38_00000000-6_MaestroRhoHT.cpp1.ii".
make: *** [tmp_build_dir/o/2d.pgi.CUDA.EXE/MaestroRhoHT.o] Error 1
```

here are the changes I made to the `GNUmakefile` file
```diff
diff --git a/Exec/TEST_PROBLEMS/reacting_bubble/GNUmakefile b/Exec/TEST_PROBLEMS/reacting_bubble/GNUmakefile
index 2a0099d..94bcfde 100644
--- a/Exec/TEST_PROBLEMS/reacting_bubble/GNUmakefile
+++ b/Exec/TEST_PROBLEMS/reacting_bubble/GNUmakefile
@@ -1,7 +1,9 @@
 DEBUG      = FALSE
-DIM        = 3
-COMP      = gnu
-USE_MPI    = TRUE
+DIM        = 2
+COMP      := pgi
+USE_CUDA := TRUE
+INTEGRATOR_DIR   := VODE90
+USE_MPI    = FALSE
 USE_OMP    = FALSE
 USE_REACT  = TRUE
```

and here's my environment

```bash
Currently Loaded Modules:
  1) tbb/2018_U3    4) Java/1.8.0_152   7) numactl/2.0.11  10) PGI/18.4-GCC-6.4.0-2.28  13) zlib/1.2.11      16) SQLite/3.21.0  19) CUDA/10.0.130
  2) CMake/3.11.1   5) MATLAB/2018a     8) GCCcore/6.4.0   11) Boost/1.67.0             14) ncurses/6.0      17) GMP/6.1.2
  3) Tcl/8.6.8      6) powertools/1.2   9) binutils/2.28   12) bzip2/1.0.6              15) libreadline/7.0  18) libffi/3.2.1
```
