### 2019-11-03 809c64a
fix gpu compilation
