### 2019-11-03 AMReX-Astro/MAESTROeX#113 @harpolea
That looks like it was a bug from a recent commit. I've fixed it in 809c64aa so it should compile now (on the development branch).
