### 2019-11-03 AMReX-Astro/MAESTROeX#111 @harpolea
Nightly regression tests are triggered automatically, however all other tests (such as those run on PRs) are triggered manually. When triggered manually, the test suite can then be configured to point at e.g. a branch or fork and we can customise which tests are run.
