### 2019-11-03 cddc789
use the new composition_derivatives routine
