### 2019-11-04 4a69946
fix GPU compilation for 3d
