### 2019-11-04 06f05e0
move import statements to fix GPU compilation
