### 2019-11-04 f5a021d
remove X derivatives from eos_t (#119)
This removes the species derivatives from the eos_t to save memory and boost GPU performance. This coordinates with starkiller-astro/Microphysics#207
