+++ Source/Src_nd/make_S.F90
@@ -1,8 +1,6 @@
 module make_S_module
 
   use amrex_error_module
-  use eos_type_module
-  use eos_module
   use network, only: nspec
   use meth_params_module, only: rho_comp, temp_comp, spec_comp, nscal, dpdt_factor, base_cutoff_density, use_delta_gamma1_term
   use base_state_geometry_module, only:  max_radial_level, nr_fine, base_cutoff_density_coord, anelastic_cutoff_density_coord, nr, dr
@@ -26,6 +24,10 @@ contains
        p0_cart, p0_lo, p0_hi, &
        gamma1bar_cart, g1_lo, g1_hi, &
        dx) bind (C,name="make_S_cc")
+
+    use eos_type_module, only : eos_t, eos_input_rt
+    use eos_module
+    use eos_composition_module, only : eos_xderivs_t, composition_derivatives
 
     integer  , value, intent (in   ) :: lev
     integer         , intent (in   ) :: lo(3), hi(3)
@@ -56,6 +58,7 @@ contains
     integer i,j,k,r
     integer pt_index(3)
     type(eos_t) :: eos_state
+    type(eos_xderivs_t) :: eos_xderivs
 
     integer comp
     double precision sigma, xi_term, pres_term, gradp0
@@ -76,6 +79,8 @@ contains
              ! dens, temp, and xmass are inputs
              call eos(eos_input_rt, eos_state, pt_index)
 
+             call composition_derivatives(eos_state, eos_xderivs)
+
              sigma = eos_state%dpdt / &
                   (eos_state%rho * eos_state%cp * eos_state%dpdr)
 
@@ -83,10 +88,10 @@ contains
              pres_term = 0.d0
              do comp = 1, nspec
                 xi_term = xi_term - &
-                     eos_state%dhdX(comp)*rodot(i,j,k,comp)/eos_state%rho
+                     eos_xderivs % dhdX(comp)*rodot(i,j,k,comp)/eos_state%rho
 
                 pres_term = pres_term + &
-                     eos_state%dpdX(comp)*rodot(i,j,k,comp)/eos_state%rho
+                     eos_xderivs % dpdX(comp)*rodot(i,j,k,comp)/eos_state%rho
              enddo
 
              S_cc(i,j,k) = (sigma/eos_state%rho) * &
@@ -146,6 +151,10 @@ contains
        gradp0_cart, gp0_lo, gp0_hi, &
        gamma1bar_cart, g1_lo, g1_hi, &
        normal, no_lo, no_hi) bind (C,name="make_S_cc_sphr")
+       
+    use eos_type_module, only : eos_t, eos_input_rt
+    use eos_module
+    use eos_composition_module, only : eos_xderivs_t, composition_derivatives
 
     integer         , intent (in   ) :: lo(3), hi(3)
     integer         , intent (in   ) :: s_lo(3), s_hi(3)
@@ -178,7 +187,7 @@ contains
     integer i,j,k,r
     integer pt_index(3)
     type(eos_t) :: eos_state
-
+    type(eos_xderivs_t) :: eos_xderivs
     integer comp
     double precision sigma, xi_term, pres_term, Ut_dot_er
 
@@ -198,6 +207,8 @@ contains
              ! dens, temp, and xmass are inputs
              call eos(eos_input_rt, eos_state, pt_index)
 
+             call composition_derivatives(eos_state, eos_xderivs)
+
              sigma = eos_state%dpdt / &
                   (eos_state%rho * eos_state%cp * eos_state%dpdr)
 
@@ -205,10 +216,10 @@ contains
              pres_term = 0.d0
              do comp = 1, nspec
                 xi_term = xi_term - &
[…]
