### 2019-11-05 0105ab5
Add evolve_base_state=T option for sdc solver. Currently tested on reacting_bubble_2d.
