+++ Source/MaestroAdvanceSdc.cpp
@@ -114,10 +114,6 @@ Maestro::AdvanceTimeStepSDC (bool is_initIter) {
     Print() << "\nTimestep " << istep << " starts with TIME = " << t_old
 	    << " DT = " << dt << std::endl << std::endl;
 
-    if (evolve_base_state) {
-	Abort("evolve_base_state not supported with SDC");
-    } 
-
     if (maestro_verbose > 0) {
     	Print() << "Cell Count:" << std::endl;
     	for (int lev=0; lev<=finest_level; ++lev) {
@@ -292,7 +288,6 @@ Maestro::AdvanceTimeStepSDC (bool is_initIter) {
 	}
     }
     
-/*
     if (evolve_base_state) {
 
         if (split_projection) {
@@ -317,7 +312,6 @@ Maestro::AdvanceTimeStepSDC (bool is_initIter) {
             }
         }
     }
-*/
 
     
     // compute unprojected MAC velocities
@@ -329,22 +323,22 @@ Maestro::AdvanceTimeStepSDC (bool is_initIter) {
 	macphi   [lev].setVal(0.);
 	delta_gamma1_term[lev].setVal(0.);
     }
-/*
+
     if (evolve_base_state && !split_projection) {
 	for (int i=0; i<Sbar.size(); ++i) {
             Sbar[i] = 1.0/(gamma1bar_old[i]*p0_old[i]) * (p0_old[i] - p0_nm1[i])/dtold;
 	}
     }
-*/
+
     // compute RHS for MAC projection, beta0*(S_cc-Sbar) + beta0*delta_chi
     MakeRHCCforMacProj(macrhs,rho0_old,S_cc_nph,Sbar,beta0_old,delta_gamma1_term,
 		       gamma1bar_old,p0_old,delta_p_term,delta_chi,is_predictor);
-/*
+
     if (spherical == 1 && evolve_base_state && split_projection) {
 	// subtract w0mac from umac
 	Addw0(umac,w0mac,-1.);
     }
-*/
+
     // wallclock time
     Real start_total_macproj = ParallelDescriptor::second();
 
@@ -356,12 +350,12 @@ Maestro::AdvanceTimeStepSDC (bool is_initIter) {
     Real end_total_macproj = ParallelDescriptor::second() - start_total_macproj;
     ParallelDescriptor::ReduceRealMax(end_total_macproj,ParallelDescriptor::IOProcessorNumber());
 
-/*
+    
     if (spherical == 1 && evolve_base_state && split_projection) {
 	// add w0mac back to umac
 	Addw0(umac,w0mac,1.);
     }
-*/
+
 
     //////////////////////////////////////////////////////////////////////////////
     // STEP 2 -- Predictor
@@ -418,20 +412,11 @@ Maestro::AdvanceTimeStepSDC (bool is_initIter) {
     // advect rhoX and rho
     DensityAdvanceSDC(1,sold,shat,sedge,sflux,scal_force,etarhoflux_dummy,umac,w0mac_dummy,rho0_pred_edge_dummy);
     
-/*
+
     if (evolve_base_state) {
 	// correct the base state density by "averaging"
-	Average(s2, rho0_new, Rho);
+	Average(shat, rho0_new, Rho);
 	compute_cutoff_coords(rho0_new.dataPtr());
-
-	if (use_etarho) {
-            // compute the new etarho
-            if (spherical == 0) {
-                MakeEtarho(etarho_ec,etarho_cc,etarhoflux_dummy);
[…]
