### 2019-11-07 65c30c5
re-add X derivatives to eos_t to compile with Microphysics sdc-maestroex branch
