### 2019-11-10 #1757 @arfon
> @arfon what's your take on " State of the field: Do the authors describe how this software compares to other commonly-used packages?" with respect o referring to the MAESTROeX paper instead, see AMReX-Astro/MAESTROeX#107 ?

Do you mean resolve this item by simply citing the AAS paper?
