### 2019-11-13 AMReX-Astro/MAESTROeX#113 @pgrete
I was now able to compile the dev version, but running the resulting executable leads to a segfault
```bash
$ srun -n 1 ./Maestro2d.pgi.CUDA.ex inputs_2d_C
Initializing CUDA...
CUDA initialized with 1 GPU
Segfault
See Backtrace.0 file for details
srun: error: nvl-001: task 0: Exited with exit code 1

$ cat Backtrace.0
=== If no file names and line numbers are shown below, one can run
            addr2line -Cpfie my_exefile my_line_address
    to convert `my_line_address` (e.g., 0x4a6b) into file name and line number.
    Or one can use amrex/Tools/Backtrace/parse_bt.py.

=== Please note that the line number reported by addr2line may not be accurate.
    One can use
            readelf -wl my_exefile | grep my_line_address'
    to find out the offset for that line.

 0: /mnt/research/turbulence/src/MAESTROeX/Exec/TEST_PROBLEMS/reacting_bubble/./Maestro2d.pgi.CUDA.ex() [0x777156]
    amrex::BLBackTrace::print_backtrace_info(_IO_FILE*) at /mnt/research/turbulence/src/MAESTROeX/Exec/TEST_PROBLEMS/reacting_bubble/tmp_build_dir/s/2d.pgi.CUDA.EXE/AMReX_BLBackTrace.cpp:167

 1: /mnt/research/turbulence/src/MAESTROeX/Exec/TEST_PROBLEMS/reacting_bubble/./Maestro2d.pgi.CUDA.ex() [0x7796f2]
    amrex::BLBackTrace::handler(int) at /mnt/research/turbulence/src/MAESTROeX/Exec/TEST_PROBLEMS/reacting_bubble/tmp_build_dir/s/2d.pgi.CUDA.EXE/AMReX_BLBackTrace.cpp:71

 2: /lib64/libc.so.6(+0x36280) [0x2ac608fd6280]

 3: /mnt/research/turbulence/src/MAESTROeX/Exec/TEST_PROBLEMS/reacting_bubble/./Maestro2d.pgi.CUDA.ex() [0x69b56f]
    amrex::DeallocateRandomSeedDevArray() at /mnt/research/turbulence/src/MAESTROeX/Exec/TEST_PROBLEMS/reacting_bubble/tmp_build_dir/s/2d.pgi.CUDA.EXE/AMReX_Random.cpp:333
 (inlined by) amrex::InitRandom(unsigned long, int) at /mnt/research/turbulence/src/MAESTROeX/Exec/TEST_PROBLEMS/reacting_bubble/tmp_build_dir/s/2d.pgi.CUDA.EXE/AMReX_Random.cpp:60

 4: /mnt/research/turbulence/src/MAESTROeX/Exec/TEST_PROBLEMS/reacting_bubble/./Maestro2d.pgi.CUDA.ex() [0x676c5b]
    amrex::Initialize(int&, char**&, bool, int, std::function<void ()> const&, std::ostream&, std::ostream&, void (*)(char const*)) at /mnt/research/turbulence/src/MAESTROeX/Exec/TEST_PROBLEMS/reacting_bubble/tmp_build_dir/s/2d.pgi.CUDA.EXE/AMReX.cpp:490

 5: /mnt/research/turbulence/src/MAESTROeX/Exec/TEST_PROBLEMS/reacting_bubble/./Maestro2d.pgi.CUDA.ex() [0x40d602]
    main at /mnt/research/turbulence/src/MAESTROeX/Exec/TEST_PROBLEMS/reacting_bubble/tmp_build_dir/s/2d.pgi.CUDA.EXE/main.cpp:24

 6: /lib64/libc.so.6(__libc_start_main+0xf5) [0x2ac608fc23d5]

 7: /mnt/research/turbulence/src/MAESTROeX/Exec/TEST_PROBLEMS/reacting_bubble/./Maestro2d.pgi.CUDA.ex() [0x412dc2]
    _start at ??:?
```
