### 2019-11-13 #1757 @pgrete
Yes, that' what I meant.

Apart from this and the pending verification of the GPU capabilities this looks good to me.
