### 2019-11-13 AMReX-Astro/MAESTROeX#113 @zingale
Hi Philipp, where are you running, and what model GPU are you using?
