### 2019-11-13 AMReX-Astro/MAESTROeX#113 @zingale
can you post the output of `./Maestro2d.pgi.CUDA.ex --describe`?
