### 2019-11-13 AMReX-Astro/MAESTROeX#113 @pgrete
I'm trying to run on MSU's cluster using V100 GPUs.

```
$ srun -n 1 ./Maestro2d.pgi.CUDA.ex --describe
==============================================================================
 MAESTROeX Build Information
==============================================================================
build date:    2019-11-13 12:52:24.642234
build machine: Linux nvl-001 3.10.0-957.5.1.el7.x86_64 #1 SMP Fri Feb 1 14:54:57 UTC 2019 x86_64 x86_64 x86_64 GNU/Linux
build dir:     /mnt/research/turbulence/src/MAESTROeX/Exec/TEST_PROBLEMS/reacting_bubble
AMReX dir:     /mnt/research/turbulence/src/amrex

COMP:          pgi
COMP version:  6.4.0

C++ compiler:  nvcc
C++ flags:     -ccbin=g++ -Xcompiler=' -Werror=return-type -g -O3  -pthread --std=c++14' --std=c++14 -Wno-deprecated-gpu-targets -m64 -arch=compute_60 -code=sm_60 -maxrregcount=255 -lineinfo
 --ptxas-options=-O3 --ptxas-options=-v --use_fast_math  -dc -x cu --expt-relaxed-constexpr --expt-extended-lambda -DNDEBUG -DCUDA -DAMREX_GIT_VERSION="19.11" -DAMREX_USE_CUDA_FORTRAN -DAMRE
X_NVCC_VERSION=10.0 -DAMREX_NVCC_MAJOR_VERSION=10 -DAMREX_NVCC_MINOR_VERSION=0 -DBL_GCC_VERSION=6.4.0 -DBL_GCC_MAJOR_VERSION=6 -DBL_GCC_MINOR_VERSION=4 -DAMREX_USE_CUDA -DAMREX_USE_GPU_PRAGM
A -DAMREX_LAUNCH='attributes(global)' -DAMREX_DEVICE='attributes(device)' -DAMREX_CUDA_FORT_GLOBAL='attributes(global)' -DAMREX_CUDA_FORT_DEVICE='attributes(device)' -DAMREX_CUDA_FORT_HOST='
attributes(host)' -DAMREX_CUDA_FORT_HOST_DEVICE='attributes(host) attributes(device)' -DAMREX_USE_GPU -DBL_COALESCE_FABS -DBL_SPACEDIM=2 -DAMREX_SPACEDIM=2 -DBL_FORT_USE_UNDERSCORE -DAMREX_F
ORT_USE_UNDERSCORE -DBL_Linux -DAMREX_Linux -DAMREX_GPU_MAX_THREADS=256 -DCRSEGRNDOMP -DEXTRA_THERMO -DINTEGRATOR=3 -DCUDA -Itmp_build_dir/s/2d.pgi.CUDA.EXE -I. -I/opt/software/CUDA/10.0.130
/include -I. -I../../../Source -I../../../Source/Src_nd -I../../../Source/param_includes -I../../../constants -I../../../Util/model_parser -I../../../Util/simple_log -I../../../Microphysics/
EOS -I/mnt/research/turbulence/src/Microphysics/EOS/helmholtz -I../../../Microphysics/networks -I/mnt/research/turbulence/src/Microphysics/networks/ignition_simple -I/mnt/research/turbulence
/src/Microphysics/EOS -I/mnt/research/turbulence/src/Microphysics/networks -I../../../Microphysics/conductivity -I../../../Microphysics/conductivity -I/mnt/research/turbulence/src/Microphysi
cs/conductivity/stellar -I/mnt/research/turbulence/src/amrex/Src/Base -I/mnt/research/turbulence/src/amrex/Src/Boundary -I/mnt/research/turbulence/src/amrex/Src/AmrCore -I/mnt/research/turbu
lence/src/amrex/Src/LinearSolvers/C_CellMG -I/mnt/research/turbulence/src/amrex/Src/LinearSolvers/MLMG -I/mnt/research/turbulence/src/Microphysics/util -I/mnt/research/turbulence/src/Microph
ysics/integration/VODE90 -I/mnt/research/turbulence/src/Microphysics/integration/VODE90/cuVODE/source/ -I/mnt/research/turbulence/src/Microphysics/integration/utils -I/mnt/research/turbulenc
e/src/Microphysics/integration -I/mnt/research/turbulence/src/Microphysics/screening -I. -I../../../Source -I../../../Source/Src_nd -I../../../Source/param_includes -I../../../constants -I..
/../../Util/model_parser -I../../../Util/simple_log -I../../../Microphysics/EOS -I/mnt/research/turbulence/src/Microphysics/EOS/helmholtz -I../../../Microphysics/networks -I/mnt/research/tur
bulence/src/Microphysics/networks/ignition_simple -I/mnt/research/turbulence/src/Microphysics/EOS -I/mnt/research/turbulence/src/Microphysics/networks -I../../../Microphysics/conductivity -I
../../../Microphysics/conductivity -I/mnt/research/turbulence/src/Microphysics/conductivity/stellar -I/mnt/research/turbulence/src/amrex/Tools/C_scripts

Fortran comp:  pgfortran
Fortran flags:  -O2 -fast -Mnomain -Mcuda=cc60,fastmath,charstring -Mcuda=lineinfo -Mcuda=ptxinfo CUDA_HOME=/opt/software/CUDA/10.0.130 -Mcuda=maxregcount:255 -Mdclchk  -noacc

Link flags:    -L. -L/opt/software/CUDA/10.0.130/lib64
Libraries:     -lcuda -lstdc++ -pgf90libs -latomic

EOS: /mnt/research/turbulence/src/Microphysics/EOS/helmholtz
NETWORK: /mnt/research/turbulence/src/Microphysics/networks/ignition_simple
CONDUCTIVITY: /mnt/research/turbulence/src/Microphysics/conductivity/stellar
MAESTROeX git describe: 19.11-7-gb7a0b0e-dirty
AMReX git describe: 19.11
Microphysics git describe: 19.11-27-ga45dc9e
```
