### 2019-11-13 AMReX-Astro/MAESTROeX#113 @zingale
can you try adding `CUDA_ARCH=70` to your `make` line?
