### 2019-11-13 293a5c0
added information about setting the CUDA compute capability to the GPU docs
