### 2019-11-13 6d7d048
specified that the deviceQuery script must be compiled first
