### 2019-11-14 AMReX-Astro/MAESTROeX#113 @pgrete
I feel like I'm doing something wrong. After a `make clean` the make command now fails with
```
Compiling main.cpp ...
nvcc -ccbin=g++ -Xcompiler=' -Werror=return-type -g -O3  -pthread --std=c++14' --std=c++14 -Wno-deprecated-gpu-targets -m64 -arch=compute_70 -code=sm_70 -maxrregcount=255 -lineinfo --ptxas-o
ptions=-O3 --ptxas-options=-v --use_fast_math  -dc -x cu --expt-relaxed-constexpr --expt-extended-lambda -DNDEBUG -DCUDA -DAMREX_GIT_VERSION=\"19.11\" -DAMREX_USE_CUDA_FORTRAN -DAMREX_NVCC_V
ERSION=10.0 -DAMREX_NVCC_MAJOR_VERSION=10 -DAMREX_NVCC_MINOR_VERSION=0 -DBL_GCC_VERSION=6.4.0 -DBL_GCC_MAJOR_VERSION=6 -DBL_GCC_MINOR_VERSION=4 -DAMREX_USE_CUDA -DAMREX_USE_GPU_PRAGMA -DAMRE
X_LAUNCH='attributes(global)' -DAMREX_DEVICE='attributes(device)' -DAMREX_CUDA_FORT_GLOBAL='attributes(global)' -DAMREX_CUDA_FORT_DEVICE='attributes(device)' -DAMREX_CUDA_FORT_HOST='attribut
es(host)' -DAMREX_CUDA_FORT_HOST_DEVICE='attributes(host) attributes(device)' -DAMREX_USE_GPU -DBL_COALESCE_FABS -DBL_SPACEDIM=2 -DAMREX_SPACEDIM=2 -DBL_FORT_USE_UNDERSCORE -DAMREX_FORT_USE_
UNDERSCORE -DBL_Linux -DAMREX_Linux -DAMREX_GPU_MAX_THREADS=256 -DCRSEGRNDOMP -DEXTRA_THERMO -DINTEGRATOR=3 -DCUDA -Itmp_build_dir/s/2d.pgi.CUDA.EXE -I. -I/opt/software/CUDA/10.0.130/include
 -I. -I../../../Source -I../../../Source/Src_nd -I../../../Source/param_includes -I../../../constants -I../../../Util/model_parser -I../../../Util/simple_log -I../../../Microphysics/EOS -I/m
nt/research/turbulence/src/Microphysics/EOS/helmholtz -I../../../Microphysics/networks -I/mnt/research/turbulence/src/Microphysics/networks/ignition_simple -I/mnt/research/turbulence/src/Mic
rophysics/EOS -I/mnt/research/turbulence/src/Microphysics/networks -I../../../Microphysics/conductivity -I../../../Microphysics/conductivity -I/mnt/research/turbulence/src/Microphysics/condu
ctivity/stellar -I/mnt/research/turbulence/src/amrex/Src/Base -I/mnt/research/turbulence/src/amrex/Src/Boundary -I/mnt/research/turbulence/src/amrex/Src/AmrCore -I/mnt/research/turbulence/sr
c/amrex/Src/LinearSolvers/C_CellMG -I/mnt/research/turbulence/src/amrex/Src/LinearSolvers/MLMG -I/mnt/research/turbulence/src/Microphysics/util -I/mnt/research/turbulence/src/Microphysics/in
tegration/VODE90 -I/mnt/research/turbulence/src/Microphysics/integration/VODE90/cuVODE/source/ -I/mnt/research/turbulence/src/Microphysics/integration/utils -I/mnt/research/turbulence/src/Mi
crophysics/integration -I/mnt/research/turbulence/src/Microphysics/screening -I. -I../../../Source -I../../../Source/Src_nd -I../../../Source/param_includes -I../../../constants -I../../../U
til/model_parser -I../../../Util/simple_log -I../../../Microphysics/EOS -I/mnt/research/turbulence/src/Microphysics/EOS/helmholtz -I../../../Microphysics/networks -I/mnt/research/turbulence/
src/Microphysics/networks/ignition_simple -I/mnt/research/turbulence/src/Microphysics/EOS -I/mnt/research/turbulence/src/Microphysics/networks -I../../../Microphysics/conductivity -I../../..
/Microphysics/conductivity -I/mnt/research/turbulence/src/Microphysics/conductivity/stellar -I/mnt/research/turbulence/src/amrex/Tools/C_scripts -c tmp_build_dir/s/2d.pgi.CUDA.EXE/main.cpp -
o tmp_build_dir/o/2d.pgi.CUDA.EXE/main.o
g++: error: tmp_build_dir/s/2d.pgi.CUDA.EXE/main.cpp: No such file or directory
g++: warning: ‘-x c++’ after last input file has no effect
g++: fatal error: no input files
compilation terminated.
make: *** [tmp_build_dir/o/2d.pgi.CUDA.EXE/main.o] Error 1
```
The environment variables for AMReX and microphysics are set
```
$ env |grep "AMR\|MIC"
AMREX_HOME=/mnt/research/turbulence/src/amrex
MICROPHYSICS_HOME=/mnt/research/turbulence/src/Microphysics
```
