### 2019-11-14 AMReX-Astro/MAESTROeX#113 @harpolea
Could you try `make realclean` and try again? The preprocessor that generates the CUDA boilerplate for GPU code creates temporary files that are not deleted by `make clean`.
