### 2019-11-14 AMReX-Astro/MAESTROeX#113 @pgrete
Great, that did the trick. Also adding a more up to date CUDA_ARCH resulted in a working executable.

I'll close this issue now, but recommend to add a hint to `make clean` to the GPU docs and potentially also that `CUDA_ARCH` is not backwards compatible (in all cases) so that it's important to set the correct version.
