### 2019-11-14 #1757 @pgrete
GPU functionality has been confirmed. Once there's a decision on the software paper related item, I'm happy to approve @arfon
