+++ docs/_sources/makefiles.rst.txt
@@ -7,158 +7,116 @@ MAESTROeX Build System
 Build Process Overview
 ======================
 
-The MAESTRO build system uses features of GNU make (version 3.81 or
-later), which is typically the default on systems. The MAESTRO executable is built in the problem’s directory (one of the directories
-under SCIENCE/, TEST_PROBLEMS, or UNIT_TESTS). This
-directory will contain a makefile, GNUmakefile, that includes
-all the necessary information to build the executable.
+The MAESTROeX build system uses features of GNU make (version 3.81 or
+later), which is typically the default on systems. The MAESTROeX
+executable is built in the problem’s directory (one of the directories
+under ``SCIENCE/``, ``TEST_PROBLEMS/``, or ``UNIT_TESTS/``). This
+directory will contain a makefile, ``GNUmakefile``, that includes all the
+necessary information to build the executable.
 
-The main macros that define the build process are split across several
-files. The 4 main files are:
+We use the build system in AMReX in ``amrex/Tools/GNUmake``.  A set of
+MAESTROeX-specific macros are defined in ``MAESTROeX/Exec/Make.Maestro``.
 
--  ${AMREX_HOME}/Tools/F_mk/GMakedefs.mak:
+MAESTRO gets the location of the AMReX library through the
+``AMREX_HOME`` variable. This should be set as an environment
+variable in your shell start-up files (e.g. ``.bashrc`` or
+``.cshrc``).
 
-   This setups the basic macros, includes the options for the selected
-   compiler, builds the list of object and source files, and defines
-   the compile and link command lines.
-
--  ${AMREX_HOME}/Tools/F_mk/GMakeMPI.mak:
-
-   This implements any changes to the compiler names and library
-   locations necessary to build a parallel executable with MPI.
-
--  ${AMREX_HOME}/Tools/F_mk/GMakerules.mak:
-
-   This creates the various build targets and specifies the rules for
-   building the object files, the list of dependencies, and some other
-   lesser-used targets (tags for editors, documentation, etc.)
-
--  MAESTRO/GMaestro.mak:
-
-   This is a MAESTRO-specific file that gathers all of the various
-   modules that are used to build a typical MAESTRO application
-   and integrates with the AMReX build system. Every MAESTRO problem’s GNUmakefile will include this file.
-
-MAESTRO gets the location of the AMReX library through the
-AMREX_HOME variable. This should be set as an environment
-variable in your shell start-up files (e.g. .bashrc or
-.cshrc).
-
-The AMReX build system separates the compiler-specific information
+The AMReX build system separates the compiler-specific information
 from the machine-specific information—this allows for reuse of the
-compiler information. The only machine-specific parts of the build system
-are for the MPI library locations, contained in GMakeMPI.mak.
-The compiler flags for the various compilers are listed in the
-files in ${AMREX_HOME}/Tools/F_mk/comps/. The compiler
-is set via the COMP variable in the problem’s GNUmakefile.
+compiler information. The only machine-specific parts of the build
+system are for the MPI library locations, contained in
+``amrex/Tools/GNUmake/sites/``.  The compiler flags for the various
+compilers are listed in the files in
+``amrex/Tools/GNUmake/comps/``. The compiler is set via the ``COMP``
+variable in the problem’s GNUmakefile.
 
 There are several options in addition to the compiler that affect the
-build process: MPI, OMP, and NDEBUG—these turn on/off
-MPI parallelization, OpenMP parallelization, and debugging. Together,
+build process: ``USE_MPI``, ``USE_OMP``, ``USE_CUDA`` for the parallelism,
+and ``DEBUG`` and ``TEST`` for diagnostic information.. Together,
[…]
