### 2019-11-14 #1757 @arfon
> okay, just to make sure I understand. Is the suggestion that we add a cite to the ApJ MAESTROeX paper in the JOSS paper.md?

Yes.
