### 2019-11-14 #1757 @pgrete
IIRC the reference is already there and my question was around whether this satisfies the "State of the field: Do the authors describe how this software compares to other commonly-used packages?" item in the reviewing checklist.
