### 2019-11-14 5548986
also compute cell-centered pi for sdc algorithm
