### 2019-11-15 #1757 @pgrete
Yes, my previous comment was meant as a clarification for why I raised the point (I just wanted to double check with the editor if this in line with the interpretation of the review instructions).
