### 2019-11-15 #1757 @tukss
Just checked that all my raised points are fixed and they are. Great!

I noticed that the first line of the bibliography reads:
> {% bibliography –cited –file paper.bib %}
I guess that's a bug somewhere in the processing of the PDF.
