### 2019-11-15 #1757 @arfon
> I guess that's a bug somewhere in the processing of the PDF.

It's in the `paper.md` - @doreenfan please merge this pull request: https://github.com/AMReX-Astro/MAESTROeX/pull/123
