### 2019-11-15 42b65b6
read in enuc and omegadot from initial data and use during initialisation rather than calculating
