+++ Source/Maestro.H
@@ -1321,6 +1321,8 @@ private:
     /// flag for writing plotfiles
     enum plotfile_flag {plotInitData = -9999999, plotInitProj = -9999998, plotDivuIter = -9999997};
 
+    #include <Problem.H>
+
 };
 
 #endif
