### 2019-11-19 66f18bb
GPU make rho flux (GPU hydro PR 8/n) (#83)
This offloads themake_rhox_flux_* and make_rhoh_flux_* routines to the GPU. It has been tested for the 2d & 3d reacting bubble and wdconvect problems compiled for MPI and GPU.

Note: It is likely that this version will not work if compiled with OpenMP (this will come in a future PR)
