+++ Source/Src_nd/make_flux.F90
@@ -54,14 +54,11 @@ module make_flux_module
 contains
 
 #if (AMREX_SPACEDIM == 2)
-  subroutine make_rhoX_flux_2d(lev, lo, hi, &
-       sfluxx, fx_lo, fx_hi, nc_fx, &
-       sfluxy, fy_lo, fy_hi, nc_fy, &
+  subroutine make_rhoX_flux_2d(lo, hi, lev, idir, &
+       sflux, fx_lo, fx_hi, nc_fx, &
        etarhoflux, eta_lo, eta_hi, &
-       sedgex, x_lo, x_hi, nc_x, &
-       sedgey, y_lo, y_hi, nc_y, &
+       sedge, x_lo, x_hi, nc_x, &
        umac,   u_lo, u_hi, &
-       vmac,   v_lo, v_hi, &
        rho0_old, rho0_edge_old, &
        rho0_new, rho0_edge_new, &
        rho0_predicted_edge, &
@@ -69,128 +66,126 @@ contains
        startcomp, endcomp) bind(C,name="make_rhoX_flux_2d")
     ! Binds to C function ``make_rhoX_flux_2d``
 
-    integer         , intent(in   ) :: lev, lo(2), hi(2)
-    integer         , intent(in   ) :: fx_lo(2), fx_hi(2), nc_fx
-    double precision, intent(inout) :: sfluxx(fx_lo(1):fx_hi(1),fx_lo(2):fx_hi(2),nc_fx)
-    integer         , intent(in   ) :: fy_lo(2), fy_hi(2), nc_fy
-    double precision, intent(inout) :: sfluxy(fy_lo(1):fy_hi(1),fy_lo(2):fy_hi(2),nc_fy)
-    integer         , intent(in   ) :: eta_lo(2), eta_hi(2)
-    double precision, intent(inout) :: etarhoflux(eta_lo(1):eta_hi(1),eta_lo(2):eta_hi(2))
-    integer         , intent(in   ) :: x_lo(2), x_hi(2), nc_x
-    double precision, intent(inout) :: sedgex(x_lo(1):x_hi(1),x_lo(2):x_hi(2),nc_x)
-    integer         , intent(in   ) :: y_lo(2), y_hi(2), nc_y
-    double precision, intent(inout) :: sedgey(y_lo(1):y_hi(1),y_lo(2):y_hi(2),nc_y)
-    integer         , intent(in   ) :: u_lo(2), u_hi(2)
-    double precision, intent(in   ) :: umac  (u_lo(1):u_hi(1),u_lo(2):u_hi(2))
-    integer         , intent(in   ) :: v_lo(2), v_hi(2)
-    double precision, intent(in   ) :: vmac  (v_lo(1):v_hi(1),v_lo(2):v_hi(2))
+    integer         , intent(in   ) :: lo(3), hi(3)
+    integer  , value, intent(in   ) :: lev, idir
+    integer         , intent(in   ) :: fx_lo(3), fx_hi(3)
+    integer  , value, intent(in   ) :: nc_fx
+    double precision, intent(inout) :: sflux(fx_lo(1):fx_hi(1),fx_lo(2):fx_hi(2),fx_lo(3):fx_hi(3),nc_fx)
+    integer         , intent(in   ) :: eta_lo(3), eta_hi(3)
+    double precision, intent(inout) :: etarhoflux(eta_lo(1):eta_hi(1),eta_lo(2):eta_hi(2),eta_lo(3):eta_hi(3))
+    integer         , intent(in   ) :: x_lo(3), x_hi(3)
+    integer  , value, intent(in   ) :: nc_x
[…]
