### 2019-11-20 252a1da
For simplified SDC, switch to Castro approach to call the integrator subroutine directly instead of burner, since that will not require changes to the individual networks.
