### 2019-11-20 3b0a41a
Switch from using SDC_METHOD to defining SDC_EVOLVE_ENTHALPY now supported in Microphysics. Revert changes to the burner for SDC since we now call the integrator directly.
