+++ Microphysics/EOS/eos_type.F90
@@ -154,11 +154,6 @@ module eos_type_module
     real(rt) :: mu
     real(rt) :: mu_e
     real(rt) :: y_e
-#ifdef EXTRA_THERMO
-    real(rt) :: dedX(nspec)
-    real(rt) :: dpdX(nspec)
-    real(rt) :: dhdX(nspec)
-#endif
     real(rt) :: gam1
     real(rt) :: cs
 
@@ -226,10 +221,6 @@ contains
     to_eos % zbar = from_eos % zbar
 
 #ifdef EXTRA_THERMO
-    to_eos % dedX(:) = from_eos % dedX(:)
-    to_eos % dpdX(:) = from_eos % dpdX(:)
-    to_eos % dhdX(:) = from_eos % dhdX(:)
-    
     to_eos % dpdA = from_eos % dpdA
     to_eos % dpdZ = from_eos % dpdZ
     to_eos % dedA = from_eos % dedA
@@ -240,71 +231,6 @@ contains
 
   end subroutine copy_eos_t
 
-  
-  ! Given a set of mass fractions, calculate quantities that depend
-  ! on the composition like abar and zbar.
-
-  subroutine composition(state)
-
-    !$acc routine seq
-
-    use amrex_constants_module, only: ONE
-    use network, only: aion, aion_inv, zion
-
-    implicit none
-
-    type (eos_t), intent(inout) :: state
-
-    !$gpu
-
-    ! Calculate abar, the mean nucleon number,
-    ! zbar, the mean proton number,
-    ! mu, the mean molecular weight,
-    ! mu_e, the mean number of nucleons per electron, and
-    ! y_e, the electron fraction.
-
-    state % mu_e = ONE / (sum(state % xn(:) * zion(:) * aion_inv(:)))
-    state % y_e = ONE / state % mu_e
-
-    state % abar = ONE / (sum(state % xn(:) * aion_inv(:)))
-    state % zbar = state % abar / state % mu_e
-
-  end subroutine composition
-
-#ifdef EXTRA_THERMO
-  ! Compute thermodynamic derivatives with respect to xn(:)
-
-  subroutine composition_derivatives(state)
-
-    use amrex_constants_module, only: ZERO
-    use network, only: aion, aion_inv, zion
-
-    implicit none
-
-    type (eos_t), intent(inout) :: state
-
-    !$gpu
-
-    state % dpdX(:) = state % dpdA * (state % abar * aion_inv(:))   &
-                                   * (aion(:) - state % abar) &
-                    + state % dpdZ * (state % abar * aion_inv(:))   &
-                                   * (zion(:) - state % zbar)
-
-    state % dEdX(:) = state % dedA * (state % abar * aion_inv(:))   &
-                                   * (aion(:) - state % abar) &
-                    + state % dedZ * (state % abar * aion_inv(:))   &
-                                   * (zion(:) - state % zbar)
-
-    if (state % dPdr .ne. ZERO) then
-
-       state % dhdX(:) = state % dedX(:) &
-                       + (state % p / state % rho**2 - state % dedr) &
-                       *  state % dPdX(:) / state % dPdr
-
-    endif
-
-  end subroutine composition_derivatives
-#endif
 
 
   ! Normalize the mass fractions: they must be individually positive
