### 2019-11-20 07ec1c5
Revert "re-add X derivatives to eos_t to compile with Microphysics sdc-maestroex branch"
This reverts commit 65c30c584b9afd6870354cfa3deb448b0f355f84.
