+++ Source/Src_nd/make_explicit_thermal.F90
@@ -17,7 +17,8 @@ contains
     !
     ! note: we explicitly fill the ghostcells by looping over them directly
 
-    use eos_type_module, only : eos_t, eos_input_rt, composition_derivatives
+    use eos_type_module, only : eos_t, eos_input_rt
+    use eos_composition_module, only : eos_xderivs_t, composition_derivatives
     use eos_module
     use conductivity_module
     use network, only: nspec
@@ -42,6 +43,7 @@ contains
     ! Local
     integer :: i,j,k,comp
     type (eos_t) :: eos_state
+    type (eos_xderivs_t) :: eos_xderivs
 
     !$gpu
 
@@ -83,8 +85,10 @@ contains
                      (1.0d0-eos_state%p/(eos_state%rho*eos_state%dpdr))+ &
                      eos_state%dedr/eos_state%dpdr)
 
+                call composition_derivatives(eos_state, eos_xderivs)
+
                 do comp=1,nspec
-                   Xkcoeff(i,j,k,comp) = (eos_state % conductivity/eos_state%cp) * eos_state % dhdX(comp)
+                   Xkcoeff(i,j,k,comp) = (eos_state % conductivity/eos_state%cp) * eos_xderivs % dhdX(comp)
                 enddo
 
              endif
