### 2019-11-20 912e946
If we are using Simplified SDC then in the burner, use the integrator interface directly instead of actual_burner.
