### 2019-11-20 #1757 @zingale
just checking in on this -- I think we are all in sync now, right?  Is there anything else we need to do?
