### 2019-11-21 #1757 @arfon
@zingale - I forgot to ask for the AAS DOI before accepting this so while the JOSS paper is accepted and published we'll need to add the AAS metadata to the submission too when we have it.

@crawfordsm - is the AAS paper accepted yet? If so, do we know what the DOI is?
