### 2019-11-21 #1757 @doreenfan
The AAS paper has been accepted, but I'm not sure what the DOI is.
