### 2019-11-21 64d01a6
change SIMPLIFIED_SDC to SDC
