+++ docs/_sources/file/make_intra_coeffs_f.rst.txt
@@ -1,0 +1,8 @@
+make_intra_coeffs.F90
+=====================
+
+make\_intra\_coeffs\_module
+---------------------------
+
+.. f:automodule:: make_intra_coeffs_module
+
