+++ docs/_sources/Src_nd_files.rst.txt
@@ -32,6 +32,7 @@ Src_nd
    file/make_gamma_f
    file/make_grav_f
    file/make_heating_f
+   file/make_intra_coeffs_f
    file/make_pi_cc_f
    file/make_plot_variables_f
    file/make_psi_f
