### 2019-11-22 944ca61
add burner invert cutoff density flag to SDC burner loop
