### 2019-11-22 3cfd369
zero omegadot and Hnuc if do_burning=false
