### 2019-11-22 969d1cd
remove extra w0 terms from flux calculations that were added due to SDC merge
