+++ Source/Src_nd/make_flux.F90
@@ -143,17 +143,17 @@ contains
                       ! edge states are rho' and X.  To make the (rho X) flux,
                       ! we need the edge state of rho0
                       sflux(i,j,k,comp) = &
-                           (umac(i,j,k)+w0(lev,j))*(rho0_edge+sedge(i,j,k,rho_comp))*sedge(i,j,k,comp)
+                           umac(i,j,k)*(rho0_edge+sedge(i,j,k,rho_comp))*sedge(i,j,k,comp)
 
                    else if (species_pred_type == predict_rhoX) then
                       ! edge states are (rho X)
                       sflux(i,j,k,comp) = &
-                           (umac(i,j,k)+w0(lev,j))*sedge(i,j,k,comp)
+                           umac(i,j,k)*sedge(i,j,k,comp)
 
                    else if (species_pred_type == predict_rho_and_X) then
                       ! edge state are rho and X
                       sflux(i,j,k,comp) = &
-                           (umac(i,j,k)+w0(lev,j))*sedge(i,j,k,rho_comp)*sedge(i,j,k,comp)
+                           umac(i,j,k)*sedge(i,j,k,rho_comp)*sedge(i,j,k,comp)
 
                    endif
 
@@ -270,16 +270,16 @@ contains
                    if (species_pred_type == predict_rhoprime_and_X) then
                       ! edge states are rho' and X.  To make the (rho X)
                       ! flux, we need the edge state of rho0
-                      sflux(i,j,k,comp) = (umac(i,j,k)+w0(lev,k))* &
+                      sflux(i,j,k,comp) = umac(i,j,k)* &
                            (rho0_edge+sedge(i,j,k,rho_comp))*sedge(i,j,k,comp)
 
                    else if (species_pred_type == predict_rhoX) then
                       ! edge states are (rho X)
-                      sflux(i,j,k,comp) = (umac(i,j,k)+w0(lev,k))*sedge(i,j,k,comp)
+                      sflux(i,j,k,comp) = umac(i,j,k)*sedge(i,j,k,comp)
 
                    else if (species_pred_type == predict_rho_and_X) then
                       ! edge states are rho and X
-                      sflux(i,j,k,comp) = (umac(i,j,k)+w0(lev,k))* &
+                      sflux(i,j,k,comp) = umac(i,j,k)* &
                            sedge(i,j,k,rho_comp)*sedge(i,j,k,comp)
 
                    endif
@@ -349,16 +349,16 @@ contains
                 if (species_pred_type == predict_rhoprime_and_X) then
                    ! edge states are rho' and X.  To make the (rho X)
                    ! flux, we need the edge state of rho0
-                   sflux(i,j,k,comp) = (umac(i,j,k) + w0mac(i,j,k)) * &
+                   sflux(i,j,k,comp) = umac(i,j,k) * &
                         (rho0_edge(i,j,k) + sedge(i,j,k,rho_comp))*sedge(i,j,k,comp)
 
                 else if (species_pred_type == predict_rhoX) then
                    ! edge states are (rho X)
-                   sflux(i,j,k,comp) = (umac(i,j,k) + w0mac(i,j,k)) * sedge(i,j,k,comp)
+                   sflux(i,j,k,comp) = umac(i,j,k) * sedge(i,j,k,comp)
 
[…]
