### 2019-11-22 4ab7490
add back w0 before calling EnthalpyAdvance in advance subroutine
