+++ CHANGES.md
@@ -1,0 +1,273 @@
+# 19.11
+
+  * Port 3d hydro subroutine to GPU
+
+  * Add enthalpy_pred_type options: predict_T_then_rhohprime,
+    predict_T_then_h, predict_Tprime_then_h
+
+  * Updates to documentation for JOSS paper, including new logo
+
+  * Add scripts for plotting and comparing plotfiles
+
+  * Bugfixes: boundary condition for thermal diffusion, fill w0_cart
+    after restart/regrid
+
+  * Update tests: rotating_star, code_comp, multilevel reacting_bubble
+
+
+# 19.10
+
+  * Move base state from 1D array to cartesian grid in many subroutines
+
+  * Start porting hydro subroutines to GPU (ongoing)
+
+  * Bug fix: fill w0_cart based on edge-centered w0, not cell-centered
+
+  * Add option to output Nonaka plot (USE_NONAKA_PLOT)
+
+  * Add diagnostics for planar geometries
+
+  * Updates to documentation and JOSS paper
+
+
+# 19.09
+
+  * Minor changes to some problem setups and bugfixes
+
+  * Remove 1D support (with compile-time error)
+
+  * Significant updates to sphinx documentation, including doxygen
+
+  * JOSS paper added; ready for submission
+
+
+# 19.08
+
+  * Option to plot gravity
+
+  * Added checkpoint at simulation time level option
+
+  * Fixed a ppm_type=2 index bug
+
+  * GPU ports: FirstDtm EstDt
+
+  * Various GPU optimizations
+
+  * Boundary condition bugs in make_edge_scal
+
+
+# 19.07
+
+  * Add BSD-3 license
+
+  * Minor bugfixes at initialization
+
+  * Add diagnostic outputs of radially averaged profiles
+
+  * Add tests: test_stability, code_comp (code comparison project for Exeter)
+
+
+# 19.06
+
+  * Implement multiple tagging criteria for multiple levels of AMR
+
+  * Specify how often to print diagnostics using sum_interval and
+    sum_per
+
+  * Plot files now include base states output
+
+  * Bugfix to new temporal algorithm in MaestroAdvanceAvg.cpp
+
+  * Many thermodynamics and source term subroutines offloaded to GPU:
+    burner_loop, make_thermal_coeffs, mk_sponge, rhoh_vs_t,
+    make_gamma, mkrhohforce, make_S, update_scal, update_vel,
+    compute_dt, make_pi_cc
+
+  * Add cuda managed attributes to meth_params.F90 using
+    parse_maestro_params.py
+
+  * Add RealVector typedef so that Real Vectors can be CUDA managed
+
+  * Move put_1d_array_on_sphr routine calls to C++ to avoid calls on the GPU
+
+  * If USE_CUDA=TRUE, then set USE_GPU_PRAGMA=TRUE and add CUDA to
+    define
+
+  * Remove loop over level in 1d divu_cart and gradp0
+
+  * Offloaded plotting routines in make_plot_variables.F90 to GPU
+
+
+# 19.05
+
+  * Travis documentation updates
+
+  * Plotfile and small_plotfile logic fixed
+
+  * Additional variables in plotfiles
+
+  * Some progress toward the do_smallscale option
+
+  * Bugfix to parsing strings with spaces in parse_maestro_params.py
+
+  * Added mach_jet problem and inflow routine
+
+  * Added fill_ext_bc routines framework to local copies of
+    bc_fill_nd.F90 to support Dirichlet boundary conditions
+
+  * Custom BndryFuncArrayMaestro that allows the bc fill routines to
+    be component-aware
+
+  * MLMG boundary conditions fixes
+
[…]
