### 2019-12-03 191129f
fixed GPU non-determinism
