+++ sphinx_docs/source/conf.py
@@ -54,6 +54,7 @@ extensions = ['sphinx.ext.autodoc',
               'sphinx.ext.viewcode',
               'sphinxcontrib.bibtex',
               'sphinx.ext.autosummary',
+              'numpydoc',
               'sphinx.ext.githubpages',
               'breathe',
               'sphinxfortran.fortran_domain',
