### 2019-12-03 0d338f2
added a couple more launch regions
