+++ Source/MaestroFill3dData.cpp
@@ -255,7 +255,7 @@ Maestro::MakeS0mac (const RealVector& s0,
         Abort("Error: only call MakeS0mac for spherical");
     }
 
-    // Construct a cartesian version of w0
+    // Construct a cartesian version of s0
     Vector<MultiFab> s0_cart(finest_level+1);
     for (int lev=0; lev<=finest_level; ++lev) {
         s0_cart[lev].define(grids[lev], dmap[lev], 1, 2);
@@ -319,6 +319,12 @@ Maestro::MakeNormal ()
     // timer for profiling
     BL_PROFILE_VAR("Maestro::MakeNormal()",MakeNormal);
 
+#ifdef AMREX_USE_CUDA
+    auto not_launched = Gpu::notInLaunchRegion();
+    // turn on GPU
+    if (not_launched) Gpu::setLaunchRegion(true);
+#endif
+
     for (int lev=0; lev<=finest_level; ++lev) {
 
         // get references to the MultiFabs at level lev
@@ -331,13 +337,22 @@ Maestro::MakeNormal ()
 #endif
         for ( MFIter mfi(normal_mf, true); mfi.isValid(); ++mfi ) {
 
+            const Box& tileBox = mfi.tilebox();
+
             // call fortran subroutine
             // use macros in AMReX_ArrayLim.H to pass in each FAB's data,
             // lo/hi coordinates (including ghost cells), and/or the # of components
-            make_normal(BL_TO_FORTRAN_3D(normal_mf[mfi]), dx);
+#pragma gpu box(tileBox)
+            make_normal(AMREX_INT_ANYD(tileBox.loVect()),
+                        AMREX_INT_ANYD(tileBox.hiVect()),
+                        BL_TO_FORTRAN_ANYD(normal_mf[mfi]), 
+                        AMREX_REAL_ANYD(dx));
         }
     }
-
+#ifdef AMREX_USE_CUDA
+    // turn off GPU
+    if (not_launched) Gpu::setLaunchRegion(false);
+#endif
 }
 
 
