### 2019-12-04 46655b9
added in some missing launch regions
