+++ Source/Src_nd/fill_3d_data.F90
@@ -1432,122 +1432,152 @@ contains
 
   end subroutine make_normal
 
-  subroutine put_data_on_faces(lo, hi, &
+  subroutine put_data_on_faces(lo, hi, idir, &
        scc, cc_lo, cc_hi, &
-       facex, x_lo, x_hi, &
-       facey, y_lo, y_hi, &
-#if (AMREX_SPACEDIM == 3)
-       facez, z_lo, z_hi, &
-#endif
+       face, x_lo, x_hi, &
        harmonic_avg) bind(C, name="put_data_on_faces")
 
     integer         , intent(in   ) :: lo(3), hi(3)
+    integer  , value, intent(in   ) :: idir
     integer         , intent(in   ) :: cc_lo(3), cc_hi(3)
     double precision, intent(in   ) :: scc(cc_lo(1):cc_hi(1),cc_lo(2):cc_hi(2),cc_lo(3):cc_hi(3))
     integer         , intent(in   ) :: x_lo(3), x_hi(3)
-    double precision, intent(inout) :: facex(x_lo(1):x_hi(1),x_lo(2):x_hi(2),x_lo(3):x_hi(3))
-    integer         , intent(in   ) :: y_lo(3), y_hi(3)
-    double precision, intent(inout) :: facey(y_lo(1):y_hi(1),y_lo(2):y_hi(2),y_lo(3):y_hi(3))
-#if (AMREX_SPACEDIM == 3)
-    integer         , intent(in   ) :: z_lo(3), z_hi(3)
-    double precision, intent(inout) :: facez(z_lo(1):z_hi(1),z_lo(2):z_hi(2),z_lo(3):z_hi(3))
-#endif
-    integer         , intent(in   ) :: harmonic_avg
+    double precision, intent(inout) :: face(x_lo(1):x_hi(1),x_lo(2):x_hi(2),x_lo(3):x_hi(3))
+    integer  , value, intent(in   ) :: harmonic_avg
 
     ! local
     integer :: i,j,k
-    double precision :: denom
+    double precision :: denom, prod
+
+    !$gpu
 
     if (harmonic_avg .eq. 1) then
+        do k = lo(3),hi(3)
+            do j = lo(2),hi(2)
+                do i = lo(1),hi(1)
+                    if (idir == 1) then 
+                        denom = (scc(i,j,k) + scc(i-1,j,k))
+                        prod = scc(i,j,k) * scc(i-1,j,k)
+                    else if (idir == 2) then 
+                        denom = (scc(i,j,k) + scc(i,j-1,k))
+                        prod = scc(i,j,k) * scc(i-1,j,k)
+                    else 
+                        denom = (scc(i,j,k) + scc(i,j,k-1))
+                        prod = scc(i,j,k) * scc(i,j,k-1)
+                    endif
 
-       k = lo(3)
-       j = lo(2)
-#if (AMREX_SPACEDIM == 3)
-       do k = lo(3),hi(3)
-#endif
-          do j = lo(2),hi(2)
-             do i = lo(1),hi(1)+1
-                denom = (scc(i,j,k) + scc(i-1,j,k))
-                if (denom .ne. 0.d0) then
-                   facex(i,j,k) = TWO*(scc(i,j,k) * scc(i-1,j,k)) / denom
-                else
-                   facex(i,j,k) = HALF*denom
-                end if
-             end do
-          end do
-#if (AMREX_SPACEDIM == 3)
-       end do
-#endif
-
-#if (AMREX_SPACEDIM == 3)
-       do k = lo(3),hi(3)
-#endif
-          do j = lo(2),hi(2)+1
-             do i = lo(1),hi(1)
-                denom = (scc(i,j,k) + scc(i,j-1,k))
-                if (denom .ne. 0.d0) then
[…]
