### 2019-12-05 305a1ec
Initialize burn_t % self_heat before calling actual_rhs where it is checked. Makes Valgrind happy.
