### 2019-12-05 a76284f
fix floating point overflow
