+++ Source/Src_nd/compute_dt.F90
@@ -60,7 +60,7 @@ contains
     !$gpu
 
     rho_min = 1.d-20
-    dt_temp = 1.e99
+    dt_temp = 1.d99
 
     eps = 1.d-8
 
