### 2019-12-10 e6b762a
added analytic heating
