+++ Exec/SCIENCE/urca/analysis/UrcaAnalysis/__init__.py
@@ -1,2 +1,2 @@
-from .yt_urca_fields import UrcaShellFields
+from .yt_urca_fields import UrcaShellFields, DatasetHelpers
 from .FileToolkit import *
