### 2019-12-10 39b9c39
fixed a tiling but in the velocity initialization... was overwriting the entire grid velocity to be zero, not just the tile
