+++ paper-parser/papers.bib
@@ -398,3 +398,54 @@ archivePrefix = "arXiv",
   adsnote = {Provided by the SAO/NASA Astrophysics Data System},
   subject = {stellar evolution}
 }
+
+
+@ARTICLE{2019arXiv190803634F,
+       author = {{Fan}, Duoming and {Nonaka}, Andrew and {Almgren}, Ann S. and
+         {Harpole}, Alice and {Zingale}, Michael},
+        title = "{MAESTROeX: A Massively Parallel Low Mach Number Astrophysical Solver}",
+      journal = {arXiv e-prints},
+     keywords = {Physics - Computational Physics, Astrophysics - Solar and Stellar Astrophysics},
+         year = "2019",
+        month = "Aug",
+          eid = {arXiv:1908.03634},
+        pages = {arXiv:1908.03634},
+archivePrefix = {arXiv},
+       eprint = {1908.03634},
+ primaryClass = {physics.comp-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv190803634F},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+      subject = {algorithm},
+}
+
+@article{maestroex_joss,
+author = {{Fan}, D. and {Nonaka}, A. and {Almgren}, A. and {Willcox}, D.
+          and {Harpole}, A. and {Zingale}, M.},
+    title = "{MAESTROeX: A Massively Parallel Low Mach Number Astrophysical Solver}",
+  journal = {Journal of Open Source Software},
+     year = {2019},
+  volume = {4},
+  number = 43,
+  pages = 1757,
+   url = {https://joss.theoj.org/papers/10.21105/joss.01757},
+   doi = {10.21105/joss.01757},
+   subject = {algorithm},
+}
+
+@article{astronum2019_maestro,
+        author = {{Harpole}, A. and {Fan}, D. and {Katz}, M.~P. and {Nonaka}, A.~J. and {Willcox}, D.~E. and {Zingale}, M.},
+        title = "{Modelling low Mach number stellar hydrodynamics with MAESTROeX}",
+        journal = {arXiv e-prints},
+        keywords = {Physics - Computational Physics, Astrophysics - Solar and Stellar Astrophysics},
+        year = "2019",
+        month = "Oct",
+        eid = {arXiv:1910.12979},
+        pages = {arXiv:1910.12979},
+        archivePrefix = {arXiv},
+        eprint = {1910.12979},
+        primaryClass = {physics.comp-ph},
+        adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv191012979H},
+        adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+        note = "{submittied to Proceedings of AstroNum 2019}",
+        subject = {algorithms: low Mach hydro},
+}
