### 2019-12-11 d8889ec
added burning_cutoff_density_lo/hi
