### 2019-12-11 bca1814
If we are in spherical 3D geometry, find our radial index in the 1D profile.
