### 2019-12-11 0b92405
Simplify by finding rloc for 2D/3D/3D spherical and interpolating to that instead of finding the radial index.
