+++ paper-parser/papers.bib
@@ -146,7 +146,6 @@ archivePrefix = "arXiv",
    year = {2014},
   pages = {3--15},
   link = {http://www.springer.com/us/book/9783319056838},
-  subject = {algorithms: low Mach hydro},
   subject = {algorithm}
 }
 
@@ -447,5 +446,5 @@ author = {{Fan}, D. and {Nonaka}, A. and {Almgren}, A. and {Willcox}, D.
         adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv191012979H},
         adsnote = {Provided by the SAO/NASA Astrophysics Data System},
         note = "{submittied to Proceedings of AstroNum 2019}",
-        subject = {algorithms: low Mach hydro},
+        subject = {algorithm}
 }
