### 2019-12-11 827cace
Add spherical geometry for rotating star artificial heating (#127)
* If we are in spherical 3D geometry, find our radial index in the 1D profile.

* Simplify by finding rloc for 2D/3D/3D spherical and interpolating to that instead of finding the radial index.

In addition to 2D/3D cartesian, this adds 3D spherical interpolation for the artificial heating in the rotating star setup.

Slicing the S field in the plotfiles shows this is spherical interpolation with the spherical octant geometry.
