+++ docs/_sources/runtime_parameters.rst.txt
@@ -277,8 +277,9 @@ namespace: ``maestro``
 +----------------------------------------+---------------------------------------------------------+---------------+
 | ``burning_cutoff_density``             | The density below which we disable burning              | -1.0          |
 +----------------------------------------+---------------------------------------------------------+---------------+
-| ``burning_invert_cutoff_density``      | Should we instead disable burning above a certain       | false         |
-|                                        | density?                                                |               |
+| ``burning_cutoff_density_lo``          | The density below which we disable burning              | -1.0          |
++----------------------------------------+---------------------------------------------------------+---------------+
+| ``burning_cutoff_density_hi``          | The density above which we disable burning              | 1.e100        |
 +----------------------------------------+---------------------------------------------------------+---------------+
 | ``buoyancy_cutoff_factor``             | The multiplicative factor (over base\_cutoff\_density)  | 5.0           |
 |                                        | below which we do zero out the buoyancy term in the     |               |
