### 2019-12-11 9b05829
remove burning_cutoff_density completely since it was replaced by burning_cutoff_density_lo
