+++ Source/param_includes/maestro_params.H
@@ -76,10 +76,8 @@ static amrex::Real sponge_start_factor;
 static bool        plot_sponge_fdamp;
 static amrex::Real anelastic_cutoff_density;
 static amrex::Real base_cutoff_density;
-static amrex::Real burning_cutoff_density;
 static amrex::Real burning_cutoff_density_lo;
 static amrex::Real burning_cutoff_density_hi;
-static bool        burning_invert_cutoff_density;
 static amrex::Real buoyancy_cutoff_factor;
 static amrex::Real dpdt_factor;
 static bool        do_planar_invsq_grav;
