+++ docs/_sources/runtime_parameters.rst.txt
@@ -275,8 +275,6 @@ namespace: ``maestro``
 |                                        | constant. Refer to Section \ref{Sec:Base Cutoff         |               |
 |                                        | Density}                                                |               |
 +----------------------------------------+---------------------------------------------------------+---------------+
-| ``burning_cutoff_density``             | The density below which we disable burning              | -1.0          |
-+----------------------------------------+---------------------------------------------------------+---------------+
 | ``burning_cutoff_density_lo``          | The density below which we disable burning              | -1.0          |
 +----------------------------------------+---------------------------------------------------------+---------------+
 | ``burning_cutoff_density_hi``          | The density above which we disable burning              | 1.e100        |
