### 2019-12-11 77a366b
need to multiply pi by beta0 if using alt_energy_fix
