+++ Source/MaestroNodalProj.cpp
@@ -325,7 +325,8 @@ Maestro::NodalProj (int proj_type,
         for (int lev=0; lev<=finest_level; ++lev) {
             for (int dir=0; dir<AMREX_SPACEDIM; ++dir) {
                 MultiFab::Multiply(gphi[lev],beta0_cart[lev],0,dir,1,0);
-            }
+            }    
+	    MultiFab::Multiply(phi[lev],beta0_cart[lev],0,0,1,0);
         }
     }
 
@@ -372,7 +373,7 @@ Maestro::NodalProj (int proj_type,
     else if (proj_type == pressure_iters_comp) {
         // Vproj = Vproj - sig*grad(phi)
         // we do this manually instead of using mlndlap.updateVelocity() because
-        // for alt_energy_fix we neet beta0*grad(phi)
+        // for alt_energy_fix we need beta0*grad(phi)
         for (int lev=0; lev<=finest_level; ++lev) {
             for (int dir=0; dir<AMREX_SPACEDIM; ++dir) {
                 MultiFab::Multiply(gphi[lev],sig[lev],0,dir,1,0);
@@ -389,7 +390,7 @@ Maestro::NodalProj (int proj_type,
     else if (proj_type == regular_timestep_comp) {
         // Vproj = Vproj - sig*grad(phi)
         // we do this manually instead of using mlndlap.updateVelocity() because
-        // for alt_energy_fix we neet beta0*grad(phi)
+        // for alt_energy_fix we need beta0*grad(phi)
         for (int lev=0; lev<=finest_level; ++lev) {
             for (int dir=0; dir<AMREX_SPACEDIM; ++dir) {
                 MultiFab::Multiply(gphi[lev],sig[lev],0,dir,1,0);
