### 2019-12-11 5dbf3aa
Revert "need to multiply pi by beta0 if using alt_energy_fix"
This is already done in make_pi_cc subroutine.

This reverts commit 77a366bcbc194a44a1125dc859d2730154b67e37.
