### 2019-12-11 30bfae6
revert a make_heating change.  accidentally committed a version that restricts the radius based on the burning_cutoff_densities
