### 2019-12-12 3053b81
add heating_cutoff_density_lo and _hi
