+++ Source/param_includes/maestro_queries.H
@@ -78,6 +78,8 @@ pp.query("anelastic_cutoff_density", anelastic_cutoff_density);
 pp.query("base_cutoff_density", base_cutoff_density);
 pp.query("burning_cutoff_density_lo", burning_cutoff_density_lo);
 pp.query("burning_cutoff_density_hi", burning_cutoff_density_hi);
+pp.query("heating_cutoff_density_lo", heating_cutoff_density_lo);
+pp.query("heating_cutoff_density_hi", heating_cutoff_density_hi);
 pp.query("buoyancy_cutoff_factor", buoyancy_cutoff_factor);
 pp.query("dpdt_factor", dpdt_factor);
 pp.query("do_planar_invsq_grav", do_planar_invsq_grav);
