+++ Source/param_includes/maestro_defaults.H
@@ -78,6 +78,8 @@ amrex::Real Maestro::anelastic_cutoff_density = -1.0;
 amrex::Real Maestro::base_cutoff_density = -1.0;
 amrex::Real Maestro::burning_cutoff_density_lo = -1.0;
 amrex::Real Maestro::burning_cutoff_density_hi = 1.e100;
+amrex::Real Maestro::heating_cutoff_density_lo = -1.0;
+amrex::Real Maestro::heating_cutoff_density_hi = 1.e100;
 amrex::Real Maestro::buoyancy_cutoff_factor = 5.0;
 amrex::Real Maestro::dpdt_factor = 0.0;
 bool        Maestro::do_planar_invsq_grav = false;
