+++ Source/param_includes/meth_params.F90
@@ -56,6 +56,8 @@ module meth_params_module
   double precision , allocatable, save :: base_cutoff_density
   double precision , allocatable, save :: burning_cutoff_density_lo
   double precision , allocatable, save :: burning_cutoff_density_hi
+  double precision , allocatable, save :: heating_cutoff_density_lo
+  double precision , allocatable, save :: heating_cutoff_density_hi
   double precision , allocatable, save :: buoyancy_cutoff_factor
   double precision , allocatable, save :: dpdt_factor
   logical          , allocatable, save :: do_planar_invsq_grav
@@ -116,6 +118,8 @@ module meth_params_module
   attributes(managed) :: base_cutoff_density
   attributes(managed) :: burning_cutoff_density_lo
   attributes(managed) :: burning_cutoff_density_hi
+  attributes(managed) :: heating_cutoff_density_lo
+  attributes(managed) :: heating_cutoff_density_hi
   attributes(managed) :: buoyancy_cutoff_factor
   attributes(managed) :: dpdt_factor
   attributes(managed) :: do_planar_invsq_grav
@@ -216,6 +220,10 @@ contains
     burning_cutoff_density_lo = -1.0d0;
     allocate(burning_cutoff_density_hi)
     burning_cutoff_density_hi = 1.d100;
+    allocate(heating_cutoff_density_lo)
+    heating_cutoff_density_lo = -1.0d0;
+    allocate(heating_cutoff_density_hi)
+    heating_cutoff_density_hi = 1.d100;
     allocate(buoyancy_cutoff_factor)
     buoyancy_cutoff_factor = 5.0d0;
     allocate(dpdt_factor)
@@ -314,6 +322,8 @@ contains
     call pp%query("base_cutoff_density", base_cutoff_density)
     call pp%query("burning_cutoff_density_lo", burning_cutoff_density_lo)
     call pp%query("burning_cutoff_density_hi", burning_cutoff_density_hi)
+    call pp%query("heating_cutoff_density_lo", heating_cutoff_density_lo)
+    call pp%query("heating_cutoff_density_hi", heating_cutoff_density_hi)
     call pp%query("buoyancy_cutoff_factor", buoyancy_cutoff_factor)
     call pp%query("dpdt_factor", dpdt_factor)
     call pp%query("do_planar_invsq_grav", do_planar_invsq_grav)
@@ -441,6 +451,12 @@ contains
     end if
     if (allocated(burning_cutoff_density_hi)) then
         deallocate(burning_cutoff_density_hi)
+    end if
+    if (allocated(heating_cutoff_density_lo)) then
+        deallocate(heating_cutoff_density_lo)
+    end if
+    if (allocated(heating_cutoff_density_hi)) then
+        deallocate(heating_cutoff_density_hi)
     end if
     if (allocated(buoyancy_cutoff_factor)) then
         deallocate(buoyancy_cutoff_factor)
