### 2019-12-12 696cb44
apply heating_cutoff_density_lo/hi to rotating star problem provide inputs file to apply heating fro Si shell inward to the core
