+++ Exec/science/rotating_star/make_heating.F90
@@ -4,7 +4,8 @@ module make_heating_module
   use model_parser_module
   use amrex_constants_module
   use base_state_geometry_module, only: nr_fine, dr, nr, max_radial_level, center
-  use meth_params_module, only: nscal, model_file, rho_comp, prob_lo, spherical, burning_cutoff_density_lo, burning_cutoff_density_hi
+  use meth_params_module, only: nscal, model_file, rho_comp, prob_lo, spherical, &
+       heating_cutoff_density_lo, heating_cutoff_density_hi
   use probin_module, only: use_analytic_heating
   
   implicit none
@@ -60,12 +61,13 @@ contains
                     end if
 
 
-!                    rho = scal(i,j,k,rho_comp)
-!                    if (rho > burning_cutoff_density_lo .and. rho < burning_cutoff_density_hi) then
+                    rho = scal(i,j,k,rho_comp)
+                    if (rho > heating_cutoff_density_lo .and. rho < heating_cutoff_density_hi) then
                        rho_Hext(i,j,k)  = interpolate(rloc, ienuc_model) * scal(i,j,k,rho_comp)
-!                    else
-!                       rho_Hext(i,j,k) = 0.d0
-!                    end if
+                    else
+                       rho_Hext(i,j,k) = 0.d0
+                    end if
+                    
                 end do
             end do
         end do
