### 2019-12-12 646e1d8
add new option -- use_omegadot_terms_in_S
