### 2019-12-12 ce992f8
add use_omegadot_terms_in_S to spherical too
