### 2019-12-14 f494b32
update to new actual_rhs interface
