### 2019-12-16 9c7ed17
put nodal solver within launch regions so it runs on GPU
