+++ Source/MaestroAdvanceAvg.cpp
@@ -885,6 +885,12 @@ Maestro::AdvanceTimeStepAverage (bool is_initIter) {
     // STEP 11 -- update the velocity
     //////////////////////////////////////////////////////////////////////////////
 
+#ifdef AMREX_USE_CUDA
+    auto not_launched = Gpu::notInLaunchRegion();
+    // turn on GPU
+    if (not_launched) Gpu::setLaunchRegion(true);
+#endif
+
     if (maestro_verbose >= 1) {
 	Print() << "<<< STEP 11: update and project new velocity >>>" << std::endl;
     }
@@ -971,6 +977,11 @@ Maestro::AdvanceTimeStepAverage (bool is_initIter) {
 
     // call nodal projection
     NodalProj(proj_type,rhcc_for_nodalproj);
+
+#ifdef AMREX_USE_CUDA
+    // turn off GPU
+    if (not_launched) Gpu::setLaunchRegion(false);
+#endif
 
     // wallclock time
     Real end_total_nodalproj = ParallelDescriptor::second() - start_total_nodalproj;
