### 2019-12-17 a8046a4
Burner routine returns success of burner_loop
