### 2019-12-18 030bfb6
remove extra launch region ifdefs from MaestroAdvance
