### 2019-12-19 a30a100
Adding AAS DOI (#128)
