### 2019-12-19 #1757 @arfon
OK, the paper is now updated with the AAS DOI information. Thanks again @tukss & @pgrete !
